### 2025-11-19 #9472 @editorialbot
[PRE REVIEW]: Newtrinos.jl: A Julia Package for Global Neutrino Oscillation Fits
**Submitting author:** @philippeller (<a href="http://orcid.org/0000-0001-6354-5209">Philipp Eller</a>)
**Repository:** https://github.com/philippeller/Newtrinos.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v0.0.1
