### 2025-11-19 #9472 @kyleniemeyer
@philippeller two questions:
- I see a `test` directory, but could you add some brief instructions to the README on how reviewers and users may run the tests?
- In terms of documentation, I see some extensive examples in the README, and then a `notebooks` directory that seems to contain some examples. Is that the case? Is there other documentation?
