### 2025-11-23 2b5940a
CompatHelper: bump compat for Cthulhu to 3, (keep existing compat)
