+++ Project.toml
@@ -81,7 +81,7 @@ CairoMakie = "0.13, 0.14, 0.15"
 ColorSchemes = "3.29.0"
 ColorTypes = "0.11, 0.12"
 ContentHashes = "0.1.0"
-Cthulhu = "2"
+Cthulhu = "2, 3"
 DataFrames = "1.7.0"
 DataStructures = "0.18.20, 0.19"
 Dates = "1.11.0"
