### 2025-11-23 9624289
CompatHelper: bump compat for JET to 0.11, (keep existing compat)
