### 2025-11-26 0859f16
NP code and plots - plots fro analytical formula
