+++ src/physics/osc.jl
@@ -789,8 +789,9 @@ function get_params(cfg::NND)  #'New'
     std = get_params(cfg.three_flavour)
     params = OrderedDict(pairs(std))
     params[:m₀] = ftype(0.01)
-    params[:N] = ftype(20)
+    params[:N] = ftype(100)
     params[:r] = ftype(1)
+    params[:eps] = ftype(1.01)
 
     
     NamedTuple(params)
@@ -800,8 +801,9 @@ function get_priors(cfg::NND)    #'New'
     std = get_priors(cfg.three_flavour)
     priors = OrderedDict{Symbol, Distribution}(pairs(std))
     priors[:m₀] = Uniform(ftype(1e-8),ftype(0.3)) #LogUniform(ftype(1e-3),ftype(1))
-    priors[:N] = Uniform(ftype(1),ftype(100))
+    priors[:N] = Uniform(ftype(1),ftype(50))
     priors[:r] = Uniform(ftype(0),ftype(1))
+    priors[:eps] = Uniform(ftype(1.0),ftype(2.0))
 
     NamedTuple(priors)
 end
@@ -824,8 +826,8 @@ function get_priors(cfg::NNM)    #'New'
     priors = OrderedDict(pairs(std))
     priors = OrderedDict{Symbol, Distribution}(pairs(std))
     #params[:scale]=Uniform(ftype(1e-6),ftype(100))
-    priors[:m₀] = Uniform(ftype(1e-6),ftype(1)) #Uniform(ftype(1e-7),ftype(2)) 
-    priors[:N] = Uniform(ftype(1),ftype(100))
+    priors[:m₀] = Uniform(ftype(1e-2),ftype(0.1)) #Uniform(ftype(1e-7),ftype(2)) 
+    priors[:N] = Uniform(ftype(1),ftype(200))
     priors[:r] = Uniform(ftype(0),ftype(1))
 
     NamedTuple(priors)
@@ -995,7 +997,7 @@ end
 
 
 
-function get_matrices(cfg::NND)
+function get_matrices_perturbation(cfg::NND)
 
    function get_Nnaturalness(params::NamedTuple)
         
@@ -1150,6 +1152,343 @@ function get_matrices(cfg::NND)
 
 end
 
+
+function get_matrices(cfg::NND)
+
+   function get_Nnaturalness(params::NamedTuple)
+        
+        N_int = round(Int, ForwardDiff.value(params[:N])) 
+        N_dual = params[:N]   
+        m0= params[:m₀]
+        r=params[:r]  
+        
+        eps=1+(1/N_dual) #params[:eps]#
+        
+        
+        T = promote_type(
+            typeof(params[:N]), 
+            typeof(params[:m₀]),
+            typeof(params[:r]), 
+            typeof(params[:Δm²₂₁]), 
+            typeof(params[:Δm²₃₁]),
+            typeof(params[:δCP]),
+            typeof(params[:θ₁₂]),
+            typeof(params[:θ₁₃]),
+            typeof(params[:θ₂₃])
+        ) 
+
+        m1, m2, m3 = get_abs_masses(params)
+
+        
+        m1_T = T(m1)
+        m2_T = T(m2) 
+        m3_T = T(m3)
+        
+        factor=(eps-1) * 2^(1/(N_dual-1))
+
+        if r>=0.0
+            scale_1= (m1_T ^2)/(r*factor)
+            scale_2= (m2_T ^2)/(r*factor)
+            scale_3= (m3_T ^2)/(r*factor)
+        end
+
+      
+
+        matrix_e=zeros(T, N_int,N_int)
+        matrix_m=zeros(T, N_int,N_int)
+        matrix_t=zeros(T, N_int,N_int)
[…]
