+++ src/experiments/gerda/gerda.jl
@@ -42,7 +42,7 @@ function get_assets(physics; datadir = @__DIR__)
 
     assets = (
 
-        observed = 0.9*1e26, #gerda 
+        observed = 1e28, #0.9*1e26, #gerda 
         # 1*1e28, #legend 1000
         
     )
@@ -320,7 +320,7 @@ function get_forward_model_correct(physics, assets)
     function forward_model(params)
     
         cfg = Newtrinos.osc.NNM()
-        observed = 0.9*1e26#legend 200
+        observed = 1e28#legend 200
         #predicted_value =get_neutrinomass(cfg)(params) #get_neutrinomass_SM(cfg)(params) 
         fun=get_halftime()
         predicted_value_T=fun(params)
@@ -328,7 +328,7 @@ function get_forward_model_correct(physics, assets)
         if predicted_value_T >= observed 
            predicted_value_T=observed
         end   
-        sigma= 0.01*1e26 #0.1*1e28 #
+        sigma= 0.01*1e28#0.1*1e28 #
         #println("Predicted m_nu: ", predicted_value_T)
        
         return Normal(predicted_value_T, sigma)
