+++ src/experiments/gerda/gerda.jl
@@ -199,20 +199,20 @@ function get_neutrinomass(cfg=NNM)
         N_m = length(mass_m)
         N_t = length(mass_t)
 
-        if any(mass_e .> 1e12) 
-            mass_e = mass_e[mass_e .<= 1e12]
+        if any(mass_e .> 1e6) 
+            mass_e = mass_e[mass_e .<= 1e6]
             N_e = length(mass_e)
             x_1_e = V_e[1, 1:N_e]
         end
 
-        if any(mass_m .> 1e12) 
-            mass_m = mass_m[mass_m .<= 1e12]
+        if any(mass_m .> 1e6) 
+            mass_m = mass_m[mass_m .<= 1e6]
             N_m = length(mass_m)
             x_1_m = V_m[1, 1:N_m]
         end
 
-        if any(mass_t .> 1e12) 
-            mass_t = mass_t[mass_t .<= 1e12]
+        if any(mass_t .> 1e6) 
+            mass_t = mass_t[mass_t .<= 1e6]
             N_t = length(mass_t)
             x_1_t = V_t[1, 1:N_t]
         end
