### 2025-12-02 143beac
osc plots
