### 2025-12-04 #9472 @philippeller
Hi @kyleniemeyer, thank you for your questions!
- I have now added the correct testing script in the "test" directory that runs the defined tests
- I have extended the README, which is currently all official documentation
