### 2025-12-07 decff20
CompatHelper: bump compat for JET to 0.11, (keep existing compat)
