+++ Project.toml
@@ -103,7 +103,7 @@ HDF5 = "0.17.2"
 Interpolations = "0.15.1, 0.16"
 InverseFunctions = "0.1.17"
 IterTools = "1.10.0"
-JET = "0.9.18, 0.10"
+JET = "0.9.18, 0.10, 0.11"
 JLD2 = "0.5.11, 0.6"
 KernelDensity = "0.6.10"
 LaTeXStrings = "1.4.0"
