+++ src/physics/osc.jl
@@ -801,7 +801,7 @@ function get_priors(cfg::NND)    #'New'
     std = get_priors(cfg.three_flavour)
     priors = OrderedDict{Symbol, Distribution}(pairs(std))
     priors[:m₀] = Uniform(ftype(1e-2),ftype(0.4)) #LogUniform(ftype(1e-3),ftype(1))
-    priors[:N] = Uniform(ftype(2),ftype(100))
+    priors[:N] = DiscreteUniform(ftype(2),ftype(40))
     priors[:r] = Uniform(ftype(1e-8),ftype(1))
     priors[:η] = Uniform(ftype(1.1),ftype(100))
 
@@ -827,7 +827,7 @@ function get_priors(cfg::NNM)    #'New'
     priors = OrderedDict(pairs(std))
     priors = OrderedDict{Symbol, Distribution}(pairs(std))
     priors[:m₀] = Uniform(ftype(1e-6),ftype(0.4)) #Uniform(ftype(1e-7),ftype(2)) 
-    priors[:N] = Uniform(ftype(2),ftype(40))
+    priors[:N] = DiscreteUniform(ftype(2),ftype(40))
     priors[:r] = Uniform(ftype(1e-8),ftype(1))
     priors[:η] = Uniform(ftype(1.000000001),ftype(1.1))
 
@@ -998,7 +998,7 @@ end
 
 
 
-function get_matrices_perturbation(cfg::NND)
+function get_matrices(cfg::NND) #_perturbation
 
    function get_Nnaturalness(params::NamedTuple)
         
@@ -1031,10 +1031,10 @@ function get_matrices_perturbation(cfg::NND)
         gamma= Vector{T}(undef, 3)
         
 
-        gamma[1] = (m1_T^2/m0^2)/N_dual
-        gamma[2] =(m2_T^2/m0^2)/N_dual
-        gamma[3] =(m3_T^2/m0^2)/N_dual
-        scale=N_dual*m0^2
+        gamma[1] = (m1_T^2/m0^2)#/N_dual
+        gamma[2] =(m2_T^2/m0^2)#/N_dual
+        gamma[3] =(m3_T^2/m0^2)#/N_dual
+        scale=m0^2/params[:r] #N_dual*m0^2
 
         gamma_sq = Vector{T}(undef, 3)
         gamma_sq[1]=gamma[1]^2
@@ -1063,13 +1063,13 @@ function get_matrices_perturbation(cfg::NND)
                 sqrt_j =sqrt(T(2*(j-1)) + T(params[:r]))
 
                 if i == j
-                    matrix_e[i, j] =sqrt_i * sqrt_j  + squared*gamma[1]
-                    matrix_m[i, j] =sqrt_i * sqrt_j  + squared*gamma[2]
-                    matrix_t[i, j] =sqrt_i * sqrt_j  + squared*gamma[3]
+                    matrix_e[i, j] =(1-N_dual*gamma[1])*sqrt_i * sqrt_j  + squared*gamma[1]
+                    matrix_m[i, j] =(1-N_dual*gamma[2])*sqrt_i * sqrt_j  + squared*gamma[2]
+                    matrix_t[i, j] =(1-N_dual*gamma[3])*sqrt_i * sqrt_j  + squared*gamma[3]
                 else
-                    matrix_e[i, j] =sqrt_i * sqrt_j 
-                    matrix_m[i, j] =sqrt_i * sqrt_j 
-                    matrix_t[i, j] =sqrt_i * sqrt_j 
+                    matrix_e[i, j] =(1-N_dual*gamma[1])*sqrt_i * sqrt_j 
[…]
