### 2025-12-16 60bdb09
gerda new weights for high modes
