+++ src/physics/osc.jl
@@ -788,10 +788,10 @@ end
 function get_params(cfg::NND)  #'New'
     std = get_params(cfg.three_flavour)
     params = OrderedDict(pairs(std))
-    params[:m₀] = ftype(0.01)
-    params[:N] = ftype(40)
+    params[:m₀] = ftype(0.1)
+    params[:N] = ftype(100)
     params[:r] = ftype(1)
-    params[:η] = ftype(1.01)
+    params[:η] = ftype(99)
 
     
     NamedTuple(params)
@@ -800,10 +800,10 @@ end
 function get_priors(cfg::NND)    #'New'
     std = get_priors(cfg.three_flavour)
     priors = OrderedDict{Symbol, Distribution}(pairs(std))
-    priors[:m₀] = Uniform(ftype(1e-2),ftype(0.4)) #LogUniform(ftype(1e-3),ftype(1))
+    priors[:m₀] = LogUniform(ftype(1e-6),ftype(1e-1))#Uniform(ftype(1e-2),ftype(0.4)) #LogUniform(ftype(1e-3),ftype(1))
     priors[:N] = DiscreteUniform(ftype(2),ftype(40))
     priors[:r] = Uniform(ftype(1e-8),ftype(1))
-    priors[:η] = Uniform(ftype(1.1),ftype(100))
+    priors[:η] =  Uniform(ftype(1.01),ftype(100))
 
     NamedTuple(priors)
 end
@@ -814,8 +814,8 @@ function get_params(cfg::NNM)  #'New'
     std = get_params(cfg.three_flavour)
     params = OrderedDict(pairs(std))
     #params[:scale]=ftype(1)
-    params[:m₀] = ftype(0.01)
-    params[:N] = ftype(20)
+    params[:m₀] = ftype(0.1)
+    params[:N] = ftype(100)
     params[:r] = ftype(1)
     params[:η] = ftype(1.1)
     
@@ -826,7 +826,7 @@ function get_priors(cfg::NNM)    #'New'
     std = get_priors(cfg.three_flavour)
     priors = OrderedDict(pairs(std))
     priors = OrderedDict{Symbol, Distribution}(pairs(std))
-    priors[:m₀] = Uniform(ftype(1e-6),ftype(0.4)) #Uniform(ftype(1e-7),ftype(2)) 
+    priors[:m₀] = LogUniform(ftype(1e-6),ftype(0.05))#Uniform(ftype(1e-6),ftype(0.4)) #Uniform(ftype(1e-7),ftype(2)) 
     priors[:N] = DiscreteUniform(ftype(2),ftype(40))
     priors[:r] = Uniform(ftype(1e-8),ftype(1))
     priors[:η] = Uniform(ftype(1.000000001),ftype(1.1))
@@ -998,7 +998,7 @@ end
 
 
 
-function get_matrices(cfg::NND) #_perturbation
+function get_matrices_perturbation(cfg::NND) #_perturbation
 
    function get_Nnaturalness(params::NamedTuple)
         
@@ -1027,14 +1027,16 @@ function get_matrices(cfg::NND) #_perturbation
         m2_T = T(m2) 
         m3_T = T(m3)
         
+        
+        scale=N_dual*m0^2/params[:r] #N_dual*m0^2
+
     
         gamma= Vector{T}(undef, 3)
         
 
-        gamma[1] = (m1_T^2/m0^2)#/N_dual
-        gamma[2] =(m2_T^2/m0^2)#/N_dual
-        gamma[3] =(m3_T^2/m0^2)#/N_dual
-        scale=m0^2/params[:r] #N_dual*m0^2
+        gamma[1] = m1_T^2/scale
+        gamma[2] = m2_T^2/scale
+        gamma[3] = m3_T^2/scale
 
         gamma_sq = Vector{T}(undef, 3)
[…]
