+++ src/experiments/gerda/gerda.jl
@@ -42,8 +42,8 @@ function get_assets(physics; datadir = @__DIR__)
 
     assets = (
 
-        observed = 1e28, #0.9*1e26, #gerda 
-        # 1*1e28, #legend 1000
+        observed =0.9 *1e26, #gerda 
+        # 1*1e28,, # #legend 1000
         
     )
     return assets
@@ -172,7 +172,78 @@ end
 
 
 
-function get_neutrinomass(cfg=NNM)
+function get_neutrinomass_STD(cfg=NNM(three_flavour=Newtrinos.osc.ThreeFlavour(ordering=:NO)))
+    function NeutrinoMassNNM_STD(params::NamedTuple)
+
+        
+
+        U= Newtrinos.osc.get_PMNS(params)
+
+        N = round(Int,params[:N])
+
+        func=  Newtrinos.osc.get_matrices(cfg)
+
+        final, h, eigen, V_e, V_m, V_t = func(params)
+        #final, h= func(params)
+        masses_NN= eigen
+
+        x_e = U[1,:]
+        x_1_e = V_e[1,1: N]
+        x_1_m = V_m[1,1: N]
+        x_1_t = V_t[1,1: N]
+
+        mass_e = eigen[1:3:end]      
+        mass_m = eigen[2:3:end]   
+        mass_t = eigen[3:3:end]      
+        N_e = length(mass_e)
+        N_m = length(mass_m)
+        N_t = length(mass_t)
+
+        #=if any(mass_e .> 1e12) 
+            mass_e = mass_e[mass_e .<= 1e12]
+            N_e = length(mass_e)
+            x_1_e = V_e[1, 1:N_e]
+        end
+
+        if any(mass_m .> 1e12) 
+            mass_m = mass_m[mass_m .<= 1e6]
+            N_m = length(mass_m)
+            x_1_m = V_m[1, 1:N_m]
+        end
+
+        if any(mass_t .> 1e12) 
+            mass_t = mass_t[mass_t .<= 1e12]
+            N_t = length(mass_t)
+            x_1_t = V_t[1, 1:N_t]
+        end=#
+
+        N = [N_e, N_m, N_t]
+        masses_NN = [mass_e, mass_m, mass_t]
+
+        X=[x_1_e, x_1_m, x_1_t]
+        sum=Float64(0.0)
+        
+        for i in 1:3
+            
+            
+            for j in 1:N[i]
+
+                mass = masses_NN[i][j]
+                integrand= abs((X[i][j].*(x_e[i])))^2 * sqrt(mass)
+                sum += integrand
+            end
+
+        end
+   
+
+        return sum
+     
+    end
+    return NeutrinoMassNNM
+end
+
+
+function get_neutrinomass(cfg=NNM(three_flavour=Newtrinos.osc.ThreeFlavour(ordering=:NO)))
     function NeutrinoMassNNM(params::NamedTuple)
 
         
@@ -199,23 +270,6 @@ function get_neutrinomass(cfg=NNM)
         N_m = length(mass_m)
         N_t = length(mass_t)
 
-        if any(mass_e .> 1e6) 
-            mass_e = mass_e[mass_e .<= 1e6]
-            N_e = length(mass_e)
-            x_1_e = V_e[1, 1:N_e]
-        end
-
-        if any(mass_m .> 1e6) 
-            mass_m = mass_m[mass_m .<= 1e6]
-            N_m = length(mass_m)
-            x_1_m = V_m[1, 1:N_m]
-        end
-
-        if any(mass_t .> 1e6) 
-            mass_t = mass_t[mass_t .<= 1e6]
-            N_t = length(mass_t)
-            x_1_t = V_t[1, 1:N_t]
[…]
