### 2025-12-17 bc9fbaf
Reparametrized, no base shape imposed. Alternative form-factors added. Backwards compatible, default Helm.
