+++ src/Newtrinos.jl
@@ -14,8 +14,8 @@ using .osc
 include("physics/earth_layers.jl")
 include("physics/atm_flux.jl")
 include("physics/xsec.jl")
-#include("physics/cevns_xsec.jl")
-#include("physics/sns_flux.jl")
+include("physics/cevns_xsec.jl")
+include("physics/sns_flux.jl")
 include("analysis/analysis_tools.jl")
 include("analysis/molewhacker.jl")
 include("utils/plotting.jl")
