+++ src/physics/cevns_xsec.jl
@@ -18,11 +18,11 @@ const mtau=1776.86
 const mpi = 139.57
 const alph = 1/137
 const ep= (mpi^2-mmu^2)/(2*mpi)
+const hcut_c = 197.326963 # MeV*fm
 
-# Only keep the dynamic configure (isotope_keys) and the (params, priors) version
-function configure(isotopes, er_centers, enu_centers)
-    # Build assets from isotopes
-    assets = get_assets(isotopes, er_centers, enu_centers)
+function configure(isotopes, er_centers, enu_centers; ff_model::Symbol = :helm, ff_kwargs::NamedTuple = (;))
+    # Build assets from isotopes (and store FF choice in assets)
+    assets = get_assets(isotopes, er_centers, enu_centers; ff_model = ff_model, ff_kwargs = ff_kwargs)
     params, priors = build_params_and_priors(isotopes)
     CevnsXsec(
         params = params,
@@ -43,16 +43,16 @@ end
 # Dynamic parameter/prior builder for isotope-specific Rn keys, using isotope list
 function build_params_and_priors(isotopes)
     param_dict = Dict(
-        :cevns_xsec_a => 0.0,
-        :cevns_xsec_b => 0.0,
-        :cevns_xsec_c => 0.0,
-        :cevns_xsec_d => 0.0,
+        :cevns_xsec_a => 1.0,
+        :cevns_xsec_b => -0.5,
+        :cevns_xsec_c => -1.0,
+        :cevns_xsec_d => 1.0,
         :sin2thetaW => 0.231,
     )
     prior_dict = Dict{Symbol, Distributions.Distribution}(
-        :cevns_xsec_a => Uniform(-2, 2),
-        :cevns_xsec_b => Uniform(-2000.0, 2000.0),
-        :cevns_xsec_c => Uniform(-3, 3),
+        :cevns_xsec_a => Uniform(0, 2),
+        :cevns_xsec_b => Uniform(-2.0, 1.0),
+        :cevns_xsec_c => Uniform(-3000, 3000),
         :cevns_xsec_d => Uniform(-1e6, 1e6),
         :sin2thetaW => truncated(Normal(0.231, 0.00013), 0.2, 0.26),
     )
@@ -63,8 +63,7 @@ function build_params_and_priors(isotopes)
     return ((; param_dict...), (; prior_dict...))
 end
 
-function get_assets(isotopes, er_centers, enu_centers)
-    # Extract isotope data into a structured format
+function get_assets(isotopes, er_centers, enu_centers; ff_model::Symbol = :helm, ff_kwargs::NamedTuple = (;))
     @info "Configuring CEvNS cross-section assets"
     isotope_data = Dict(iso.Rn_key => (
         mass = iso.mass,
@@ -74,60 +73,151 @@ function get_assets(isotopes, er_centers, enu_centers)
         Rn_nom = iso.Rn_nom
     ) for iso in isotopes)
 
-    # Return assets as a NamedTuple
     return (
-        isotopes = isotope_data,  # Dictionary of isotope data keyed by Rn_key
+        isotopes = isotope_data,
         er_centers = er_centers,
-        enu_centers = enu_centers
+        enu_centers = enu_centers,
+        ff_model = ff_model,
+        ff_kwargs = ff_kwargs,
     )
 end
 
-# Helm-like nuclear form factor squared, generic in type
-function ffsq(er, mn, rn)
-    r0 = rn / 197.326963
+# ---- Nuclear form factors ----------------------------------------------------
[…]
