### 2025-12-18 79e8c3c
data-driven Flux implemented for l-Ar
