### 2026-01-05 625a254
perturbation vs full eigenvalues
