+++ src/physics/osc.jl
@@ -998,7 +998,7 @@ end
 
 
 
-function get_matrices(cfg::NND) #_perturbation
+function get_matrices_p(cfg::NND) #_perturbation
 
    function get_Nnaturalness(params::NamedTuple)
         
@@ -1034,9 +1034,9 @@ function get_matrices(cfg::NND) #_perturbation
         gamma= Vector{T}(undef, 3)
         
 
-        gamma[1] = m1_T^2/scale#*params[:r]
-        gamma[2] = m2_T^2/scale#*params[:r]
-        gamma[3] = m3_T^2/scale#*params[:r]
+        gamma[1] = m1_T^2/(scale*params[:r])#m1_T^2/m0^2#scale#*params[:r]
+        gamma[2] = m2_T^2/(scale*params[:r])#scale#*params[:r]
+        gamma[3] = m3_T^2/(scale*params[:r])#scale#*params[:r]
 
         gamma_sq = Vector{T}(undef, 3)
         gamma_sq[1]=gamma[1]^2
@@ -1053,11 +1053,11 @@ function get_matrices(cfg::NND) #_perturbation
         for i in 1:N_int
             
             sqrt_i = sqrt(T(2*(i-1)) + T(params[:r]))
-            squared= one(T)
+            squared=sqrt_i^2 #one(T)
             
-            if i!=1
+            #=if i!=1
              squared=sqrt_i^2
-            end    
+            end =#   
            
             for j in 1:N_int
                 
@@ -1125,16 +1125,19 @@ function get_matrices(cfg::NND) #_perturbation
         #println("FinalUmatrix size: ", size(FinalUmatrix))
 
         delta_mass = Vector{T}(undef, 3*N_int)
-        delta_mass[1] =(scale)*(eigenvalues_e[1])- m1_T^2 #zero(T)
-        delta_mass[2] =(scale)*(eigenvalues_m[1])- (scale)*(eigenvalues_e[1]) #T(params.Δm²₂₁)#
-        delta_mass[3] = (scale)*(eigenvalues_t[1])- (scale)*(eigenvalues_e[1]) #T(params.Δm²₃₁)#
-
+        #delta_mass[1] =zero(T)#(scale)*(eigenvalues_e[1])- m1_T^2 
+        #delta_mass[2] =T(params.Δm²₂₁)#(scale)*(eigenvalues_m[1])- (scale)*(eigenvalues_e[1]) #
+        #delta_mass[3] = T(params.Δm²₃₁)#(scale)*(eigenvalues_t[1])- (scale)*(eigenvalues_e[1]) #
+      
+        delta_mass[1] =(scale)*(eigenvalues_e[1])- (scale)*(eigenvalues_e[1])#m1_T^2 
+        delta_mass[2] =(scale)*(eigenvalues_m[1])- (scale)*(eigenvalues_e[1]) #
+        delta_mass[3] =(scale)*(eigenvalues_t[1])- (scale)*(eigenvalues_e[1]) #
         
 
         for i in 2:N_int
-            delta_mass[3*i-2] =(scale)*(eigenvalues_e[i])- m1_T^2
-            delta_mass[3*i-1] =(scale)*(eigenvalues_m[i])- m1_T^2
-            delta_mass[3*i] = (scale)*(eigenvalues_t[i])- m1_T^2
+            delta_mass[3*i-2] =(scale)*(eigenvalues_e[i])- (scale)*(eigenvalues_e[1])
+            delta_mass[3*i-1] =(scale)*(eigenvalues_m[i])- (scale)*(eigenvalues_e[1])
+            delta_mass[3*i] = (scale)*(eigenvalues_t[i])- (scale)*(eigenvalues_e[1])
         end
         
[…]
