+++ plots_NP/pert_full/eigenvalues_pert_NND_ NO_100.0 1 0.01.txt
@@ -1,0 +1,300 @@
+0.000101
+0.000177
+0.002601
+0.000303
+0.000531
+0.007799
+0.000505
+0.000885
+0.012995
+0.000706
+0.001238
+0.018189
+0.000908
+0.001592
+0.023382
+0.001110
+0.001945
+0.028573
+0.001311
+0.002298
+0.033763
+0.001513
+0.002652
+0.038952
+0.001714
+0.003005
+0.044140
+0.001915
+0.003358
+0.049328
+0.002117
+0.003711
+0.054515
+0.002318
+0.004064
+0.059701
+0.002520
+0.004417
+0.064887
+0.002721
+0.004770
+0.070072
+0.002922
+0.005123
+0.075256
+0.003124
+0.005476
+0.080440
+0.003325
+0.005828
+0.085624
+0.003526
+0.006181
+0.090808
+0.003727
+0.006534
+0.095991
+0.003929
+0.006887
+0.101174
+0.004130
+0.007240
+0.106356
+0.004331
+0.007592
+0.111538
+0.004532
+0.007945
+0.116720
+0.004734
+0.008298
+0.121902
+0.004935
+0.008651
+0.127084
+0.005136
+0.009003
+0.132265
+0.005337
+0.009356
+0.137446
+0.005538
+0.009709
+0.142627
+0.005740
+0.010061
+0.147808
+0.005941
+0.010414
+0.152989
+0.006142
+0.010767
+0.158169
+0.006343
+0.011119
+0.163350
+0.006544
+0.011472
+0.168530
+0.006745
+0.011825
+0.173711
+0.006946
+0.012177
+0.178891
+0.007148
+0.012530
+0.184071
+0.007349
+0.012882
+0.189251
+0.007550
+0.013235
+0.194431
+0.007751
+0.013588
+0.199611
+0.007952
+0.013940
+0.204792
+0.008153
+0.014293
+0.209972
+0.008354
+0.014645
+0.215152
+0.008556
+0.014998
+0.220332
+0.008757
+0.015351
+0.225512
+0.008958
+0.015703
+0.230692
+0.009159
+0.016056
+0.235872
+0.009360
+0.016408
+0.241052
+0.009561
+0.016761
+0.246232
+0.009763
+0.017114
+0.251413
+0.009964
+0.017466
+0.256593
+0.010165
+0.017819
+0.261774
+0.010366
+0.018172
+0.266954
+0.010567
+0.018524
+0.272135
+0.010768
+0.018877
+0.277316
+0.010970
+0.019230
+0.282497
+0.011171
+0.019582
+0.287678
+0.011372
+0.019935
+0.292859
+0.011573
+0.020288
+0.298041
+0.011774
+0.020640
+0.303222
+0.011976
+0.020993
+0.308404
+0.012177
+0.021346
+0.313586
+0.012378
+0.021699
+0.318769
+0.012579
+0.022051
+0.323951
+0.012780
+0.022404
+0.329134
+0.012982
+0.022757
+0.334317
+0.013183
+0.023110
+0.339501
+0.013384
+0.023463
+0.344685
+0.013586
+0.023816
+0.349869
+0.013787
+0.024168
+0.355053
+0.013988
+0.024521
+0.360238
+0.014190
+0.024874
+0.365424
+0.014391
+0.025227
+0.370609
+0.014592
+0.025580
+0.375796
+0.014794
+0.025933
+0.380983
+0.014995
+0.026287
+0.386170
+0.015197
+0.026640
+0.391358
+0.015398
+0.026993
+0.396547
+0.015600
+0.027346
+0.401736
+0.015801
+0.027700
+0.406927
+0.016003
+0.028053
+0.412118
[…]

[…]
