### 2026-01-16 c73c58b
method fixing Lambda
