+++ src/physics/osc.jl
@@ -814,7 +814,7 @@ function get_params(cfg::NNM)  #'New'
     std = get_params(cfg.three_flavour)
     params = OrderedDict(pairs(std))
     #params[:scale]=ftype(1)
-    params[:m₀] = ftype(0.1)
+    params[:m₀] = ftype(0.01)
     params[:N] = ftype(100)
     params[:r] = ftype(1)
     params[:η] = ftype(1.1)
@@ -1165,9 +1165,17 @@ function get_matrices(cfg::NND) #full
         N_int = round(Int, ForwardDiff.value(params[:N])) 
         N_dual = params[:N]   
         r=params[:r]  
+        m0= params[:m₀]
+        Lambda= 1e4*1e9  #cutoff scale eV
+        l=0.05#higgs coupling
+
+        b=1/(N_dual)^4 #choice of b
         
-        η=1+(1/N_dual)#
+
+
+        η=1+ (m0^2*l*N_dual)/(Lambda^2*b^2*r)  #(1+ (m0^2 * l * N_dual)/(Lambda^2 * b^2))  #eta parameter
         
+        println("Eta value: ", η)
         
         T = promote_type(
             typeof(params[:N]), 
@@ -1330,7 +1338,7 @@ end
 
 
 
-function get_matrices(cfg::NNM)
+function get_matrices(cfg::NNM) #full
 
    function get_Nnaturalness(params::NamedTuple)
 
@@ -1340,6 +1348,16 @@ function get_matrices(cfg::NNM)
         N_dual = params[:N]   
     
         r=params[:r]   
+
+        m0= params[:m₀]
+        Lambda= 1e4*1e9 #cutoff scale eV
+        l=0.05#higgs coupling
+        ms=100*1e9 #rehaton mass eV
+        b=1/sqrt(N_dual) #choice of b
+        
+
+
+        η=1+(m0*l*N_dual*ms)/(Lambda^2*b^2*r)  #(1+ (m0^2 * l * N_dual)/(Lambda^2 * b^2))  #eta parameter
         
         
         T = promote_type(
@@ -1361,7 +1379,7 @@ function get_matrices(cfg::NNM)
         m2_T = T(m2) 
         m3_T = T(m3)
         
-        η=1+(N_dual)#params[:η]#1+N_dual#1+(1/N_dual)
+        η=1+(1/N_dual)#params[:η]#1+N_dual#1+(1/N_dual)
 
         factor=(η-1) * 2^(1/(N_dual-1))
 
@@ -1463,7 +1481,7 @@ function get_matrices(cfg::NNM)
             #println("First three delta masses set to: ", delta_mass[1], ", ", delta_mass[2], ", ", delta_mass[3])
         else
 
-            delta_mass[1] =((eigenvalues_e[1])*scale_1)^2- m1_T^2 #zero(T)
+            delta_mass[1] =((eigenvalues_e[1])*scale_1)^2-((eigenvalues_e[1])*scale_1)^2#((eigenvalues_e[1])*scale_1)^2- m1_T^2 #zero(T)
             delta_mass[2] =((eigenvalues_m[1])*scale_2)^2- ((eigenvalues_e[1])*scale_1)^2 #T(params.Δm²₂₁)#
             delta_mass[3] =((eigenvalues_t[1])*scale_3)^2- ((eigenvalues_e[1])*scale_1)^2 #T(params.Δm²₃₁)#
         
@@ -1473,9 +1491,9 @@ function get_matrices(cfg::NNM)
             
         
             for i in 2:N_int
-                delta_mass[3*i-2] =((eigenvalues_e[i])*scale_1)^2- m1_T^2
[…]
