### 2026-01-21 #9644 @rafaelab
I have finally finished reviewing the code and the paper.

## The code

I have several comments about the code

- The main problem is the lack of adequate documentation and examples.
I spent several hours trying to make things work and when they did, trying to make sense of the results. I tried checking the source code and found it to be messy in many places.

- I ran the example from the README file, as well as the files test.jl for several (but not all) experiments. They work as expected, but they are too complex to count as tests. They are more like examples than tests, and they are not easy to find.

- The “documentation” (which essentially does not exist) does not introduce the user to the main parts of the code. For instance, I would have liked to test the physics modules separately. Since there are no unit tests, this turned out to be time-consuming without clear examples to guide me (and, honestly, I did not want to invest more time digging into the code).
If there is a mistake, for instance, in the neutrino oscillation part, there are no tests and it takes the user too much time to find the right functions and figure out how they work.

- From a design point of view, there are far too many dependencies, many of which are not needed in the core code, I guess. Maybe some of them could be moved to the sub-modules.

- My concrete suggestions are to:
. implement unit tests;
. improve documentation, including descriptions of functions;
. provide basic examples that do not require running the full analysis;
. add guidelines for contributors.

## The paper

I find this paper quite underwhelming. The statement of need is extremely weak, with no framing of this work within the broader context.
It lists the main features, but it falls short of clearly explaining to non-expert readers (e.g., beginning graduate students and people from related fields) what the code actually does.
