### 2026-01-21 #9644 @matthewfeickert
@rafaelab thanks for your review. Please open up Issues in https://github.com/philippeller/Newtrinos.jl that are scoped to each comment area (generally prefecated with "JOSS Review:") to have discussion in the GitHub repository. When doing so, please mention `openjournals/joss-reviews#9644` so that a link is created to this Issue thread.
