### 2026-01-22 c66cad4
plots Lambda_H
