+++ src/physics/osc.jl
@@ -788,7 +788,7 @@ end
 function get_params(cfg::NND)  #'New'
     std = get_params(cfg.three_flavour)
     params = OrderedDict(pairs(std))
-    params[:m₀] = ftype(0.01)
+    params[:m₀] = ftype(0.1)
     params[:N] = ftype(100)
     params[:r] = ftype(1)
     params[:η] = ftype(99)
@@ -1005,7 +1005,14 @@ function get_matrices_p(cfg::NND) #_perturbation
         N_int = round(Int, ForwardDiff.value(params[:N])) 
         N_dual = params[:N]   
         m0= params[:m₀]
-           
+ 
+        r=params[:r]  
+      
+        Lambda= 1e4*1e9  #cutoff scale eV
+        l=0.05#higgs coupling
+
+        dif= (m0^2*l*N_dual)/(Lambda^2*r) 
+       
         
         
         T = promote_type(
@@ -1028,7 +1035,7 @@ function get_matrices_p(cfg::NND) #_perturbation
         m3_T = T(m3)
         
         
-        scale=N_dual*m0^2/params[:r] #N_dual*m0^2
+        scale= dif *m0^2/r #N_dual*m0^2/params[:r] 
 
     
         gamma= Vector{T}(undef, 3)
@@ -1169,13 +1176,13 @@ function get_matrices(cfg::NND) #full
         Lambda= 1e4*1e9  #cutoff scale eV
         l=0.05#higgs coupling
 
-        b=1/(N_dual)^4 #choice of b
+        b=1/sqrt(N_dual) #choice of b
         
 
 
         η=1+ (m0^2*l*N_dual)/(Lambda^2*b^2*r)  #(1+ (m0^2 * l * N_dual)/(Lambda^2 * b^2))  #eta parameter
         
-        println("Eta value: ", η)
+        #println("Eta value: ", η)
         
         T = promote_type(
             typeof(params[:N]), 
@@ -1353,11 +1360,12 @@ function get_matrices(cfg::NNM) #full
         Lambda= 1e4*1e9 #cutoff scale eV
         l=0.05#higgs coupling
         ms=100*1e9 #rehaton mass eV
-        b=1/sqrt(N_dual) #choice of b
-        
-
-
-        η=1+(m0*l*N_dual*ms)/(Lambda^2*b^2*r)  #(1+ (m0^2 * l * N_dual)/(Lambda^2 * b^2))  #eta parameter
+       
+        η=1+(1/N_dual)
+        b=(m0*l*N_dual*ms)/(Lambda^2*r*(η-1))  #choice of b
+       
+        #b=1/sqrt(N_dual) #choice of b 
+        #η=1+(m0*l*N_dual*ms)/(Lambda^2*b^2*r)  #(1+ (m0^2 * l * N_dual)/(Lambda^2 * b^2))  #eta parameter
         
         
         T = promote_type(
@@ -1379,7 +1387,7 @@ function get_matrices(cfg::NNM) #full
         m2_T = T(m2) 
         m3_T = T(m3)
         
-        η=1+(1/N_dual)#params[:η]#1+N_dual#1+(1/N_dual)
+        #η=1+(1/N_dual)#params[:η]#1+N_dual#1+(1/N_dual)
 
         factor=(η-1) * 2^(1/(N_dual-1))
 
@@ -1513,6 +1521,180 @@ function get_matrices(cfg::NNM) #full
 end
 
 
+
+
+function get_matrice_p(cfg::NNM) #perturbation
+   function get_Nnaturalness(params::NamedTuple)
+        
+        N_int = round(Int, ForwardDiff.value(params[:N])) 
+        N_dual = params[:N]    
[…]
