### 2026-01-25 33d9911
CompatHelper: bump compat for Optim to 2, (keep existing compat)
