+++ Project.toml
@@ -124,7 +124,7 @@ Memoize = "0.4.4"
 MonotonicSplines = "0.3.3"
 Mooncake = "0.4.180"
 NPZ = "0.4.3"
-Optim = "1.11.0"
+Optim = "1.11.0, 2"
 Optimization = "4.1.1, 5"
 PCHIPInterpolation = "0.2.1"
 PDMats = "0.11"
