### 2026-01-26 #9644 @philippeller
Thank you @rafaelab for your review!
I do agree with some of these concerns, and we'll be working to address them and get back with a detailed reply. It might take a little while, as we will need to implement some additions to the repository (mostly more tests and better documentation).
