### 2026-01-30 87669fd
new anyltical eigenvalues_all_comp_NNM_NO_20
