+++ minos_test.jl
@@ -1,29 +1,20 @@
 
-# %%
 using LinearAlgebra
 using Distributions
-#using Plots
 using LaTeXStrings
 using Printf
 using FileIO
 import JLD2
 
-# %%
 using DataFrames
-#using StatsPlots
 
-# %%
-#using Revise
 using Newtrinos
 using Newtrinos.osc
 
-# %%
-#using Pkg
-#Pkg.status() 
 
-# %%
+
 osc_cfg = Newtrinos.osc.OscillationConfig(
-    flavour=Newtrinos.osc.NND(),
+    flavour=Newtrinos.osc.NNM(),
     propagation=Newtrinos.osc.Basic(),
     states=Newtrinos.osc.All(),
     interaction=Newtrinos.osc.SI()
@@ -31,7 +22,6 @@ osc_cfg = Newtrinos.osc.OscillationConfig(
 
 osc = Newtrinos.osc.configure(osc_cfg)
 
-# %%
 
 atm_flux = Newtrinos.atm_flux.configure()
 earth_layers = Newtrinos.earth_layers.configure()
@@ -39,21 +29,15 @@ xsec=Newtrinos.xsec.configure()
 
 physics = (; osc, atm_flux, earth_layers, xsec);
 
-# %%
+
 experiments = (
  
-   minos= Newtrinos.minos.configure(physics),
+   dayabay= Newtrinos.dayabay.configure(physics),
 );
 
-# %%
 p = Newtrinos.get_params(experiments)
 
-# %%
-img = experiments.minos.plot(p)
-#display("image/png", img)
-#save("/home/sofialon/Newtrinos.jl/natural plot/minos/minos_data_NND_20.png", img)
 
-# %%
 
 all_priors = Newtrinos.get_priors(experiments)
 
@@ -61,7 +45,7 @@ all_priors = Newtrinos.get_priors(experiments)
 vars_to_scan = (r=31,N=31)  
 
 modified_priors = (
-    N = all_priors.N, 
+    N = DiscreteUniform(2, 200),
     m₀= p.m₀,
     r =all_priors.r,
     
@@ -79,24 +63,11 @@ modified_priors = (
 )
 
 
-# %%
 
 likelihood = Newtrinos.generate_likelihood(experiments);
 
 
-# %%
-
 result = Newtrinos.scan(likelihood, modified_priors, vars_to_scan, p)
 
-# %%
-JLD2.@save "scan_minos_rN_NND.jld2" result
-
-# %%
-using CairoMakie
-
-# %%
-img = CairoMakie.plot(result)
-#display("image/png", img)
-save("/home/sofialon/Newtrinos.jl/natural plot/minos/minos_rN_NND.png", img)
-
+JLD2.@save "/home/sofialon/Newtrinos.jl/plots_fix/dayabay/scan_dayabay_Nr_m0=0.01_NNM_NO.jld2" result
 
