+++ plots_fix/eigen/h_pert_NNM_NO_20.0 0.5 0.01.txt
@@ -1,0 +1,60 @@
+0.000000
+0.000005
+0.000170
+0.000163
+0.000292
+0.004364
+0.000542
+0.000955
+0.014077
+0.001133
+0.001991
+0.029270
+0.001937
+0.003399
+0.049920
+0.002952
+0.005179
+0.076019
+0.004179
+0.007329
+0.107563
+0.005617
+0.009850
+0.144556
+0.007267
+0.012743
+0.187003
+0.009130
+0.016007
+0.234919
+0.011205
+0.019645
+0.288322
+0.013495
+0.023659
+0.347238
+0.015999
+0.028049
+0.411707
+0.018722
+0.032821
+0.481781
+0.021664
+0.037980
+0.557542
+0.024832
+0.043533
+0.639115
+0.028233
+0.049495
+0.726722
+0.031886
+0.055898
+0.820817
+0.035839
+0.062828
+0.922711
+62.914695
+63.596259
+71.823354
