+++ plots_fix/eigen/h_fullr1_NNM_NO_20.0 0.01 0.01.txt
@@ -1,0 +1,60 @@
+0.000000
+0.000078
+0.002549
+4.108210
+7.201769
+105.801274
+16.241582
+28.471571
+418.272009
+36.317794
+63.665170
+935.294692
+64.297219
+112.713102
+1655.848822
+100.159282
+175.579299
+2579.404536
+143.895111
+252.248207
+3705.733338
+195.504613
+342.719663
+5034.832836
+254.995446
+447.007094
+6566.900262
+322.383138
+565.137718
+8302.335493
+397.692198
+697.154500
+10241.769715
+480.958434
+843.120213
+12386.125105
+572.233106
+1003.124713
+14736.721740
+671.590314
+1177.297898
+17295.467906
+779.140823
+1365.833941
+20065.216175
+895.060390
+1569.040941
+23050.492766
+1019.656458
+1787.457848
+26259.215303
+1153.564410
+2022.198487
+29707.746787
+1298.623022
+2276.486235
+33443.441230
+1448.179074
+2538.657995
+37294.958249

[…]
