+++ src/experiments/katrin/katrin_NN.jl
@@ -108,7 +108,7 @@ function get_posterior_SM(params)
 
 end    
 
-# function to get m0 posterior from m_nu posterior in NND-NNM
+# function to get m0 posterior from m_nu posterior in NNM-NNM
 
 
 function get_posterior_NN(params, cfg)
@@ -233,7 +233,7 @@ end
 
 
 function get_neutrinomass_old(cfg=NNM)
-    function NeutrinoMassNND_old(params::NamedTuple)
+    function NeutrinoMassNNM_old(params::NamedTuple)
 
         U= Newtrinos.osc.get_PMNS(params)
 
@@ -315,7 +315,7 @@ function get_neutrinomass_old(cfg=NNM)
         return sum
      
     end
-    return NeutrinoMassNND
+    return NeutrinoMassNNM
 end
 
 
@@ -499,7 +499,7 @@ end
 
 
 function get_neutrinomass_new(cfg=NNM)
-    function NeutrinoMassNND_new(params::NamedTuple)
+    function NeutrinoMassNNM_new(params::NamedTuple)
 
         U= Newtrinos.osc.get_PMNS(params)
 
@@ -582,7 +582,7 @@ function get_neutrinomass_new(cfg=NNM)
 
      
     end
-    return NeutrinoMassNND
+    return NeutrinoMassNNM
 end
 
 
