+++ plots_fix/eigen/eigenvalues_full_NNM_NO_20.0 0.5 0.01.txt
@@ -1,60 +1,0 @@
-0.000102
-0.000180
-0.002640
-0.002535
-0.004444
-0.065291
-0.008165
-0.014313
-0.210267
-0.016964
-0.029738
-0.436882
-0.028921
-0.050698
-0.744791
-0.044027
-0.077178
-1.133815
-0.062279
-0.109176
-1.603882
-0.083680
-0.146691
-2.155006
-0.108231
-0.189729
-2.787276
-0.135940
-0.238303
-3.500860
-0.166816
-0.292429
-4.296017
-0.200875
-0.352133
-5.173121
-0.238136
-0.417452
-6.132708
-0.278630
-0.488438
-7.175547
-0.322401
-0.565169
-8.302789
-0.369520
-0.647769
-9.516257
-0.420112
-0.736456
-10.819139
-0.474431
-0.831678
-12.218027
-0.533220
-0.934734
-13.732010
-0.609464
-1.068390
-15.695514
