+++ plots_fix/eigen/eigenvalues_fullr1_NNM_NO_20.0 0.01 0.01.txt
@@ -1,60 +1,0 @@
-0.000103
-0.000181
-0.002652
-4.108313
-7.201872
-105.801377
-16.241685
-28.471674
-418.272112
-36.317897
-63.665273
-935.294795
-64.297322
-112.713205
-1655.848925
-100.159385
-175.579402
-2579.404639
-143.895214
-252.248310
-3705.733441
-195.504716
-342.719766
-5034.832939
-254.995549
-447.007197
-6566.900365
-322.383241
-565.137821
-8302.335596
-397.692301
-697.154603
-10241.769818
-480.958537
-843.120316
-12386.125208
-572.233209
-1003.124816
-14736.721843
-671.590417
-1177.298001
-17295.468009
-779.140926
-1365.834044
-20065.216277
-895.060493
-1569.041044
-23050.492869
-1019.656561
-1787.457951
-26259.215406
-1153.564512
-2022.198590
-29707.746890
-1298.623125
-2276.486338
-33443.441333
-1448.179177
-2538.658098
-37294.958352
