+++ plots_fix/eigen/h_full_NNM_NO_20.0 0.5 0.01.txt
@@ -1,60 +1,0 @@
-0.000000
-0.000077
-0.002537
-0.002433
-0.004342
-0.065189
-0.008062
-0.014210
-0.210164
-0.016862
-0.029636
-0.436779
-0.028818
-0.050595
-0.744688
-0.043924
-0.077076
-1.133712
-0.062177
-0.109073
-1.603780
-0.083577
-0.146588
-2.154904
-0.108129
-0.189627
-2.787174
-0.135837
-0.238200
-3.500757
-0.166714
-0.292326
-4.295914
-0.200772
-0.352031
-5.173019
-0.238033
-0.417349
-6.132605
-0.278527
-0.488335
-7.175445
-0.322298
-0.565066
-8.302687
-0.369418
-0.647667
-9.516154
-0.420009
-0.736353
-10.819036
-0.474329
-0.831575
-12.217925
-0.533117
-0.934632
-13.731908
-0.609361
-1.068287
-15.695412

[…]
