### 2026-02-03 23e5c40
plota 90% CL
