+++ plots_fix/eigen/h_pertnew_NNM_NO_20.0 0.5 0.01.txt
@@ -1,0 +1,60 @@
+0.000000
+0.000005
+0.000171
+0.000164
+0.000292
+0.004382
+0.000542
+0.000955
+0.014129
+0.001134
+0.001993
+0.029364
+0.001938
+0.003402
+0.050064
+0.002953
+0.005182
+0.076218
+0.004181
+0.007334
+0.107820
+0.005619
+0.009856
+0.144871
+0.007270
+0.012749
+0.187378
+0.009133
+0.016015
+0.235352
+0.011209
+0.019654
+0.288811
+0.013499
+0.023668
+0.347778
+0.016004
+0.028060
+0.412291
+0.018726
+0.032832
+0.482402
+0.021669
+0.037991
+0.558187
+0.024837
+0.043544
+0.639769
+0.028238
+0.049507
+0.727363
+0.031890
+0.055908
+0.821411
+0.035842
+0.062837
+0.923199
+22663.216705
+22676.060223
+22825.068040
