+++ plots_fix/eigen/h_pertnew_NNM_NO_100.0 0.5 0.01.txt
@@ -1,0 +1,300 @@
+0.000000
+0.000005
+0.000158
+0.000153
+0.000273
+0.004094
+0.000509
+0.000898
+0.013272
+0.001069
+0.001879
+0.027686
+0.001832
+0.003216
+0.047331
+0.002798
+0.004910
+0.072204
+0.003967
+0.006959
+0.102302
+0.005339
+0.009363
+0.137623
+0.006913
+0.012123
+0.178165
+0.008690
+0.015238
+0.223925
+0.010670
+0.018709
+0.274904
+0.012852
+0.022534
+0.331098
+0.015237
+0.026715
+0.392507
+0.017824
+0.031250
+0.459131
+0.020614
+0.036141
+0.530967
+0.023606
+0.041386
+0.608015
+0.026800
+0.046986
+0.690275
+0.030197
+0.052940
+0.777746
+0.033796
+0.059249
+0.870427
+0.037598
+0.065913
+0.968317
+0.041602
+0.072932
+1.071417
+0.045808
+0.080305
+1.179726
+0.050216
+0.088033
+1.293243
+0.054826
+0.096115
+1.411969
+0.059639
+0.104551
+1.535903
+0.064654
+0.113343
+1.665044
+0.069871
+0.122488
+1.799393
+0.075291
+0.131988
+1.938950
+0.080912
+0.141843
+2.083714
+0.086736
+0.152052
+2.233685
+0.092762
+0.162616
+2.388864
+0.098990
+0.173534
+2.549251
+0.105421
+0.184806
+2.714845
+0.112053
+0.196433
+2.885646
+0.118888
+0.208415
+3.061656
+0.125926
+0.220751
+3.242874
+0.133165
+0.233441
+3.429299
+0.140606
+0.246487
+3.620933
+0.148250
+0.259886
+3.817776
+0.156097
+0.273640
+4.019828
+0.164145
+0.287749
+4.227090
+0.172396
+0.302213
+4.439561
+0.180849
+0.317031
+4.657242
+0.189504
+0.332204
+4.880135
+0.198362
+0.347731
+5.108238
+0.207422
+0.363613
+5.341553
+0.216684
+0.379851
+5.580080
+0.226149
+0.396442
+5.823821
+0.235817
+0.413389
+6.072775
+0.245686
+0.430691
+6.326944
+0.255759
+0.448348
+6.586328
+0.266034
+0.466360
+6.850928
+0.276511
+0.484726
+7.120745
+0.287191
+0.503449
+7.395780
+0.298074
+0.522526
+7.676033
+0.309159
+0.541959
+7.961508
+0.320447
+0.561747
+8.252203
+0.331938
+0.581890
+8.548121
+0.343632
+0.602389
+8.849263
+0.355529
+0.623244
+9.155630
+0.367628
+0.644455
+9.467225
+0.379931
+0.666021
+9.784048
+0.392437
+0.687944
+10.106101
+0.405146
+0.710222
+10.433387
+0.418058
+0.732857
+10.765907
+0.431173
+0.755849
+11.103664
+0.444492
+0.779197
+11.446660
+0.458014
+0.802901
+11.794898
+0.471741
+0.826963
+12.148380
+0.485670
+0.851382
+12.507110
+0.499804
+0.876158
+12.871091
+0.514142
+0.901292
+13.240326
+0.528684
+0.926784
+13.614820
+0.543430
+0.952634
+13.994576
+0.558381
+0.978843
+14.379600
+0.573536
+1.005411
+14.769897
+0.588896
+1.032337
+15.165473
+0.604462
+1.059624
+15.566334
+0.620233
+1.087271
+15.972488
+0.636210
+1.115278
+16.383943
+0.652393
[…]
