+++ plots_fix/eigen/eigenvalues_pertnew_NNM_NO_20.0 0.5 0.01.txt
@@ -1,0 +1,60 @@
+0.000007
+0.000012
+0.000177
+0.000170
+0.000299
+0.004389
+0.000549
+0.000962
+0.014136
+0.001141
+0.002000
+0.029371
+0.001945
+0.003409
+0.050071
+0.002960
+0.005189
+0.076224
+0.004187
+0.007340
+0.107827
+0.005626
+0.009863
+0.144878
+0.007277
+0.012756
+0.187385
+0.009140
+0.016022
+0.235359
+0.011216
+0.019661
+0.288818
+0.013506
+0.023675
+0.347785
+0.016011
+0.028067
+0.412298
+0.018733
+0.032839
+0.482409
+0.021676
+0.037998
+0.558194
+0.024844
+0.043551
+0.639776
+0.028245
+0.049514
+0.727369
+0.031897
+0.055915
+0.821418
+0.035849
+0.062844
+0.923206
+22663.216712
+22676.060230
+22825.068046

[…]
