+++ plots_fix/eigen/eigenvalues_pert9_NNM_NO_20.0 1 0.01.txt
@@ -1,0 +1,60 @@
+0.000109
+0.000191
+0.002776
+0.000976
+0.001709
+0.024842
+0.002698
+0.004724
+0.068762
+0.005269
+0.009229
+0.134445
+0.008687
+0.015217
+0.221846
+0.012951
+0.022688
+0.330955
+0.018060
+0.031640
+0.461781
+0.024015
+0.042075
+0.614352
+0.030817
+0.053994
+0.788716
+0.038468
+0.067402
+0.984944
+0.046971
+0.082303
+1.203130
+0.056330
+0.098705
+1.443404
+0.066551
+0.116619
+1.705945
+0.077642
+0.136060
+1.991002
+0.089616
+0.157049
+2.298939
+0.102493
+0.179621
+2.630324
+0.116307
+0.203837
+2.986137
+0.131128
+0.229822
+3.368370
+0.147161
+0.257936
+3.782709
+17.019538
+17.753706
+27.715242
