+++ plots_fix/eigen/eigenvalues_full9_NNM_NO_20.0 1 0.01.txt
@@ -1,0 +1,60 @@
+0.000102
+0.000179
+0.002628
+0.000910
+0.001596
+0.023443
+0.002515
+0.004409
+0.064767
+0.004911
+0.008608
+0.126462
+0.008094
+0.014189
+0.208452
+0.012065
+0.021149
+0.310702
+0.016821
+0.029488
+0.433197
+0.022364
+0.039204
+0.575943
+0.028694
+0.050301
+0.738967
+0.035814
+0.062781
+0.922311
+0.043725
+0.076649
+1.126042
+0.052431
+0.091911
+1.350255
+0.061938
+0.108577
+1.595085
+0.072253
+0.126659
+1.860727
+0.083387
+0.146178
+2.147470
+0.095359
+0.167164
+2.455774
+0.108199
+0.189672
+2.786441
+0.121971
+0.213816
+3.141131
+0.136864
+0.239922
+3.524653
+5989.388498
+10499.398038
+154244.721999
