+++ plots_fix/eigen/h_pert9_NNM_NO_20.0 1 0.01.txt
@@ -1,0 +1,60 @@
+0.000000
+0.000082
+0.002667
+0.000867
+0.001600
+0.024732
+0.002588
+0.004615
+0.068653
+0.005160
+0.009119
+0.134335
+0.008578
+0.015108
+0.221737
+0.012842
+0.022579
+0.330846
+0.017951
+0.031531
+0.461671
+0.023906
+0.041965
+0.614242
+0.030708
+0.053885
+0.788607
+0.038359
+0.067292
+0.984834
+0.046862
+0.082194
+1.203020
+0.056221
+0.098596
+1.443295
+0.066442
+0.116510
+1.705836
+0.077533
+0.135950
+1.990892
+0.089507
+0.156939
+2.298829
+0.102384
+0.179512
+2.630215
+0.116198
+0.203728
+2.986027
+0.131019
+0.229713
+3.368261
+0.147052
+0.257827
+3.782600
+17.019429
+17.753596
+27.715133
