+++ plots_fix/eigen/h_full9_NNM_NO_20.0 1 0.01.txt
@@ -1,0 +1,60 @@
+0.000000
+0.000077
+0.002526
+0.000808
+0.001494
+0.023341
+0.002413
+0.004307
+0.064665
+0.004809
+0.008506
+0.126360
+0.007992
+0.014087
+0.208350
+0.011963
+0.021047
+0.310600
+0.016719
+0.029386
+0.433095
+0.022262
+0.039102
+0.575841
+0.028592
+0.050199
+0.738865
+0.035712
+0.062679
+0.922209
+0.043623
+0.076547
+1.125940
+0.052329
+0.091809
+1.350153
+0.061836
+0.108475
+1.594983
+0.072151
+0.126557
+1.860625
+0.083285
+0.146076
+2.147368
+0.095257
+0.167062
+2.455672
+0.108097
+0.189570
+2.786339
+0.121869
+0.213714
+3.141029
+0.136762
+0.239820
+3.524551
+5989.388396
+10499.397936
+154244.721897
