+++ plots_fix/eigen/eigenvalues_pert9_NNM_NO_20.0 0.5 0.01.txt
@@ -1,0 +1,60 @@
+0.000007
+0.000012
+0.000176
+0.000170
+0.000298
+0.004370
+0.000549
+0.000961
+0.014082
+0.001140
+0.001998
+0.029273
+0.001943
+0.003406
+0.049922
+0.002959
+0.005185
+0.076019
+0.004185
+0.007336
+0.107561
+0.005624
+0.009857
+0.144551
+0.007274
+0.012749
+0.186996
+0.009136
+0.016014
+0.234909
+0.011212
+0.019652
+0.288310
+0.013501
+0.023665
+0.347224
+0.016006
+0.028056
+0.411691
+0.018728
+0.032828
+0.481764
+0.021671
+0.037986
+0.557524
+0.024839
+0.043539
+0.639097
+0.028240
+0.049502
+0.726704
+0.031892
+0.055904
+0.820800
+0.035845
+0.062835
+0.922698
+58.560172
+59.217910
+67.169423
