+++ plots_fix/eigen/h_pert9_NNM_NO_20.0 1.0e-8 0.01.txt
@@ -1,0 +1,60 @@
+0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+-0.000000
+0.000000
+0.000000
+0.000000
+134239345404518464.000000
+134239345404518464.000000
+134239345404518576.000000
