+++ plots_fix/eigen/h_pert9_NNM_NO_20.0 0.5 0.01.txt
@@ -1,0 +1,60 @@
+0.000000
+0.000005
+0.000170
+0.000163
+0.000292
+0.004363
+0.000542
+0.000954
+0.014076
+0.001133
+0.001991
+0.029266
+0.001936
+0.003399
+0.049915
+0.002952
+0.005178
+0.076012
+0.004178
+0.007329
+0.107554
+0.005617
+0.009850
+0.144544
+0.007267
+0.012742
+0.186989
+0.009129
+0.016007
+0.234902
+0.011205
+0.019645
+0.288303
+0.013494
+0.023658
+0.347217
+0.015999
+0.028049
+0.411684
+0.018721
+0.032821
+0.481757
+0.021664
+0.037979
+0.557517
+0.024832
+0.043532
+0.639090
+0.028233
+0.049495
+0.726697
+0.031885
+0.055898
+0.820793
+0.035839
+0.062828
+0.922691
+58.560165
+59.217903
+67.169416
