+++ plots_fix/eigen/eigenvalues_full9_NNM_NO_20.0 1.0e-8 0.01.txt
@@ -1,0 +1,60 @@
+0.000103
+0.000181
+0.002652
+4067767400876.358887
+7130796253736.253906
+104757213874768.828125
+16161464095889.607422
+28331046560094.476562
+416206184861444.875000
+36198386462185.484375
+63455771468211.132812
+932217046560662.375000
+64138770828864.132812
+112435265262998.796875
+1651765765155736.750000
+99961971699339.156250
+175233336388941.468750
+2574320657173079.500000
+143659073574640.937500
+251834355976345.437500
+3699652121767725.500000
+195229954506988.375000
+342238110250750.625000
+5027757018418469.000000
+254682252020801.406250
+446457987792464.625000
+6558832036291694.000000
+322031477024568.312500
+564521179224068.125000
+8293276627813705.000000
+397302124840489.062500
+696470624845377.250000
+10231721621017110.000000
+480529991472516.812500
+842369075051321.625000
+12375088870391718.000000
+571766323566449.750000
+1002306365211986.250000
+14724698130806770.000000
+671085206736708.000000
+1176412367409448.750000
+17282457329090434.000000
+778597389983974.625000
+1364881224641907.000000
+20051218584257288.000000
+894478603011328.375000
+1568020991078858.000000
+23035507463350732.000000
+1019036246952766.125000
+1786370540908198.000000
+26243240467774568.000000
+1152905611654720.250000
+2021043537230724.000000
+29690778216943996.000000
+1297925185411035.750000
+2275262850025545.000000
+33425467299890384.000000
+54063915093902532608.000000
+94774043159611097088.000000
+1392308005413271109632.000000
