### 2026-02-26 #9644 @szabo137
Finally, I was able to spend some time here. I'm very sorry for the delay.

First of all, many thanks for providing the package and paper. I  highly appreciate the efforts of putting complicated analysis pipelines into Julia packages to make them accessible to a wider audience. The following review is separated into the paper part and the package part.

## Paper review
Unfortunately, the paper itself is a bit on the shorter side. The *technical limitations* could be described in more detail to understand the need for the presented package, and how it helps to overcome these shortcomings of other approaches. This makes the key features presented hard to connect to and appreciate. Especially the claim that the *different approach* taken here is *open-source, modular, and computationally efficient* sounds very generic in this context. Finally, the review of the related development efforts correctly lists packages that try to fulfill similar purposes. However, it is not fully clear here what the presented software adds to this list, especially since the listed software, *GAMBIT*, seems to provide similar functionality.

I propose to make the shortcomings in the current neutrino analysis toolchain clearer in the statement of need, explicitly point to the features of the presented software, which try to solve these, and present a more balanced perspective in the *Related Work* part. This will help the reader to better understand the software and judge the usability in their own work.

## Package review
Looking at the code, the package seems to have gone through a decent number of development cycles, which shows substantial scholarly effort. The readme provides an example of an analysis done with the package, which serves as an entry point. However, I was wondering if a more quickstart-like section with a more rudimentary example and a bit more explanation would make the expected learning curve less steep. Caused by the lack of documentation, it is very hard to infer the purpose and/or usage of the functions presented in the readme. Therefore, I propose to shorten the example in the readme significantly and move the more complex examples to a dedicated section in the docs. Furthermore, having at least an API reference section in the docs would also be very much appreciated. Additionally, adding some contribution guidelines would help to interact with your development process.

Regarding the tests, in the package, there are examples implemented, which can be checked by hand to see if they produce reasonable results or reproduce known ones. However, having more fine-grained unit-tests, which test not only the results, but also the expected functionality, would serve at least two purposes: 1) users could see how the code is used, 2) developers could check if the code still works after changes. Therefore, I propose to at least add a minimal suite of unit tests in the package.

## Summary
Regarding the paper, the author presents a good starting point, but should enrich the material a bit, as suggested in the review above.
Regarding the package, the two major concerns are the missing docs and unit tests, and the minor issue is in parts of the example in the readme, which could be a more concise quickstart section.
After addressing these issues, I think the submission is eligible for publication in JOSS.
