### 2026-02-26 #9644 @philippeller
Thank you @szabo137 for your review!

We will be working to address the raised issues, which are mostly shared concerns among both reviewers.
As already noted above, it will take a little time to implement this (There is actually a working student right now tasked with implementing unit tests and better documentation)

We will let you know once we have a revised draft and, importantly, an updated repo.
