### 2026-03-04 #9644 @matthewfeickert
@rafaelab, @szabo137, @philippeller to keep the review tracking issue from becoming increasingly complex to follow (given that not all notifications are relevant to all participants) please have all relevant discussions on issues or questions for the review in GitHub Issues on https://github.com/philippeller/Newtrinos.jl/issues that reference this Issue (`https://github.com/openjournals/joss-reviews/issues/9644`). It would be great to see https://github.com/openjournals/joss-reviews/issues/9644#issuecomment-3965984312 ported over as Issues (like @rafaelab did with https://github.com/philippeller/Newtrinos.jl/issues/143) and probably split out into multiple more specific ones as needed. Iteration can happen as needed and then once things have concluded please report the status here.

> As already noted above, it will take a little time to implement this

@philippeller Do you have an estimated timeline for us to set `editorialbot` reminders to check back?
