+++ src/physics/osc.jl
@@ -815,7 +815,7 @@ function get_params(cfg::NNM)  #'New'
     params = OrderedDict(pairs(std))
    
     params[:m₀] = ftype(0.01)
-    params[:N] = ftype(20)
+    params[:N] = ftype(50)
     params[:r] = ftype(1)
     params[:η] = ftype(1+1/params[:N])
     
@@ -1165,7 +1165,7 @@ function get_matrices_p(cfg::NND) #_perturbation
 end
 
 
-function get_matrices(cfg::NND) #full
+function get_matrices_f(cfg::NND) #full
 
    function get_Nnaturalness(params::NamedTuple)
         
@@ -1181,7 +1181,7 @@ function get_matrices(cfg::NND) #full
 
 
         #η=1+ (m0^2*l*N_dual)/(Lambda^2*b^2*r)  #(1+ (m0^2 * l * N_dual)/(Lambda^2 * b^2))  #eta parameter
-        η=params[:η]
+        η=1+ 1/N_dual #params[:η]
         #println("Eta value: ", η)
         
         T = promote_type(
@@ -1358,6 +1358,178 @@ function get_matrices(cfg::NND) #full
 end
 
 
+function get_matrices(cfg::NND) #full schur NND
+
+   function get_Nnaturalness(params::NamedTuple)
+        
+        N_int = round(Int, ForwardDiff.value(params[:N])) 
+        N_dual = params[:N]   
+        r=params[:r]  
+        m0= params[:m₀]
+        Lambda= 1e4*1e9  #cutoff scale eV
+        l=0.05#higgs coupling
+
+        b=1/sqrt(N_dual) #choice of b
+        
+        η=1+1/N_dual #params[:η]
+        
+        T = promote_type(
+            typeof(params[:N]), 
+            typeof(params[:m₀]),
+            typeof(params[:r]), 
+            typeof(params[:Δm²₂₁]), 
+            typeof(params[:Δm²₃₁]),
+            typeof(params[:δCP]),
+            typeof(params[:θ₁₂]),
+            typeof(params[:θ₁₃]),
+            typeof(params[:θ₂₃])
+        ) 
+
+        m1, m2, m3 = get_abs_masses(params)
+
+        m1_T = T(m1)
+        m2_T = T(m2) 
+        m3_T = T(m3)
+        
+        factor=(η-1) * 2^(1/(N_dual-1)) 
+        factorN=factor
+
+        if r >= zero(T)
+            scale_1= (m1_T ^2)/(r*factor)
+            scale_2= (m2_T ^2)/(r*factor)
+            scale_3= (m3_T ^2)/(r*factor)
+        end
+
+        matrix_e=zeros(T, N_int,N_int)
+        matrix_m=zeros(T, N_int,N_int)
+        matrix_t=zeros(T, N_int,N_int)
+    
+        for i in 1:N_int
+            sqrt_i = sqrt(T(2*(i-1)) + T(params[:r]))
+            
+            for j in 1:N_int
+                sqrt_j =sqrt(T(2*(j-1)) + T(params[:r]))
+
+                if i == j
+                    matrix_e[i, j] =sqrt_i * sqrt_j *η
+                    matrix_m[i, j] =sqrt_i * sqrt_j *η
+                    matrix_t[i, j] =sqrt_i * sqrt_j *η
+                else
+                    matrix_e[i, j] =sqrt_i * sqrt_j 
+                    matrix_m[i, j] =sqrt_i * sqrt_j 
+                    matrix_t[i, j] =sqrt_i * sqrt_j 
+                end
+            end
+        end
+
+        # PMNS matrix 
+        U = get_PMNS(params)
+
+        eigenvalues_e, V_e = eigen(Hermitian(matrix_e))
[…]
