### 2026-03-15 c505b4b
CompatHelper: bump compat for Optim to 2, (keep existing compat)
