### 2026-03-19 b816e43
Add Super-K atmospheric neutrino systematics: xsec shapes, PID migration, flux model, split energy scale
Expand Super-K from ~30 to ~50 nuisance parameters to better approximate
the 193 systematic sources in the official SK analysis (PRD 109, 072014).
Contours were too small due to missing systematics over-constraining the fit.

Phase 1 (xsec.jl): Energy-dependent M_A^QE and M_A^Res shape params via
dipole form factor ratios, isospin-1/2 background (I12) on CCother, FSI
cross-channel migration between 1p1h↔1pi. Widened cc2p2h (100%) and
nutau_cc (25%) priors per paper.

Phase 2 (super_k.jl): P
[…]
