+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -226,6 +226,23 @@ function get_assets(physics; datadir = @__DIR__)
         sk_i_v_multigev_multiring_nuebar = occursin.("sk1-5_fc_multigev_multiring_nuebarlike", bininfo.Sample),
         sk_i_v_multigev_multiring_mu = occursin.("sk1-5_fc_multigev_multiring_mulike", bininfo.Sample),
         sk_i_v_multigev_multiring_other = occursin.("sk1-5_fc_multigev_multiring_other", bininfo.Sample),
+        # PID migration masks
+        sk_i_iii_subgev_elike = occursin.(r"sk1-3_fc_subgev_1ring_elike", bininfo.Sample),
+        sk_i_iii_subgev_mulike = occursin.(r"sk1-3_fc_subgev_1ring_mulike", bininfo.Sample),
+        sk_iv_v_subgev_elike = occursin.(r"sk4-5_fc_subgev_1ring_nue", bininfo.Sample),
+        sk_iv_v_subgev_mulike = occursin.(r"sk4-5_fc_subgev_1ring_numu", bininfo.Sample),
+        sk_i_iii_multigev_1ring_elike = occursin.(r"sk1-3_fc_multigev_1ring_(elike|nue)", bininfo.Sample),
+        sk_i_iii_multigev_1ring_mulike = occursin.(r"sk1-3_fc_multigev_1ring_(mulike|numu)", bininfo.Sample),
+        sk_iv_v_multigev_1ring_elike = occursin.(r"sk4-5_fc_multigev_1ring_nue", bininfo.Sample),
+        sk_iv_v_multigev_1ring_mulike = occursin.(r"sk4-5_fc_multigev_1ring_numu", bininfo.Sample),
+        # Ring counting masks
+        sk_1ring = occursin.(r"_1ring_", bininfo.Sample),
+        sk_multiring = occursin.(r"_(2ring|multiring)_", bininfo.Sample),
+        # E-like mask (for nue contamination and NC pi0)
+        sk_elike = occursin.(r"(elike|nuebarlike)", bininfo.Sample) .| occursin.(r"_nuelike", bininfo.Sample),
+        # SK phase masks (for split energy scale)
+        sk_i_iii_bins = occursin.(r"^sk1-3_", bininfo.Sample),
+        sk_iv_v_bins = .!occursin.(r"^sk1-3_", bininfo.Sample),
     )
 
     data = CSV.read(joinpath(datadir, "bins/sk_2023_Data.txt"), DataFrame; delim=' ', ignorerepeated=true, comment="#", header=false)
@@ -282,7 +299,8 @@ end
     
 function get_params()
     params = (
-        sk_energy_scale = 1.0,
+        sk_i_iii_energy_scale = 1.0,
+        sk_iv_v_energy_scale = 1.0,
         sk_updown_energy_scale = 1.0,
         sk_fc_norm = 1.0,
         sk_pc_norm = 1.0,
@@ -299,12 +317,20 @@ function get_params()
         sk_i_v_btd_1 = 1.0,
         sk_i_v_btd_2 = 1.0,
         sk_i_v_btd_3 = 1.0,
+        sk_i_iii_subgev_pid = 1.0,
+        sk_iv_v_subgev_pid = 1.0,
+        sk_i_iii_multigev_pid = 1.0,
+        sk_iv_v_multigev_pid = 1.0,
+        sk_ring_counting = 1.0,
+        sk_nue_contamination = 1.0,
+        sk_ncpi0_norm = 1.0,
         )
 end
 
 function get_priors()
     priors = (
-        sk_energy_scale = Normal(1.0, 0.025),
+        sk_i_iii_energy_scale = Normal(1.0, 0.03),
+        sk_iv_v_energy_scale = Normal(1.0, 0.018),
[…]
