+++ test/test_regression.jl
@@ -10,7 +10,7 @@ const ALL_REFERENCE_VALUES = (
     kamland      = -63.111403860037875,
     minos        = -268.3363280363475,
     deepcore     = -950.7063452304332,
-    super_k      = -3706.345109466843,
+    super_k      = -3706.3369078597534,
     orca         = -1164.2506083927215,
     coherent_csi = -574.3416032522271,
     coherent_lAr = -1754.9946939034153,

[…]
