### 2026-03-19 2e0039f
Auto-select PolyesterForwardDiff for large parameter counts
Add select_ad(n_params) helper that uses AutoPolyesterForwardDiff (threaded
chunked ForwardDiff) when n_params > 12 (one chunk), otherwise plain
AutoForwardDiff. Applied in find_mle, molewhacker, and analysis CLI.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
