### 2026-03-19 f7b778a
Extend Mooncake reverse-mode AD support to all 8 experiments
Replace dynamic NamedTuple(generator) patterns with static map/explicit
construction (ORCA, Super-K, xsec.jl), replace vcat-based meshgrid
splatting with vec() (atm_flux.jl), and precompute ORCA nominal flux
at configuration time to avoid file I/O during differentiation.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
