+++ test/test_mooncake.jl
@@ -1,0 +1,193 @@
+using Mooncake
+
+@testset "Mooncake Reverse-Mode AD" begin
+
+    @testset "Vacuum oscillation gradient" begin
+        osc = Newtrinos.osc.configure()
+        E = [2.0]
+        L = [300.0]
+
+        function vacuum_prob(x)
+            p = NamedTuple{keys(osc.params)}(Tuple(x))
+            P = osc.osc_prob(E, L, p)
+            P[1, 1, 2, 2]  # P(νμ→νμ)
+        end
+
+        x0 = collect(Float64, values(osc.params))
+        rule = Mooncake.build_rrule(vacuum_prob, x0)
+        val, (_, dx) = Mooncake.value_and_gradient!!(rule, vacuum_prob, x0)
+
+        @test isfinite(val)
+        @test all(isfinite, dx)
+
+        # Verify against finite differences
+        eps_fd = 1e-7
+        for i in eachindex(x0)
+            x_plus = copy(x0); x_plus[i] += eps_fd
+            x_minus = copy(x0); x_minus[i] -= eps_fd
+            fd = (vacuum_prob(x_plus) - vacuum_prob(x_minus)) / (2eps_fd)
+            @test dx[i] ≈ fd atol=1e-4
+        end
+    end
+
+    @testset "Matter oscillation gradient (SI)" begin
+        cfg = Newtrinos.osc.OscillationConfig(interaction=Newtrinos.osc.SI())
+        osc_si = Newtrinos.osc.configure(cfg)
+        el = Newtrinos.earth_layers.configure()
+        layers = el.compute_layers()
+        paths = el.compute_paths([-0.5], layers)
+        E = [5.0]
+
+        function matter_prob(x)
+            p = NamedTuple{keys(osc_si.params)}(Tuple(x))
+            P = osc_si.osc_prob(E, paths, layers, p)
+            P[1, 1, 2, 2]
+        end
+
+        x0 = collect(Float64, values(osc_si.params))
+        rule = Mooncake.build_rrule(matter_prob, x0)
+        val, (_, dx) = Mooncake.value_and_gradient!!(rule, matter_prob, x0)
+
+        @test isfinite(val)
+        @test all(isfinite, dx)
+
+        # Verify against finite differences
+        eps_fd = 1e-7
+        for i in eachindex(x0)
+            x_plus = copy(x0); x_plus[i] += eps_fd
+            x_minus = copy(x0); x_minus[i] -= eps_fd
+            fd = (matter_prob(x_plus) - matter_prob(x_minus)) / (2eps_fd)
+            @test dx[i] ≈ fd rtol=1e-4
+        end
+    end
+
+    @testset "Sterile matter oscillation gradient (DefaultEigen)" begin
+        cfg = Newtrinos.osc.OscillationConfig(
+            flavour=Newtrinos.osc.Sterile(),
+            interaction=Newtrinos.osc.SI(),
+            eigen_method=Newtrinos.osc.DefaultEigen()
+        )
+        osc_st = Newtrinos.osc.configure(cfg)
+        el = Newtrinos.earth_layers.configure()
+        layers = el.compute_layers()
+        paths = el.compute_paths([-0.5], layers)
+        E = [5.0]
+
+        function sterile_matter_prob(x)
+            p = NamedTuple{keys(osc_st.params)}(Tuple(x))
+            P = osc_st.osc_prob(E, paths, layers, p)
+            P[1, 1, 2, 2]
+        end
+
+        x0 = collect(Float64, values(osc_st.params))
+        rule = Mooncake.build_rrule(sterile_matter_prob, x0)
+        val, (_, dx) = Mooncake.value_and_gradient!!(rule, sterile_matter_prob, x0)
+
+        @test isfinite(val)
[…]
