### 2026-03-20 1e34483
Add Mooncake support for JUNO, TAO, and IC Upgrade (Asimov data)
Future experiments lack observed data, so tests use generate_asimov_data
to create unfluctuated expectations. IC Upgrade gets the same fixes as
ORCA: static NamedTuple construction and precomputed nominal flux.
Mooncake tests now build all 11 rules in parallel via Threads.@spawn.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
