+++ test/test_mooncake.jl
@@ -96,98 +96,172 @@ using Mooncake
         end
     end
 
-    @testset "DeepCore full likelihood gradient" begin
-        dc = Newtrinos.deepcore.configure()
-        params = Newtrinos.get_params((deepcore=dc,))
-        likelihood = Newtrinos.generate_likelihood((deepcore=dc,))
-
-        function neg_ll_dc(x)
-            p = NamedTuple{keys(params)}(Tuple(x))
-            Float64(-log(likelihood(p)))
-        end
-
-        x0 = collect(Float64, values(params))
-        rule = Mooncake.build_rrule(neg_ll_dc, x0)
-        val, (_, dx) = Mooncake.value_and_gradient!!(rule, neg_ll_dc, x0)
-
-        @test isfinite(val)
-        @test all(isfinite, dx)
-        @test val ≈ 950.706345 rtol=1e-6
-
-        # Spot check a few gradients against finite differences
-        eps_fd = 1e-6
-        for i in [1, length(x0)÷2, length(x0)]  # first, middle, last param
-            scale = max(1.0, abs(x0[i]))
-            x_plus = copy(x0); x_plus[i] += eps_fd * scale
-            x_minus = copy(x0); x_minus[i] -= eps_fd * scale
-            fd = (neg_ll_dc(x_plus) - neg_ll_dc(x_minus)) / (2eps_fd * scale)
-            @test dx[i] ≈ fd rtol=1e-4
-        end
-    end
-
-    # Helper to test a full experiment likelihood gradient
-    function test_experiment_gradient(name, exp_tuple, ref_val)
-        params = Newtrinos.get_params(exp_tuple)
-        likelihood = Newtrinos.generate_likelihood(exp_tuple)
-
-        neg_ll = let params=params, likelihood=likelihood
-            x -> begin
-                p = NamedTuple{keys(params)}(Tuple(x))
-                Float64(-log(likelihood(p)))
-            end
-        end
-
-        x0 = collect(Float64, values(params))
+    # Helper: build rule, compute gradient, check finite, check value, spot-check against FD
+    function check_mooncake_gradient(neg_ll, x0, ref_val; rtol_val=1e-6, rtol_grad=1e-4)
         rule = Mooncake.build_rrule(neg_ll, x0)
         val, (_, dx) = Mooncake.value_and_gradient!!(rule, neg_ll, x0)
 
         @test isfinite(val)
         @test all(isfinite, dx)
-        @test val ≈ ref_val rtol=1e-6
+        @test val ≈ ref_val rtol=rtol_val
 
-        # Spot check gradients against finite differences
         eps_fd = 1e-6
         for i in [1, length(x0)÷2, length(x0)]
             scale = max(1.0, abs(x0[i]))
             x_plus = copy(x0); x_plus[i] += eps_fd * scale
             x_minus = copy(x0); x_minus[i] -= eps_fd * scale
             fd = (neg_ll(x_plus) - neg_ll(x_minus)) / (2eps_fd * scale)
-            @test dx[i] ≈ fd rtol=1e-4
+            @test dx[i] ≈ fd rtol=rtol_grad
         end
     end
 
+    # Configure all experiments up front (sequential, I/O heavy)
+    dc = Newtrinos.deepcore.configure()
+    db = Newtrinos.dayabay.configure()
+    kl = Newtrinos.kamland.configure()
+    mi = Newtrinos.minos.configure()
+    or = Newtrinos.orca.configure()
+    sk = Newtrinos.super_k.configure()
[…]
