### 2026-03-20 1b92604
Add --ad CLI option to select AD backend (forwarddiff, polyester, mooncake)
Introduces set_ad_backend/select_ad API with global backend setting.
The analysis CLI now accepts --ad auto|forwarddiff|polyester|mooncake,
which propagates to all find_mle calls. Default remains :auto (FD for
small param counts, PolyesterFD for large). Also adds bench_ad_comparison.jl
benchmark script comparing all three backends.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
