### 2026-03-20 abf76e0
Fix AD backend selection being overridden by hardcoded ForwardDiff default
Remove the global set_batcontext(ad = AutoForwardDiff()) at module load time
that was overriding user's --ad selection. Propagate the AD backend choice to
Distributed workers, which previously always used ForwardDiff. Also run
set_batcontext on the main process unconditionally (not just in non-distributed
mode).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
