+++ Project.toml
@@ -52,6 +52,7 @@ Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 MeasureBase = "fa1605e6-acd5-459c-a1e6-7e635759db14"
 Memoize = "c03570c3-d221-55d1-a50c-7939bbd78826"
 MonotonicSplines = "568f7cb4-8305-41bc-b90d-d32b39cc99d1"
+Mooncake = "da2b9cff-9c12-43a0-ae48-6db2b0edb7d6"
 NPZ = "15e1cf62-19b3-5cfa-8e77-841668bca605"
 Optim = "429524aa-4258-5aef-a3af-852621145aeb"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
@@ -121,6 +122,7 @@ Makie = "0.21, 0.22, 0.23, 0.24"
 MeasureBase = "0.14.13"
 Memoize = "0.4.4"
 MonotonicSplines = "0.3.3"
+Mooncake = "0.5.23"
 NPZ = "0.4.3"
 Optim = "1.11.0"
 Optimization = "4.8.0"
