### 2026-03-20 09a567a
making things work with reverse mode AD (mooncake)
