+++ src/experiments/icecube/deepcore_9y_verification_sample/deepcore.jl
@@ -220,14 +220,22 @@ function reweight(params, physics, assets)
     p = physics.osc.osc_prob(assets.binning.e_fine, assets.paths, assets.layers, params)
     p_flux = reshape(sys_flux.nue, s) .* p[:, :, 1, :] .+ reshape(sys_flux.numu, s) .* p[:, :, 2, :]
     
-    nus = NamedTuple(ch=>gather_flux(p_flux, assets.mc[ch].ef_idx, assets.mc[ch].cf_idx, i) for (i, ch) in enumerate([:nue_cc, :numu_cc, :nutau_cc]))
+    nus = (
+        nue_cc = gather_flux(p_flux, assets.mc.nue_cc.ef_idx, assets.mc.nue_cc.cf_idx, 1),
+        numu_cc = gather_flux(p_flux, assets.mc.numu_cc.ef_idx, assets.mc.numu_cc.cf_idx, 2),
+        nutau_cc = gather_flux(p_flux, assets.mc.nutau_cc.ef_idx, assets.mc.nutau_cc.cf_idx, 3),
+    )
 
     nu_ncs = (nu_nc = gather_flux_nc(reshape(sys_flux.nue .+ sys_flux.numu, s),  assets.mc[:nu_nc].ef_idx, assets.mc[:nu_nc].cf_idx),)
     
     p = physics.osc.osc_prob(assets.binning.e_fine, assets.paths, assets.layers, params, anti=true)
     p_flux = reshape(sys_flux.nuebar, s) .* p[:, :, 1, :] .+ reshape(sys_flux.numubar, s) .* p[:, :, 2, :]
 
-    nubars = NamedTuple(ch=>gather_flux(p_flux, assets.mc[ch].ef_idx, assets.mc[ch].cf_idx, i) for (i, ch) in enumerate([:nuebar_cc, :numubar_cc, :nutaubar_cc]))
+    nubars = (
+        nuebar_cc = gather_flux(p_flux, assets.mc.nuebar_cc.ef_idx, assets.mc.nuebar_cc.cf_idx, 1),
+        numubar_cc = gather_flux(p_flux, assets.mc.numubar_cc.ef_idx, assets.mc.numubar_cc.cf_idx, 2),
+        nutaubar_cc = gather_flux(p_flux, assets.mc.nutaubar_cc.ef_idx, assets.mc.nutaubar_cc.cf_idx, 3),
+    )
 
     nubar_ncs = (nubar_nc = gather_flux_nc(reshape(sys_flux.nue .+ sys_flux.numu, s),  assets.mc[:nubar_nc].ef_idx, assets.mc[:nubar_nc].cf_idx),)
     
@@ -262,10 +270,10 @@ function apply_hypersurfaces(hists, params, physics, assets)
     f_numu_cc = get_hypersurface_factor(assets.hypersurfaces.numu_cc, idx, fraction, params)
     f_nutau_cc = get_hypersurface_factor(assets.hypersurfaces.nutau_cc, idx, fraction, params) 
 
-    nu_nc = hists[:nu_nc] .+ hists[:nubar_nc]
-    nue_cc = hists[:nue_cc] .+ hists[:nuebar_cc]
-    numu_cc = hists[:numu_cc] .+ hists[:numubar_cc]
-    nutau_cc = hists[:nutau_cc] .+ hists[:nutaubar_cc]
+    nu_nc = hists.nu_nc .+ hists.nubar_nc
+    nue_cc = hists.nue_cc .+ hists.nuebar_cc
+    numu_cc = hists.numu_cc .+ hists.numubar_cc
+    nutau_cc = hists.nutau_cc .+ hists.nutaubar_cc
             
     return (
     nu_nc .* f_nu_nc_nue_cc .* physics.xsec.scale(:Any, :NC, params) .+ 
[…]
