+++ test/runtests.jl
@@ -1,6 +1,8 @@
 using Newtrinos
 using Test
 
+@info "Running tests with $(Threads.nthreads()) threads"
+
 @testset "Newtrinos.jl" begin
     include("test_helpers.jl")
     include("test_osc.jl")
@@ -9,4 +11,10 @@ using Test
     include("test_analysis.jl")
     include("test_autodiff.jl")
     include("test_regression.jl")
+
+    if get(ENV, "NEWTRINOS_TEST_MOONCAKE", "") == "1"
+        include("test_mooncake.jl")
+    else
+        @info "Skipping Mooncake tests (set NEWTRINOS_TEST_MOONCAKE=1 to enable)"
+    end
 end
