+++ test/test_regression.jl
@@ -2,6 +2,7 @@ using Test
 using Newtrinos
 using DensityInterface
 using Printf
+
 # All reference likelihood values at default parameters.
 # If a value changes beyond rtol=1e-6, it indicates a regression.
 # To update: run the experiment and replace the reference value.
@@ -37,33 +38,61 @@ end
 selected = select_experiments()
 
 @testset "Likelihood Regression" begin
-    # Collect results for summary table
-    results = NamedTuple{(:name, :ref, :actual, :rdiff, :status), Tuple{Symbol, Float64, Float64, Float64, Symbol}}[]
+    # Run all experiments in parallel using threads, then check results
+    exp_names = collect(keys(selected))
+    n_exps = length(exp_names)
 
-    for (name, ref_llh) in pairs(selected)
-        @testset "$name" begin
+    # Storage for results (thread-safe: each index written by one task)
+    actual_llh = Vector{Float64}(undef, n_exps)
+    ref_llh = Vector{Float64}(undef, n_exps)
+
+    # Configure, evaluate likelihood for each experiment in parallel
+    t_start = time()
+    timings = Vector{Float64}(undef, n_exps)
+    @sync for idx in 1:n_exps
+        Threads.@spawn begin
+            t0 = time()
+            name = exp_names[idx]
+            ref_llh[idx] = selected[name]
             mod = getproperty(Newtrinos, name)
             exp = mod.configure()
             experiments = (; name => exp)
             params = Newtrinos.get_params(experiments)
             likelihood = Newtrinos.generate_likelihood(experiments)
-            llh = logdensityof(likelihood, params)
+            actual_llh[idx] = logdensityof(likelihood, params)
+            timings[idx] = time() - t0
+            @info "  $(name) done on thread $(Threads.threadid()) in $(round(timings[idx], digits=1))s"
+        end
+    end
+    t_total = time() - t_start
+    t_sequential = sum(timings)
+    @info @sprintf("Regression tests: %.1fs wall time (%.1fs sequential, %.1fx speedup from %d threads)",
+                   t_total, t_sequential, t_sequential / t_total, Threads.nthreads())
 
-            diff = llh - ref_llh
-            rdiff = abs(diff) / abs(ref_llh)
+    # Now run the assertions on the main thread
+    results = NamedTuple{(:name, :ref, :actual, :rdiff, :status), Tuple{Symbol, Float64, Float64, Float64, Symbol}}[]
+
+    for idx in 1:n_exps
+        name = exp_names[idx]
+        llh = actual_llh[idx]
+        ref = ref_llh[idx]
+
+        @testset "$name" begin
+            diff = llh - ref
+            rdiff = abs(diff) / abs(ref)
 
             if rdiff > 1e-6
-                if llh > ref_llh
-                    push!(results, (name=name, ref=ref_llh, actual=llh, rdiff=rdiff, status=:improved))
-                    @warn "$name likelihood improved: $ref_llh → $llh (rdiff=$rdiff). Update reference value."
-                    @test_broken llh ≈ ref_llh rtol=1e-6
+                if llh > ref
+                    push!(results, (name=name, ref=ref, actual=llh, rdiff=rdiff, status=:improved))
+                    @warn "$name likelihood improved: $ref → $llh (rdiff=$rdiff). Update reference value."
+                    @test_broken llh ≈ ref rtol=1e-6
                 else
-                    push!(results, (name=name, ref=ref_llh, actual=llh, rdiff=rdiff, status=:regressed))
-                    @test llh ≈ ref_llh rtol=1e-6
+                    push!(results, (name=name, ref=ref, actual=llh, rdiff=rdiff, status=:regressed))
[…]
