+++ src/analysis/analysis.jl
@@ -145,7 +145,7 @@ end
 
 priors = Newtrinos.condition(priors, conditional_vars, p)
 
-@reset priors.Δm²₃₁ = Uniform(0.0018, 0.0028)
+@reset priors.Δm²₃₁ = Uniform(0.002, 0.003)
 @reset priors.θ₂₃ = Uniform(pi/4-0.2, pi/4+0.2)
 
 if lowercase(args["task"]) == "nestedsampling"
