### 2026-03-20 1552c2c
Fix distributed profiling crash with Mooncake AD
Workers were receiving a serialized likelihood closure from the main
process instead of using their own locally-built likelihood. This fails
with Mooncake (and is wasteful). Now _profile passes only serializable
data to map_func, and workers reference their local likelihood via an
@everywhere-defined function.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
