### 2026-03-20 27b1da8
Fix distributed profiling: zero serialization of complex objects
Workers now build everything locally (likelihood, scanpoints, AD context)
from scratch. Only integer indices are sent via pmap — no closures,
function chains, or distribution products cross process boundaries.
Reverts _profile back to original signature (threaded-only use).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
