+++ src/analysis/analysis.jl
@@ -61,64 +61,93 @@ n_workers = args["workers"]
 n_threads = args["threads"]
 use_distributed = n_workers > 1
 
-# Set up distributed workers if requested
-map_func = nothing
-if use_distributed
+"""
+    _distributed_profile(args, conditional_vars, prior_overrides, vars_to_scan, cache_dir, n_workers, n_threads)
+
+Run profiling on distributed workers. Each worker builds its own likelihood, scanpoints,
+and AD context locally — nothing is serialized except integer indices and scalar results.
+`prior_overrides` is a Dict{Symbol,Distribution} of prior replacements to apply after conditioning.
+"""
+function _distributed_profile(args, conditional_vars, prior_overrides, vars_to_scan, cache_dir, n_workers, n_threads)
+    t1 = time()
+
     addprocs(n_workers; exeflags="--threads=$n_threads")
 
+    # Send only plain data to workers (strings, numbers, dicts, simple distributions)
     @everywhere args = $args
+    @everywhere conditional_vars = $conditional_vars
+    @everywhere prior_overrides = $prior_overrides
+    @everywhere vars_to_scan = $vars_to_scan
+    @everywhere cache_dir = $cache_dir
 
     @everywhere begin
         using Distributions
         using DensityInterface
         using BAT
+        using DataStructures
         using MeasureBase
         using ADTypes
         using Newtrinos
+        using Accessors
 
         include(joinpath(@__DIR__, "cli_common.jl"))
 
-        ##### PHYSICS CONFIG #####
-        # To use defaults:
+        # Each worker builds everything from scratch — no serialization needed
         experiments = configure_experiments(args["experiments"])
-
-        # To override physics (e.g. for IO, sterile models, custom flux, etc.), uncomment and modify:
-        # osc = Newtrinos.osc.configure(Newtrinos.osc.OscillationConfig(
-        #     flavour=Newtrinos.osc.ThreeFlavour(ordering=:IO),
-        #     interaction=Newtrinos.osc.SI(),
-        # ))
-        # atm_flux = Newtrinos.atm_flux.configure()
-        # earth_layers = Newtrinos.earth_layers.configure()
-        # xsec = Newtrinos.xsec.configure()
-        # physics = (; osc, atm_flux, earth_layers, xsec)
-        # experiments = configure_experiments(args["experiments"], physics)
-
+        p = Newtrinos.get_params(experiments)
+        priors = Newtrinos.get_priors(experiments)
         likelihood = Newtrinos.generate_likelihood(experiments)
 
-        # Set AD backend on each worker
         ad_backend = Symbol(args["ad"])
         Newtrinos.set_ad_backend(ad_backend)
-        p = Newtrinos.get_params(experiments)
         set_batcontext(ad = Newtrinos.select_ad(length(p)))
+
+        priors = Newtrinos.condition(priors, conditional_vars, p)
+        for (k, v) in prior_overrides
+            @reset priors[k] = v
+        end
+
+        _, scanpoints = Newtrinos.generate_scanpoints(vars_to_scan, priors)
     end
 
-    # Define remote work function that uses each worker's local likelihood
-    @everywhere function _remote_find_mle(scanpoint, params, cache_dir)
-        Newtrinos.find_mle_cached(likelihood, scanpoint, deepcopy(params), cache_dir)
+    # Workers have everything — just send integer indices
[…]
