+++ src/analysis/analysis.jl
@@ -61,93 +61,44 @@ n_workers = args["workers"]
 n_threads = args["threads"]
 use_distributed = n_workers > 1
 
-"""
-    _distributed_profile(args, conditional_vars, prior_overrides, vars_to_scan, cache_dir, n_workers, n_threads)
-
-Run profiling on distributed workers. Each worker builds its own likelihood, scanpoints,
-and AD context locally — nothing is serialized except integer indices and scalar results.
-`prior_overrides` is a Dict{Symbol,Distribution} of prior replacements to apply after conditioning.
-"""
-function _distributed_profile(args, conditional_vars, prior_overrides, vars_to_scan, cache_dir, n_workers, n_threads)
-    t1 = time()
-
+# Set up distributed workers if requested
+map_func = nothing
+if use_distributed
     addprocs(n_workers; exeflags="--threads=$n_threads")
 
-    # Send only plain data to workers (strings, numbers, dicts, simple distributions)
     @everywhere args = $args
-    @everywhere conditional_vars = $conditional_vars
-    @everywhere prior_overrides = $prior_overrides
-    @everywhere vars_to_scan = $vars_to_scan
-    @everywhere cache_dir = $cache_dir
 
     @everywhere begin
         using Distributions
         using DensityInterface
         using BAT
-        using DataStructures
         using MeasureBase
         using ADTypes
         using Newtrinos
-        using Accessors
 
-        include(joinpath(@__DIR__, "cli_common.jl"))
-
-        # Each worker builds everything from scratch — no serialization needed
-        experiments = configure_experiments(args["experiments"])
-        p = Newtrinos.get_params(experiments)
-        priors = Newtrinos.get_priors(experiments)
+        experiments = Newtrinos.configure_experiments(args["experiments"])
         likelihood = Newtrinos.generate_likelihood(experiments)
 
         ad_backend = Symbol(args["ad"])
         Newtrinos.set_ad_backend(ad_backend)
         set_batcontext(ad = Newtrinos.select_ad(length(p)))
-
-        priors = Newtrinos.condition(priors, conditional_vars, p)
-        for (k, v) in prior_overrides
-            @reset priors[k] = v
-        end
-
-        _, scanpoints = Newtrinos.generate_scanpoints(vars_to_scan, priors)
     end
 
-    # Workers have everything — just send integer indices
-    @everywhere function _do_work(i)
-        Newtrinos.find_mle_cached(likelihood, scanpoints[i], deepcopy(p), cache_dir)
+    # Define work function on workers that uses their local likelihood
+    @everywhere function _do_work(scanpoint, params, cache_dir)
+        Newtrinos.find_mle_cached(likelihood, scanpoint, deepcopy(params), cache_dir)
     end
 
-    n_points = prod(values(vars_to_scan))
-    work = collect(1:n_points)
-
-    if !isnothing(cache_dir)
-        if isdir(cache_dir)
-            @info "Reusing cache dir `$(cache_dir)`"
-        else
-            mkdir(cache_dir)
-        end
+    map_func = (work, scanpoints, params, cache_dir) -> pmap(work) do i
+        _do_work(scanpoints[i], params, cache_dir)
     end
-
-    opt_results_flat = pmap(_do_work, work)
-
-    # Collect scanpoint grid shape and assemble results
-    grid_shape = Tuple(values(vars_to_scan))
[…]
