+++ src/analysis/analysis_tools.jl
@@ -64,6 +64,25 @@ end
 # AD backend is set via set_ad_backend() + set_batcontext(ad = select_ad(n))
 # Do NOT set a hardcoded default here — it overrides user selection
 
+# ── Experiment Configuration ───────────────────────────────────────
+
+"""
+    configure_experiments(experiment_list)
+    configure_experiments(experiment_list, physics)
+
+Construct a NamedTuple of configured experiments from a list of experiment name strings.
+Optionally pass a shared `physics` override.
+"""
+function configure_experiments(experiment_list)
+    pairs = (Symbol(lowercase(exp)) => getproperty(getproperty(Newtrinos, Symbol(lowercase(exp))), :configure)() for exp in experiment_list)
+    return (; pairs...)
+end
+
+function configure_experiments(experiment_list, physics)
+    pairs = (Symbol(lowercase(exp)) => getproperty(getproperty(Newtrinos, Symbol(lowercase(exp))), :configure)(physics) for exp in experiment_list)
+    return (; pairs...)
+end
+
 # ── Core Types ──────────────────────────────────────────────────────
 
 """
@@ -436,18 +455,17 @@ function assemble_profile_results(opt_results, result_size)
 end
 
 function _profile(likelihood, scanpoints, params, cache_dir; map_func=nothing)
-    do_work(i) = find_mle_cached(likelihood, scanpoints[i], deepcopy(params), cache_dir)
-
     if isnothing(map_func)
         # Default: threaded execution
         opt_results = Array{Any}(undef, size(scanpoints))
         @showprogress Threads.@threads for i in eachindex(scanpoints)
-            opt_results[i] = do_work(i)
+            opt_results[i] = find_mle_cached(likelihood, scanpoints[i], deepcopy(params), cache_dir)
         end
     else
-        # Custom map (e.g., pmap for distributed)
+        # Distributed: pass scanpoints, params, cache_dir as data
+        # map_func is responsible for calling find_mle with its own likelihood
         work = collect(eachindex(scanpoints))
-        opt_results_flat = map_func(do_work, work)
+        opt_results_flat = map_func(work, scanpoints, params, cache_dir)
         opt_results = reshape(opt_results_flat, size(scanpoints))
     end
     assemble_profile_results(opt_results, size(scanpoints))
