### 2026-03-20 55fb611
Add Super-K detector systematics and adjust priors based on Wester thesis
Add 6 new nuisance parameters: relative normalizations for multi-GeV FC
and PC+upmu samples (flux model differences at high energy), FC/PC
separation migration, pi0 selection uncertainty, and split sub-GeV vs
multi-GeV ring counting.

Adjust energy scale priors to weighted averages from Table 5.6, and
tighten sub-GeV PID priors to 2% (better constrained than multi-GeV).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
