+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -243,6 +243,20 @@ function get_assets(physics; datadir = @__DIR__)
         # SK phase masks (for split energy scale)
         sk_i_iii_bins = occursin.(r"^sk1-3_", bininfo.Sample),
         sk_iv_v_bins = .!occursin.(r"^sk1-3_", bininfo.Sample),
+        # Multi-GeV FC mask (for relative normalization)
+        fc_multigev = occursin.(r"_fc_multigev_", bininfo.Sample),
+        # PC + Up-mu mask (for relative normalization)
+        pc_upmu = occursin.("_pc_", bininfo.Sample) .| occursin.("_upmu_", bininfo.Sample),
+        # FC multi-GeV mu-like single-ring (for FC/PC separation)
+        fc_multigev_mulike = occursin.(r"_fc_multigev_1ring_mu", bininfo.Sample),
+        # pi0 samples
+        sk_1ring_pi0 = occursin.("_1ring_ncpi0", bininfo.Sample),
+        sk_2ring_pi0 = occursin.("_2ring_ncpi0", bininfo.Sample),
+        # Ring separation sub-GeV vs multi-GeV
+        sk_subgev_1ring = occursin.(r"_fc_subgev_1ring_", bininfo.Sample),
+        sk_subgev_multiring = occursin.(r"_fc_subgev.*(2ring|multiring)", bininfo.Sample),
+        sk_multigev_1ring = occursin.(r"_fc_multigev_1ring_", bininfo.Sample),
+        sk_multigev_multiring = occursin.(r"_fc_multigev.*(2ring|multiring)", bininfo.Sample),
     )
 
     data = CSV.read(joinpath(datadir, "bins/sk_2023_Data.txt"), DataFrame; delim=' ', ignorerepeated=true, comment="#", header=false)
@@ -324,14 +338,28 @@ function get_params()
         sk_ring_counting = 1.0,
         sk_nue_contamination = 1.0,
         sk_ncpi0_norm = 1.0,
+        # Relative normalizations (flux model differences at high energy)
+        sk_fc_multigev_rel_norm = 1.0,
+        sk_pc_upmu_rel_norm = 1.0,
+        # FC/PC separation
+        sk_fc_pc_separation = 1.0,
+        # pi0 selection
+        sk_pi0_norm = 1.0,
+        # Split ring counting: sub-GeV and multi-GeV
+        sk_subgev_ring_counting = 1.0,
+        sk_multigev_ring_counting = 1.0,
         )
 end
 
 function get_priors()
     priors = (
-        sk_i_iii_energy_scale = Normal(1.0, 0.03),
-        sk_iv_v_energy_scale = Normal(1.0, 0.018),
-        sk_updown_energy_scale = Normal(1.0, 0.01),
+        # Energy scales: weighted averages from Table 5.6 (conventional FV)
+        # SK I: 3.3%, SK II: 2.0%, SK III: 2.4% → weighted ~2.5%
+        # SK IV: 2.1%, SK V: 1.8% → weighted ~2.0%
+        sk_i_iii_energy_scale = Normal(1.0, 0.025),
+        sk_iv_v_energy_scale = Normal(1.0, 0.02),
+        # Up/down: SK I-III ~0.6-1.3%, SK IV-V ~0.5-0.7% → use 0.8% average
+        sk_updown_energy_scale = Normal(1.0, 0.008),
         sk_fc_norm = Normal(1.0, 0.05),
         sk_pc_norm = Normal(1.0, 0.05),
         sk_upmu_norm = Normal(1.0, 0.05),
@@ -347,13 +375,26 @@ function get_priors()
[…]
