### 2026-03-20 de42e3b
Add configurable earth density model and move layer computation to forward model
Add VariableDensity config type to earth_layers with matter_density_scale
parameter (Normal(1.0, 0.068) prior, 6.8% from Wester thesis). The default
PREM config remains parameter-free for backward compatibility.

Move layers/paths computation from get_assets() to the forward model
(reweight/calc_weights) in all 5 atmospheric experiments (super_k,
deepcore 9y, deepcore 3y, orca, ic_upgrade). This enables parameterized
earth density in the likelihood. Path recomputation costs ~11μs, negligible
compared to oscillatio
[…]
