+++ src/physics/earth_layers.jl
@@ -4,12 +4,13 @@ using CSV, DataFrames
 using StatsBase
 using StaticArrays, ArraysOfArrays, StructArrays
 using DataStructures
+using Distributions
 
 using ..Newtrinos
 export configure
 export PREM
 
-const datadir = @__DIR__ 
+const datadir = @__DIR__
 
 abstract type DensityModel end
 
@@ -17,6 +18,10 @@ abstract type DensityModel end
     zones::Array{Float64} = [0., 4., 7.5, 12.5, 13.1]
     p_fractions::Float64 = 0.5
     atm_heihgt::Float64 = 20.
+end
+
+@kwdef struct VariableDensity <: DensityModel
+    prem::PREM = PREM()
 end
 
 @kwdef struct EarthLayers <: Newtrinos.Physics
@@ -27,7 +32,7 @@ end
     compute_paths::Function
 end
 
-function configure(cfg::DensityModel=PREM())
+function configure(cfg::PREM=PREM())
     EarthLayers(
         cfg=cfg,
         params = (;),
@@ -37,28 +42,43 @@ function configure(cfg::DensityModel=PREM())
         )
 end
 
+function configure(cfg::VariableDensity)
+    EarthLayers(
+        cfg=cfg,
+        params = (matter_density_scale = 1.0,),
+        priors = (matter_density_scale = Normal(1.0, 0.068),),
+        compute_layers = get_compute_layers(cfg.prem),
+        compute_paths = compute_paths
+        )
+end
+
 
 function get_compute_layers(cfg::PREM)
     function compute_layers()
-        
+
         PREM = CSV.read(joinpath(datadir, "PREM_1s.csv"), DataFrame, header=["radius","depth","density","Vpv","Vph","Vsv","Vsh","eta","Q-mu","Q-kappa"])
         # density boundaries to define the constant density zones
-        
+
         radii = Float64[]
         ave_densities = Float64[]
-        
+
         push!(radii, 6371+cfg.atm_heihgt)
         push!(ave_densities, 0.)
-        
+
         for i in 1:length(cfg.zones)-1
             mask = (PREM.density .< cfg.zones[i+1]) .& (PREM.density .>= cfg.zones[i])
             push!(radii, maximum(PREM.radius[mask]))
             push!(ave_densities, mean(PREM.density[mask]))
         end
-        
+
         layers = StructArray{Newtrinos.Layer}((radii, ave_densities .* cfg.p_fractions, ave_densities .* (1 .- cfg.p_fractions)))
     end
 end
+
+function scale_densities(layers, scale)
+    StructArray{Newtrinos.Layer}((layers.radius, layers.p_density .* scale, layers.n_density .* scale))
+end
+
 function ray_circle_path_length(r, y, cz)
     # Compute the discriminant
     disc = r^2 - y^2 + (y * cz)^2
@@ -112,7 +132,7 @@ function compute_paths(cz::Number, layers, r_detector)
             lengths_traversed[i] = inter/2 + len_det
             layer_idx_traversed[i] = i
             lengths_traversed[end-i+n_layers_outside] = inter/2
-            layer_idx_traversed[end-i+n_layers_outside] = i              
+            layer_idx_traversed[end-i+n_layers_outside] = i
         else
             lengths_traversed[i] = intersections[i]/2
             layer_idx_traversed[i] = i
@@ -122,7 +142,7 @@ function compute_paths(cz::Number, layers, r_detector)
     end
 
     la = StructArray{Newtrinos.Path}((lengths_traversed, layer_idx_traversed))
-    
+
 end
 
 function compute_paths(cz::AbstractArray, layers; r_detector = 6369)
