### 2026-03-20 0affddc
Update documentation for earth_layers refactoring and Super-K systematics
- CLAUDE.md: Document VariableDensity config, earth layers/paths in
  forward model pattern, update test count, note compute_paths exception
  to precompute rule
- README.md: Add VariableDensity example to physics configuration section

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
