### 2026-03-20 de82eb0
Fix Layer struct to support mixed types for ForwardDiff compatibility
Split Layer{T} into Layer{T,U} so radius (always Float64) and densities
(Dual when differentiating through matter_density_scale) can have
different types.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
