### 2026-03-20 e7506cd
Fix typo: rename btd -> bdt (Boosted Decision Tree) in Super-K params
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
