### 2026-03-20 7999db6
Add flux normalization, CC xsec norm, and tighten SK detector priors
Add atm_flux_norm (10%) for overall atmospheric flux normalization and
cc_norm (15%) for CC cross-section normalization in Differential_H2O.
Tighten SK sample norms to reduction-only uncertainties (FC 1.5%, PC 3%,
up-mu 1%) now that flux norm freedom is handled by atm_flux_norm.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
