### 2026-03-20 a610d1b
setting non-constant density for SK as default; free CP phase
