+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -26,7 +26,7 @@ end
 function default_physics()
     osc = Newtrinos.osc.configure(Newtrinos.osc.OscillationConfig(interaction=Newtrinos.osc.SI()))
     atm_flux = Newtrinos.atm_flux.configure(Newtrinos.atm_flux.AtmFluxConfig(nominal_model=Newtrinos.atm_flux.HKKM("kam-ally-20-01-mtn-solmin.d")))
-    earth_layers = Newtrinos.earth_layers.configure()
+    earth_layers = Newtrinos.earth_layers.configure(Newtrinos.earth_layers.VariableDensity())
     xsec = Newtrinos.xsec.configure(Newtrinos.xsec.Differential_H2O())
     (; osc, atm_flux, earth_layers, xsec)
 end
