### 2026-03-20 4ce9984
setting CP default to pi/2
