### 2026-03-21 634e06d
Move flux norm from atm_flux to SK with energy-dependent bathtub shape
Remove global atm_flux_norm parameter. Instead add experiment-specific
energy-dependent flux normalization to Super-K with two parameters:
- sk_flux_norm_low: 25% at 0.1 GeV, linear in logE to 7% at 1 GeV
- sk_flux_norm_high: 7% flat from 1-10 GeV, rising to 20% at 1 TeV
Applied in true energy space before response matrix contraction.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
