### 2026-03-22 78e99bd
Fix Fij up/down energy scale to split downgoing/upgoing computation
The old approach fused shifted R matrices by resetting downgoing bins to
nominal, but used nominal energy for flux/osc everywhere. Now compute
calc_weights separately for downgoing (nominal R, nominal E) and upgoing
(shifted R, shifted E), then add the results.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
