### 2026-03-22 bf09d29
Implement layer-dependent Ye (electron fraction) in Earth matter effects
The matter potential V_CC ∝ Ye × ρ, where Ye = Z/A varies between mantle
(~0.495) and core (~0.467). Previously Ye=0.5 was used uniformly, overestimating
the core matter potential by ~7%. Now p_fractions is per-zone in the PREM struct.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
