+++ src/analysis/ye_study.jl
@@ -1,0 +1,235 @@
+#!/usr/bin/env julia
+"""
+Estimate the impact of proper layer-dependent Ye (electron fraction) on Δm² fit in Super-K.
+
+The matter effect depends on electron density Ne = Ye × ρ × NA. Previously Ye=0.5
+was used for all layers. In reality:
+  - Mantle (silicate rock): Ye ≈ 0.494–0.496
+  - Core (iron-nickel):     Ye ≈ 0.466–0.468
+
+The core Ye is ~7% lower than 0.5, reducing the matter potential for core-crossing
+(upgoing) neutrinos. This script quantifies the effect on Super-K bin predictions
+and the resulting shift in best-fit Δm².
+"""
+
+using Newtrinos
+using StatsBase: midpoints
+using Accessors
+using CairoMakie
+using Printf
+
+# --- Helper: compute bin predictions with a given PREM config ---
+
+function compute_prediction(params, assets, physics, layers)
+    E = 10.0 .^ midpoints(assets.loge_grid)
+    logE = midpoints(assets.loge_grid)
+
+    paths = physics.earth_layers.compute_paths(assets.cz_midpoints, layers)
+
+    p = physics.osc.osc_prob(E, paths, layers, params)
+    p_anti = physics.osc.osc_prob(E, paths, layers, params, anti=true)
+
+    flux = physics.atm_flux.sys_flux(assets.flux_nominal, params)
+    s = (size(p, 1), size(p, 2))
+
+    fnl = haskey(params, :sk_flux_norm_low) ? params.sk_flux_norm_low : zero(eltype(E))
+    fnh = haskey(params, :sk_flux_norm_high) ? params.sk_flux_norm_high : zero(eltype(E))
+    flux_norm = 1 .+ fnl .* Newtrinos.super_k.flux_norm_sigma_low.(logE) .+
+                      fnh .* Newtrinos.super_k.flux_norm_sigma_high.(logE)
+
+    xsec_nue     = physics.xsec.scale(E, :nue,   :CC, false, params)
+    xsec_numu    = physics.xsec.scale(E, :numu,  :CC, false, params)
+    xsec_nutau   = physics.xsec.scale(E, :nutau, :CC, false, params)
+    xsec_nuebar  = physics.xsec.scale(E, :nue,   :CC, true,  params)
+    xsec_numubar = physics.xsec.scale(E, :numu,  :CC, true,  params)
+    xsec_nutaubar= physics.xsec.scale(E, :nutau, :CC, true,  params)
+    xsec_nc      = physics.xsec.scale(E, :nue,   :NC, false, params)
+
+    nue_flux   = (reshape(flux.nue,    s) .* p[:, :, 1, 1] .+
+                  reshape(flux.numu,   s) .* p[:, :, 2, 1]) .* xsec_nue .* flux_norm
+    numu_flux  = (reshape(flux.nue,    s) .* p[:, :, 1, 2] .+
+                  reshape(flux.numu,   s) .* p[:, :, 2, 2]) .* xsec_numu .* flux_norm
+    nutau_flux = (reshape(flux.nue,    s) .* p[:, :, 1, 3] .+
+                  reshape(flux.numu,   s) .* p[:, :, 2, 3]) .* xsec_nutau .* flux_norm
+    nuebar_flux  = (reshape(flux.nuebar,  s) .* p_anti[:, :, 1, 1] .+
+                    reshape(flux.numubar, s) .* p_anti[:, :, 2, 1]) .* xsec_nuebar .* flux_norm
[…]
