### 2026-03-22 fb06731
Rename matter_density_scale to electron_density_scale and activate for Super-K
scale_densities now only scales p_density (electron density), leaving
n_density unchanged — making it a true Ye uncertainty as described in the
Wester thesis (Section 5.2.4, 6.8% prior). Activated VariableDensity for
Super-K's default physics to match the real SK analysis.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
