### 2026-03-22 2f06635
Fix type mismatch in scale_densities for ForwardDiff compatibility
Promote n_density to match Dual type of scaled p_density via one.(p).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
