### 2026-03-22 414dc4c
Conserve total density when scaling electron density
When electron_density_scale changes p_density, transfer the difference
to n_density so that ρ_total = p_density + n_density remains constant.
This correctly models a Ye (electron fraction) uncertainty at fixed mass density.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
