### 2026-03-23 84dd5a5
Add 5% resolution broadening to Super-K response matrix construction
Broaden quantile spreads by 5% around the median before building spline
CDFs, to account for uncertainty in the reconstruction distribution
(known only from 5 quantile points). Also add response_matrix_study.jl
comparing spline vs split-Gaussian CDF construction.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
