+++ src/analysis/response_matrix_study.jl
@@ -1,0 +1,483 @@
+#!/usr/bin/env julia
+"""
+Compare spline vs split-Gaussian response matrix construction for Super-K.
+
+The current approach fits RQSpline CDFs through 5 quantile points + 2 linearly
+extrapolated tail points. This study compares against a split-Gaussian CDF
+constructed from the same quantiles, to check whether the spline tails
+underestimate smearing.
+"""
+
+using Newtrinos
+using StatsBase: midpoints
+using Accessors
+using CairoMakie
+using Printf
+using Distributions: Normal, cdf as dist_cdf
+
+# ── Split-Gaussian CDF from quantiles ──────────────────────────────────────
+
+"""
+    make_split_gaussian_log_e_cdf(bin)
+
+Build a split-Gaussian CDF in log10(E) from the 5 quantile points.
+μ = Q50 in log10, σ_low from lower quantiles, σ_high from upper quantiles.
+"""
+function make_split_gaussian_log_e_cdf(bin)
+    q = log10.([bin.EnergyQuantile2_3Percent, bin.EnergyQuantile15_9Percent,
+                bin.EnergyQuantile50_0Percent, bin.EnergyQuantile84_1Percent,
+                bin.EnergyQuantile97_7Percent])
+
+    μ = q[3]
+    # Average of 1σ and 2σ estimates for each side
+    σ_low  = 0.5 * ((q[3] - q[2]) + (q[3] - q[1]) / 2)
+    σ_high = 0.5 * ((q[4] - q[3]) + (q[5] - q[3]) / 2)
+
+    # Guard against degenerate bins
+    σ_low  = max(σ_low, 1e-6)
+    σ_high = max(σ_high, 1e-6)
+
+    return x -> begin
+        if x <= μ
+            dist_cdf(Normal(μ, σ_low), x)
+        else
+            dist_cdf(Normal(μ, σ_high), x)
+        end
+    end
+end
+
+"""
+    make_split_gaussian_cosz_cdf(bin)
+
+Build a split-Gaussian CDF in cos(zenith) from the 5 quantile points.
+"""
+function make_split_gaussian_cosz_cdf(bin)
+    q = [bin.CosZQuantile2_3Percent, bin.CosZQuantile15_9Percent,
+         bin.CosZQuantile50_0Percent, bin.CosZQuantile84_1Percent,
+         bin.CosZQuantile97_7Percent]
+
+    μ = q[3]
+    σ_low  = 0.5 * ((q[3] - q[2]) + (q[3] - q[1]) / 2)
+    σ_high = 0.5 * ((q[4] - q[3]) + (q[5] - q[3]) / 2)
+    σ_low  = max(σ_low, 1e-6)
+    σ_high = max(σ_high, 1e-6)
+
+    return x -> begin
+        if x <= μ
+            dist_cdf(Normal(μ, σ_low), x)
+        else
+            dist_cdf(Normal(μ, σ_high), x)
+        end
+    end
+end
+
+# ── Response matrix with split-Gaussian CDFs ───────────────────────────────
+
+function make_response_matrix_splitgauss(MC_component, logE_grid, cosZ_grid)
+    n_bins = size(MC_component, 1)
+    n_logE = length(logE_grid)
+    n_cosZ = length(cosZ_grid)
+
+    response_matrix = zeros(Float64, n_bins, n_logE-1, n_cosZ-1)
+
+    for bin_idx in 1:n_bins
+        bin = MC_component[bin_idx, :]
+        bin.Counts == 0 && continue
+
+        log_e_cdf = make_split_gaussian_log_e_cdf(bin)
+        cosz_cdf = make_split_gaussian_cosz_cdf(bin)
+
[…]
