### 2026-03-23 e51521d
Add --profile-vars CLI option and NO vs IO quantile comparison
analysis.jl: Add --profile-vars flag to specify which variables to
profile/scan over from the command line (e.g. --profile-vars theta23:21
dm231). Supports CLI-friendly aliases and direct parameter names.

response_matrix_study.jl: Add Part 5 comparing normal vs inverted
ordering MC quantiles to verify reconstruction kernels are
ordering-independent.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
