+++ src/analysis/response_matrix_study.jl
@@ -14,6 +14,7 @@ using Accessors
 using CairoMakie
 using Printf
 using Distributions: Normal, cdf as dist_cdf
+using CSV, DataFrames
 
 # ── Split-Gaussian CDF from quantiles ──────────────────────────────────────
 
@@ -474,6 +475,191 @@ function main()
     outpath = joinpath(@__DIR__, "response_matrix_study.png")
     save(outpath, fig, px_per_unit=2)
     println("\nSaved plot to $outpath")
+
+    # ── Part 5: NO vs IO quantile comparison ───────────────────────────
+
+    println("\n=== Part 5: Normal vs Inverted ordering quantile comparison ===")
+
+    datadir = joinpath(dirname(pathof(Newtrinos)), "experiments", "super_k", "sk_atm_2023")
+    read_sk = Newtrinos.super_k.read_sk_file
+
+    MC_IO = (
+        nue     = read_sk(joinpath(datadir, "bins/inverted/sk_2023_MCNueIO.txt")),
+        numu    = read_sk(joinpath(datadir, "bins/inverted/sk_2023_MCNumuIO.txt")),
+        nutau   = read_sk(joinpath(datadir, "bins/inverted/sk_2023_MCNutauIO.txt")),
+        nuebar  = read_sk(joinpath(datadir, "bins/inverted/sk_2023_MCNueBarIO.txt")),
+        numubar = read_sk(joinpath(datadir, "bins/inverted/sk_2023_MCNumuBarIO.txt")),
+        nunc    = read_sk(joinpath(datadir, "bins/inverted/sk_2023_MCNCIO.txt")),
+    )
+
+    # Compare quantiles bin-by-bin between NO and IO
+    e_quantile_cols = [:EnergyQuantile2_3Percent, :EnergyQuantile15_9Percent,
+                       :EnergyQuantile50_0Percent, :EnergyQuantile84_1Percent,
+                       :EnergyQuantile97_7Percent]
+    cz_quantile_cols = [:CosZQuantile2_3Percent, :CosZQuantile15_9Percent,
+                        :CosZQuantile50_0Percent, :CosZQuantile84_1Percent,
+                        :CosZQuantile97_7Percent]
+
+    # Collect fractional differences in energy quantiles and absolute diffs in cosZ
+    e_frac_diffs = Float64[]   # (IO - NO) / NO for energy quantiles
+    cz_abs_diffs = Float64[]   # IO - NO for cosZ quantiles
+    e_rms_frac_diffs = Float64[]  # (IO_RMS - NO_RMS) / NO_RMS
+    cz_rms_abs_diffs = Float64[]
+    counts_frac_diffs = Float64[]  # (IO_counts - NO_counts) / NO_counts
+
+    for key in keys(MC)
+        mc_no = MC[key]
+        mc_io = MC_IO[key]
+        for bin_idx in 1:size(mc_no, 1)
+            bin_no = mc_no[bin_idx, :]
+            bin_io = mc_io[bin_idx, :]
+            (bin_no.Counts == 0 || bin_io.Counts == 0) && continue
+
+            # Counts difference
+            push!(counts_frac_diffs, (bin_io.Counts - bin_no.Counts) / bin_no.Counts)
+
+            # Energy quantiles (fractional diff in linear E)
+            for col in e_quantile_cols
+                v_no = bin_no[col]
+                v_io = bin_io[col]
+                v_no > 0 && push!(e_frac_diffs, (v_io - v_no) / v_no)
+            end
+
+            # CosZ quantiles (absolute diff)
[…]
