### 2026-03-23 1f5ae97
Replace F_ij energy scale with reco bin overlap method; revert δCP default
Energy scale uncertainty now shifts E_reco bin edges and redistributes
counts via overlap fractions (exact, count-conserving) instead of the
old F_ij linear approximation in E_true space. Old F_ij code commented
out for reference. Also reverts δCP default from π/2 back to 1.0.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
