+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -153,6 +153,62 @@ function contract_R(R_flat, weighted_flux)
     R_flat * vec(weighted_flux)
 end
 
+# Energy scale via reco bin overlap method
+struct EnergyGroup
+    indices::Vector{Int}
+    logP_edges::Vector{Float64}
+end
+
+function build_energy_groups(bininfo, bin_mask)
+    groups = EnergyGroup[]
+    masked_indices = findall(bin_mask)
+    sub_bininfo = bininfo[masked_indices, :]
+    for key in unique(zip(sub_bininfo.Sample, sub_bininfo.CosZMin, sub_bininfo.CosZMax))
+        local_mask = (sub_bininfo.Sample .== key[1]) .& (sub_bininfo.CosZMin .== key[2]) .& (sub_bininfo.CosZMax .== key[3])
+        local_idxs = findall(local_mask)
+        order = sortperm(sub_bininfo.logPMin[local_idxs])
+        sorted_local = local_idxs[order]
+        global_idxs = masked_indices[sorted_local]
+        edges = vcat(sub_bininfo.logPMin[sorted_local], [sub_bininfo.logPMax[sorted_local[end]]])
+        push!(groups, EnergyGroup(global_idxs, edges))
+    end
+    return groups
+end
+
+function apply_energy_scale(counts, energy_groups, delta)
+    result = copy(counts)
+    for group in energy_groups
+        n = length(group.indices)
+        edges = group.logP_edges
+        for i in 1:n
+            # Clamp edge bins to group boundaries for count conservation
+            shifted_lo = (i == 1) ? edges[1] : edges[i] + delta
+            shifted_hi = (i == n) ? edges[end] : edges[i+1] + delta
+            acc = zero(eltype(counts))
+            for j in 1:n
+                overlap = max(zero(eltype(counts)), min(shifted_hi, edges[j+1]) - max(shifted_lo, edges[j]))
+                nom_width = edges[j+1] - edges[j]
+                acc += (overlap / nom_width) * counts[group.indices[j]]
+            end
+            result[group.indices[i]] = acc
+        end
+    end
+    return result
+end
+
+function apply_all_energy_scales(counts, assets, params)
+    delta_i_iii = log10(params.sk_i_iii_energy_scale)
+    delta_iv_v = log10(params.sk_iv_v_energy_scale)
+    result = apply_energy_scale(counts, assets.energy_groups_sk_i_iii, delta_i_iii)
+    result = apply_energy_scale(result, assets.energy_groups_sk_iv_v, delta_iv_v)
+
+    delta_ud_i_iii = log10(params.sk_i_iii_updown_energy_scale)
+    delta_ud_iv_v = log10(params.sk_iv_v_updown_energy_scale)
+    result = apply_energy_scale(result, assets.energy_groups_sk_i_iii_up, delta_ud_i_iii)
+    result = apply_energy_scale(result, assets.energy_groups_sk_iv_v_up, delta_ud_iv_v)
+    return result
+end
+
 function flux_norm_sigma_low(logE)
     # 25% at logE=-1 (0.1 GeV), linear in logE to 7% at logE=0 (1 GeV), zero above
     logE < zero(logE) ? max(0.07, 0.25 - 0.18 * (logE + 1)) : zero(logE)
@@ -320,38 +376,50 @@ function get_assets(physics; datadir = @__DIR__)
     R = flatten_R(R_3d)
[…]
