### 2026-03-23 8aefef3
Use truncated energy scale priors and finer E_true grid (251 points)
Truncate energy scale priors to [0.5, 1.5] to prevent non-positive
values reaching log10. Increase loge_grid from 201 to 251 points
to investigate energy resolution impact on Δm² constraints.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
