### 2026-03-23 e12fad8
Fix antineutrino matter oscillation double-conjugation bug
propagate() was re-conjugating U for antineutrinos, but osc_prob() already
passes conj(U_PMNS). This double conjugation produced the neutrino Hamiltonian
instead of the antineutrino one — an error proportional to sin(δCP)·sin(θ₁₃).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
