### 2026-03-23 e323e8e
Fix typo: rename atm_flux_uphorizonzal_sigma to atm_flux_uphorizontal_sigma
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
