### 2026-03-23 5633b0b
Fix transposed oscillation probability indices in osc_prob
osc_prob returned P[E, L, out, in] but all experiments indexed it as
P[E, L, in, out]. This swapped P(νμ→νe) with P(νe→νμ) — an error equal
to the T-violating asymmetry, proportional to J × sin(δCP). Fixed by
transposing the flavour indices in _add_rest_and_permute so osc_prob now
returns P[E, L, α, β] = P(να → νβ) matching how experiments use it.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
