+++ test/test_regression.jl
@@ -9,10 +9,10 @@ using Printf
 const ALL_REFERENCE_VALUES = (
     dayabay      = -168.90003273508322,
     kamland      = -63.111403860037875,
-    minos        = -268.3363280363475,
-    deepcore     = -950.8506524476818,
-    super_k      = -3693.676453878093,
-    orca         = -1164.226373825956,
+    minos        = -268.40461484198985,
+    deepcore     = -950.6200930498498,
+    super_k      = -3718.157000175498,
+    orca         = -1164.3663831498247,
     coherent_csi = -574.3416032522271,
     coherent_lAr = -1754.9946939034153,
 )
