### 2026-03-24 1221f7a
Use additive systematic factors in Super-K to match Wester 2023 F_ij formalism
The multiplicative chaining of ~20 systematic scale factors compounded
deviations as (1+δ)^N, creating artificial over/under-prediction in bins
where many systematics overlap. Switch to the linearized additive form
1 + Σ_j (f_ij - 1) to match the thesis formalism.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
