+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -353,7 +353,7 @@ function get_assets(physics; datadir = @__DIR__)
         nunc=read_sk_file(joinpath(datadir, "bins/normal/sk_2023_MCNCNO.txt")))
 
         
-    loge_grid = LinRange(-1,3,251)
+    loge_grid = LinRange(-1,3,201)
     cz_grid = LinRange(-1.0,1.0,101)
 
     # Bestfit from SK atm 2023 paper
@@ -548,38 +548,40 @@ function get_double_factor(total, mask1, mask2, factor1)
 end
 
 function get_all_factors(params, assets, total)
-    return (
-        get_factor(assets.masks.fc, params.sk_fc_norm * params.sk_fiducial_norm) .*
-        get_factor(assets.masks.pc, params.sk_pc_norm * params.sk_fiducial_norm) .*
-        get_factor(assets.masks.upmu, params.sk_upmu_norm) .*
-        get_double_factor(total, assets.masks.pc_stop, assets.masks.pc_thru, params.sk_pc_stopping_vs_througoing) .*
-        get_double_factor(total, assets.masks.umpmu_stop, assets.masks.upmu_thru, params.sk_upmu_stopping_vs_througoing) .*
-        get_double_factor(total, assets.masks.upmu_nonshower, assets.masks.upmu_shower, params.sk_upmu_nonshower_vs_shower) .*
-        get_double_factor(total, assets.masks.sk_i_iii_elike_1decay_e, assets.masks.sk_i_iii_elike_0decay_e, params.sk_i_iii_decay_e_tag_eff) .*
-        get_double_factor(total, assets.masks.sk_i_iii_mulike_1decay_e, assets.masks.sk_i_iii_mulike_0decay_e, params.sk_i_iii_decay_e_tag_eff) .*
-        get_double_factor(total, assets.masks.sk_i_iii_mulike_2decay_e, assets.masks.sk_i_iii_mulike_1decay_e, params.sk_i_iii_decay_e_tag_eff) .*
-        get_double_factor(total, assets.masks.sk_iv_v_1decay_e, assets.masks.sk_iv_v_0decay_e, params.sk_iv_v_decay_e_tag_eff) .*
-        get_double_factor(total, assets.masks.sk_iv_v_subgev_0neutron, assets.masks.sk_iv_v_subgev_1neutron, params.sk_iv_v_subgev_neutron_tag_eff) .*
-        get_double_factor(total, assets.masks.sk_iv_v_multigev_0neutron, assets.masks.sk_iv_v_multigev_1neutron, params.sk_iv_v_multigev_neutron_tag_eff) .*
-        get_double_factor(total, assets.masks.sk_i_v_multigev_multiring_nuebar, assets.masks.sk_i_v_multigev_multiring_nue, params.sk_i_v_bdt_1) .*
-        get_double_factor(total, assets.masks.sk_i_v_multigev_multiring_nue, assets.masks.sk_i_v_multigev_multiring_mu, params.sk_i_v_bdt_2) .*
-        get_double_factor(total, assets.masks.sk_i_v_multigev_multiring_mu, assets.masks.sk_i_v_multigev_multiring_other, params.sk_i_v_bdt_3) .*
+    # Additive/linearized formalism: expected_i = nominal_i × (1 + Σ_j f_ij × ε_j)
+    # Each factor returns 1 for unaffected bins, so (factor .- 1) gives 0 there.
+    return 1 .+ (
+        (get_factor(assets.masks.fc, params.sk_fc_norm * params.sk_fiducial_norm) .- 1) .+
+        (get_factor(assets.masks.pc, params.sk_pc_norm * params.sk_fiducial_norm) .- 1) .+
[…]
