+++ test/test_regression.jl
@@ -11,7 +11,7 @@ const ALL_REFERENCE_VALUES = (
     kamland      = -63.111403860037875,
     minos        = -268.40461484198985,
     deepcore     = -950.6200930498498,
-    super_k      = -3718.157000175498,
+    super_k      = -3719.011060553307,
     orca         = -1164.3663831498247,
     coherent_csi = -574.3416032522271,
     coherent_lAr = -1754.9946939034153,
