### 2026-03-24 3050dc2
Add GENIE-based cross-section model with flavor-dependent PCA uncertainties
Replace Differential_H2O with GENIE_H2O for Super-K. Key improvements:

1. Flavor-specific CC channel fractions from GENIE G18_10a splines on O16
   (nue, nuebar, numu, numubar) instead of using nue fractions for all flavors.
   This correctly handles the muon mass threshold effect on numu CCQE.

2. PCA-based systematic uncertainties derived from the spread across 4 GENIE
   tunes (G18_10a, G21_11a/SuSAv2, N18_10j/z-expansion, G18_02a/Llewellyn-Smith).
   Two PCA components capture 100% of inter-tune variance, replacin
[…]
