### 2026-03-24 77a07c2
adding blur to osc_prob (5%)
