### 2026-03-24 01aed21
Update Super-K regression reference for oscillation smoothing
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
