### 2026-03-25 74088f7
Replace GENIE_H2O with H2O_PCA per-process cross-section reweighting model
Redesign cross-section systematics as a reweighting model with per-interaction-
process (CC1p1h, CC2p2h, CC1pi, CCDIS, CCother, NC) normalization and shape
parameters, 100% correlated across neutrino flavors.

Key changes:
- Rename GENIE_H2O → H2O_PCA (backward compat alias kept)
- Configurable nominal MC via H2O_PCA(nominal=:Wester)
- Wester (NEUT) as default nominal; numu/numubar derived via GENIE ratio
- Per-process norm (6) + shape PCA (6) + nutau_cc_norm = 13 parameters
- Shape PCA computed from 4 input curves (Wes
[…]
