+++ src/physics/xsec.jl
@@ -14,7 +14,11 @@ struct SimpleScaling <: XsecModel end
 
 struct Differential_H2O <: XsecModel end
 
-struct GENIE_H2O <: XsecModel end
+@kwdef struct H2O_PCA <: XsecModel
+    nominal::Symbol = :Wester   # which curve is the MC nominal (:Wester, :G18_10a, :G21_11a, :G18_02a)
+end
+
+const GENIE_H2O = H2O_PCA  # backward compat alias
 
 @kwdef struct Xsec <: Newtrinos.Physics
     cfg::XsecModel
@@ -179,61 +183,172 @@ function get_scale(cfg::Differential_H2O)
     end
 end
 
-function get_params(cfg::GENIE_H2O)
+function get_params(cfg::H2O_PCA)
     (
-        nc_norm = 1.,
-        nutau_cc_norm = 1.,
-        cc_norm = 1.,
-        xsec_pc1 = 0.,
-        xsec_pc2 = 0.,
+        xsec_cc1p1h_norm = 1.,
+        xsec_cc2p2h_norm = 1.,
+        xsec_cc1pi_norm = 1.,
+        xsec_ccdis_norm = 1.,
+        xsec_ccother_norm = 1.,
+        xsec_nc_norm = 1.,
+        xsec_nutau_cc_norm = 1.,
+        xsec_cc1p1h_shape = 0.,
+        xsec_cc2p2h_shape = 0.,
+        xsec_cc1pi_shape = 0.,
+        xsec_ccdis_shape = 0.,
+        xsec_ccother_shape = 0.,
+        xsec_nc_shape = 0.,
     )
 end
 
-function get_priors(cfg::GENIE_H2O)
+function get_priors(cfg::H2O_PCA)
     (
-        nc_norm = Truncated(Normal(1, 0.2), 0.4, 1.6),
-        nutau_cc_norm = Truncated(Normal(1, 0.25), 0.3, 1.7),
-        cc_norm = Truncated(Normal(1, 0.15), 0.5, 1.5),
-        xsec_pc1 = Normal(0, 1),
-        xsec_pc2 = Normal(0, 1),
+        xsec_cc1p1h_norm = Truncated(Normal(1, 0.15), 0.4, 1.6),
+        xsec_cc2p2h_norm = Truncated(Normal(1, 0.50), 0.0, 3.0),
+        xsec_cc1pi_norm = Truncated(Normal(1, 0.15), 0.4, 1.6),
+        xsec_ccdis_norm = Truncated(Normal(1, 0.10), 0.5, 1.5),
+        xsec_ccother_norm = Truncated(Normal(1, 0.30), 0.2, 1.8),
+        xsec_nc_norm = Truncated(Normal(1, 0.15), 0.4, 1.6),
+        xsec_nutau_cc_norm = Truncated(Normal(1, 0.25), 0.3, 1.7),
+        xsec_cc1p1h_shape = Normal(0, 1),
+        xsec_cc2p2h_shape = Normal(0, 1),
+        xsec_cc1pi_shape = Normal(0, 1),
+        xsec_ccdis_shape = Normal(0, 1),
+        xsec_ccother_shape = Normal(0, 1),
+        xsec_nc_shape = Normal(0, 1),
     )
 end
 
-function get_scale(cfg::GENIE_H2O)
-    # Load precomputed GENIE data: nominal channel fractions + PCA components
+function get_scale(cfg::H2O_PCA)
     data = load(joinpath(@__DIR__, "xsec_genie_data.jld2"))
     E_grid = data["E_grid"]
-    nominal_fractions = data["nominal_fractions"]
-    pca_components = data["pca_components"]
[…]
