### 2026-03-25 fb3485f
Taper shape PCA corrections to zero above 30 GeV
Wester (NEUT) CSV data only extends to ~28 GeV. Beyond that, Flat()
extrapolation creates artificial shape differences vs GENIE splines.
Apply smooth taper (linear fade 20-30 GeV) so shape corrections are
zero at high E, where only normalization uncertainty remains.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
