### 2026-03-25 23bdb3e
Blend Wester nominal to GENIE G18_10a above 20-30 GeV
Wester CSV data ends at ~28 GeV. Flat extrapolation beyond that caused
unphysical nominal σ/E: CC bending upward (frozen CCDIS value), NC
dropping to zero (CSV ended early). Fix by smoothly blending the nominal
from Wester (below 20 GeV) to GENIE G18_10a (above 30 GeV), which has
valid splines up to 1000 GeV.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
