### 2026-03-25 d88b4a1
Fix up/down energy scale: normalization not energy shift
The up/down energy scale systematic varies the relative normalization of
upgoing vs downgoing FC and PC events (Wester thesis Sec 5.2.3), not an
energy bin shift. Replace apply_energy_scale on upgoing bins with a
multiplicative normalization: upgoing bins get (1+δ), downgoing get (1-δ).

Masks are built in reco coszen space (+1 for CosZMax≤0, -1 for CosZMin≥0)
for FC+PC bins only (up-μ excluded as they are all upgoing by definition).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
