### 2026-03-26 fa681a1
Add ν̄/ν cross-section ratio parameters to H2O_PCA and --seed CLI option
Add 6 per-process nubar_ratio parameters to the H2O_PCA xsec model that
independently scale antineutrino cross-sections relative to neutrino ones.
Prior widths derived from GENIE tune spread of the ν̄/ν total xsec ratio
per process. Also widen NC norm prior from 15% to 20% to match Wester
thesis NC/CC uncertainty. This breaks the artificial ν/ν̄ correlation that
was forcing δCP to absorb cross-section asymmetries.

Add --seed argument to analysis.jl CLI that loads a previous
NewtrinosResult from a JLD2 file and uses its b
[…]
