+++ src/physics/xsec.jl
@@ -198,6 +198,12 @@ function get_params(cfg::H2O_PCA)
         xsec_ccdis_shape = 0.,
         xsec_ccother_shape = 0.,
         xsec_nc_shape = 0.,
+        xsec_cc1p1h_nubar_ratio = 1.,
+        xsec_cc2p2h_nubar_ratio = 1.,
+        xsec_cc1pi_nubar_ratio = 1.,
+        xsec_ccdis_nubar_ratio = 1.,
+        xsec_ccother_nubar_ratio = 1.,
+        xsec_nc_nubar_ratio = 1.,
     )
 end
 
@@ -208,7 +214,7 @@ function get_priors(cfg::H2O_PCA)
         xsec_cc1pi_norm = Truncated(Normal(1, 0.15), 0.4, 1.6),
         xsec_ccdis_norm = Truncated(Normal(1, 0.10), 0.5, 1.5),
         xsec_ccother_norm = Truncated(Normal(1, 0.30), 0.2, 1.8),
-        xsec_nc_norm = Truncated(Normal(1, 0.15), 0.4, 1.6),
+        xsec_nc_norm = Truncated(Normal(1, 0.20), 0.4, 1.6),
         xsec_nutau_cc_norm = Truncated(Normal(1, 0.25), 0.3, 1.7),
         xsec_cc1p1h_shape = Normal(0, 1),
         xsec_cc2p2h_shape = Normal(0, 1),
@@ -216,6 +222,13 @@ function get_priors(cfg::H2O_PCA)
         xsec_ccdis_shape = Normal(0, 1),
         xsec_ccother_shape = Normal(0, 1),
         xsec_nc_shape = Normal(0, 1),
+        # ν̄/ν ratio priors: σ from GENIE tune spread of per-process ν̄/ν xsec ratio
+        xsec_cc1p1h_nubar_ratio = Truncated(Normal(1, 0.17), 0.3, 1.7),
+        xsec_cc2p2h_nubar_ratio = Truncated(Normal(1, 0.17), 0.3, 1.7),
+        xsec_cc1pi_nubar_ratio = Truncated(Normal(1, 0.05), 0.7, 1.3),
+        xsec_ccdis_nubar_ratio = Truncated(Normal(1, 0.05), 0.7, 1.3),
+        xsec_ccother_nubar_ratio = Truncated(Normal(1, 0.10), 0.5, 1.5),
+        xsec_nc_nubar_ratio = Truncated(Normal(1, 0.05), 0.7, 1.3),
     )
 end
 
@@ -369,6 +382,13 @@ function get_scale(cfg::H2O_PCA)
         CCDIS  = :xsec_ccdis_shape,
         CCother = :xsec_ccother_shape,
     )
+    nubar_ratio_syms = (
+        CC1p1h = :xsec_cc1p1h_nubar_ratio,
+        CC2p2h = :xsec_cc2p2h_nubar_ratio,
+        CC1pi  = :xsec_cc1pi_nubar_ratio,
+        CCDIS  = :xsec_ccdis_nubar_ratio,
+        CCother = :xsec_ccother_nubar_ratio,
+    )
 
     function get_flavor_key(flav::Symbol, anti::Bool)
         if flav == :nue
@@ -385,15 +405,16 @@ function get_scale(cfg::H2O_PCA)
 
         if interaction == :NC
             # NC reweight: norm × (1 + ε × shape_PC(E)), clamped ≥ 0
-            # shape_PC differs for ν vs ν̄
+            # shape_PC differs for ν vs ν̄; nubar_ratio scales ν̄ relative to ν
             nc_shape = anti ? process_shape_itps["NC_nubar"] : process_shape_itps["NC_nu"]
[…]
