### 2026-03-26 d1f7217
Fix global scope ambiguity for p in seed loop
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
