### 2026-03-26 6a4042f
Replace loop with merge to avoid global scope issue in seed logic
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
