### 2026-03-26 497d850
Implement ray-to-spray oscillation averaging (Maltoni 2308.00037)
Replace ad-hoc Gaussian kernel smoothing with physics-based amplitude-level
averaging for Super-K. The Spray propagation model accumulates perturbation
matrices K_E and K_Θ through the layer-by-layer propagation, encoding how
the S-matrix changes with energy and zenith shifts. Their eigenspectra then
determine which interference terms survive bin-averaging via Gaussian damping
— fast oscillations are automatically suppressed while slow ones are preserved.

Key changes:
- Add Spray propagation type with K_E (energy) and K_Θ (zeni
[…]
