+++ test/test_regression.jl
@@ -9,10 +9,10 @@ using Printf
 const ALL_REFERENCE_VALUES = (
     dayabay      = -168.90003273508322,
     kamland      = -63.111403860037875,
-    minos        = -268.40461484198985,
-    deepcore     = -950.6200930498498,
-    super_k      = -3706.3913697750395,
-    orca         = -1164.3663831498247,
+    minos        = -268.4046154382636,
+    deepcore     = -950.625021522011,
+    super_k      = -3718.5900790556525,
+    orca         = -1164.3664813530054,
     coherent_csi = -574.3416032522271,
     coherent_lAr = -1754.9946939034153,
 )
