+++ src/analysis/plot_oscillograms.jl
@@ -1,0 +1,128 @@
+using Newtrinos
+using CairoMakie
+using LinearAlgebra
+using StatsBase: midpoints
+
+# --- Setup ---
+earth_layers = Newtrinos.earth_layers.configure()
+layers = earth_layers.compute_layers()
+
+osc_basic = Newtrinos.osc.configure(Newtrinos.osc.OscillationConfig(interaction=Newtrinos.osc.SI()))
+osc_spray = Newtrinos.osc.configure(Newtrinos.osc.OscillationConfig(interaction=Newtrinos.osc.SI(), propagation=Newtrinos.osc.Spray()))
+
+params = Newtrinos.get_params((; osc=osc_basic, earth_layers))
+
+# Grid: log10(E) from -1 to 2, cz from -1 to 0 (upgoing)
+loge_grid = LinRange(-1, 2, 301)
+cz_grid = LinRange(-1.0, 0.0, 201)
+
+E = 10.0 .^ midpoints(loge_grid)
+cz = midpoints(cz_grid)
+paths = earth_layers.compute_paths(cz, layers)
+dldcz = Newtrinos.earth_layers.compute_dldcz(cz, layers)
+
+dlogE = Float64(step(loge_grid))
+dCZ = Float64(step(cz_grid))
+
+# --- Compute oscillation probabilities ---
+
+# 1) No smoothing (Basic)
+println("Computing Basic (no smoothing)...")
+p_basic = osc_basic.osc_prob(E, paths, layers, params)
+p_basic_anti = osc_basic.osc_prob(E, paths, layers, params; anti=true)
+
+# 2) Gaussian blur (Basic + kernel smoothing, old Super-K method)
+println("Computing Gaussian blur...")
+K_E_mat = Newtrinos.super_k.make_gaussian_kernel_matrix(length(E), 0.05 / dlogE)
+K_CZ_mat = Newtrinos.super_k.make_gaussian_kernel_matrix(length(cz), 0.05 / dCZ)
+p_gauss = Newtrinos.super_k.smooth_osc_prob(p_basic, K_E_mat, K_CZ_mat)
+p_gauss_anti = Newtrinos.super_k.smooth_osc_prob(p_basic_anti, K_E_mat, K_CZ_mat)
+
+# 3) Spray E-only
+println("Computing Spray (E only)...")
+Delta_E = E .* dlogE .* log(10)
+p_spray_E = osc_spray.osc_prob(E, paths, layers, params; Delta_E)
+p_spray_E_anti = osc_spray.osc_prob(E, paths, layers, params; anti=true, Delta_E)
+
+# 4) Spray E+CZ
+println("Computing Spray (E+CZ)...")
+p_spray_ECZ = osc_spray.osc_prob(E, paths, layers, params; Delta_E, Delta_CZ=dCZ, dldcz)
+p_spray_ECZ_anti = osc_spray.osc_prob(E, paths, layers, params; anti=true, Delta_E, Delta_CZ=dCZ, dldcz)
+
+# --- Plot ---
+logE_mid = midpoints(loge_grid)
+cz_mid = midpoints(cz_grid)
+
+function make_oscillogram_figure(probs_list, titles, channel_name, channel_idx; clims=(0, 1))
+    n = length(probs_list)
+    fig = Figure(size=(400*n + 80, 420))
+    for (i, (p, title)) in enumerate(zip(probs_list, titles))
+        ax = Axis(fig[1, i];
+            xlabel = "log₁₀(E / GeV)",
+            ylabel = i == 1 ? "cos θ_z" : "",
+            title = title,
+            titlesize = 14,
+        )
[…]
