### 2026-03-26 dc90ca5
Optimize Spray propagation: 3.2x faster Super-K likelihood (501→157ms)
- SMatrix-ify compute_spray_layer and spray_average (eliminate heap allocs)
- Use fast_eigen (BargerEigen) for complex Hermitian 3×3 in spray_average
  instead of LAPACK eigen (22x faster per call, zero-alloc)
- Precompute dldcz in Super-K get_assets instead of recomputing per eval
- Increase spray smearing widths to 10% (Delta_E, Delta_CZ)

Allocations: 2.4M → 26K, memory: 388 → 31 MiB (matches Basic).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
