### 2026-03-26 e0a56cc
Fix dldcz layer boundary artifacts, increase spray smearing to 10%
- compute_dldcz now uses total path length derivative distributed
  proportionally to each section, avoiding divergent per-section
  derivatives near layer boundaries
- Update oscillogram script and Super-K smearing to 10% (Delta_E, Delta_CZ)
- Update Super-K regression reference value

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
