### 2026-03-27 02a76dd
Move spray σ_E and σ_h into Spray config, set Super-K σ_E=25%
- Add σ_E (fractional energy smearing) and σ_h (production height
  uncertainty in km) fields to Spray struct with defaults 0.15 and 10.0
- Compute Delta_E and dldh inside propagate() from config, removing
  Delta_E/Delta_CZ/dldcz kwargs from osc_prob
- Production height: only first (atmosphere) layer varies (dL/dh=1),
  replacing the old cos-zenith dldcz approach
- Super-K uses σ_E=0.25 (25%), σ_h=10km (default)
- Clean up earth_layers.compute_dldcz (no longer used by super_k)
- Fix oscillogram plot axis orientation

Co-Authored-By
[…]
