+++ test/test_regression.jl
@@ -11,7 +11,7 @@ const ALL_REFERENCE_VALUES = (
     kamland      = -63.111403860037875,
     minos        = -268.4046154382636,
     deepcore     = -950.625021522011,
-    super_k      = -3714.2067977827737,
+    super_k      = -3713.422980775823,
     orca         = -1164.3664813530054,
     coherent_csi = -574.3416032522271,
     coherent_lAr = -1754.9946939034153,
