### 2026-03-27 5c892a6
Switch Super-K to Basic + 5% Gaussian blur for testing
- Revert to Basic propagation (no Spray) for Super-K
- Add 5% Gaussian blur in logE and CZ on oscillation probabilities
- Precompute blur kernel matrices in get_assets

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
