+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -24,7 +24,7 @@ using ..Newtrinos
 end
 
 function default_physics()
-    osc = Newtrinos.osc.configure(Newtrinos.osc.OscillationConfig(interaction=Newtrinos.osc.SI(), propagation=Newtrinos.osc.Spray(σ_E=0.25)))
+    osc = Newtrinos.osc.configure(Newtrinos.osc.OscillationConfig(interaction=Newtrinos.osc.SI()))
     atm_flux = Newtrinos.atm_flux.configure(Newtrinos.atm_flux.AtmFluxConfig(nominal_model=Newtrinos.atm_flux.HKKM("kam-ally-20-01-mtn-solmin.d")))
     earth_layers = Newtrinos.earth_layers.configure(Newtrinos.earth_layers.VariableDensity())
     xsec = Newtrinos.xsec.configure(Newtrinos.xsec.H2O_PCA())
@@ -262,8 +262,8 @@ function calc_weights(params, assets, physics)
     layers = haskey(params, :electron_density_scale) ? Newtrinos.earth_layers.scale_densities(assets.nominal_layers, params.electron_density_scale) : assets.nominal_layers
     paths = physics.earth_layers.compute_paths(assets.cz_midpoints, layers)
 
-    p = physics.osc.osc_prob(E, paths, layers, params);
-    p_anti = physics.osc.osc_prob(E, paths, layers, params; anti=true);
+    p = smooth_osc_prob(physics.osc.osc_prob(E, paths, layers, params), assets.K_E_blur, assets.K_CZ_blur);
+    p_anti = smooth_osc_prob(physics.osc.osc_prob(E, paths, layers, params; anti=true), assets.K_E_blur, assets.K_CZ_blur);
 
     flux = physics.atm_flux.sys_flux(assets.flux_nominal, params)
 
@@ -403,13 +403,21 @@ function get_assets(physics; datadir = @__DIR__)
     paths = physics.earth_layers.compute_paths(cz_midpoints, nominal_layers)
     flux_nominal = physics.atm_flux.nominal_flux(10. .^midpoints(loge_grid), cz_midpoints)
 
+    # Gaussian blur kernels for oscillation probability smoothing (5% in logE and CZ)
+    n_E = length(midpoints(loge_grid))
+    n_CZ = length(cz_midpoints)
+    dlogE = Float64(step(loge_grid))
+    dCZ = Float64(step(cz_grid))
+    K_E_blur = make_gaussian_kernel_matrix(n_E, 0.05 / dlogE)
+    K_CZ_blur = make_gaussian_kernel_matrix(n_CZ, 0.05 / dCZ)
+
     flatten_R(R3d) = NamedTuple(key => reshape(R3d[key], size(R3d[key], 1), :) for key in keys(R3d))
 
     # Build response matrices with 1% broadened quantile spreads to account for
     # uncertainty in the reconstruction distribution (known only from 5 quantiles)
     R_3d = NamedTuple(key => make_response_matrix(MC[key], loge_grid, cz_grid; resolution_scale=1.01) for key in keys(MC))
     R = flatten_R(R_3d)
-    nominal_weights = calc_weights(params_nominal, (;R, flux_nominal, paths, nominal_layers, loge_grid, cz_grid, cz_midpoints), physics)
+    nominal_weights = calc_weights(params_nominal, (;R, flux_nominal, paths, nominal_layers, loge_grid, cz_grid, cz_midpoints, K_E_blur, K_CZ_blur), physics)
 
     # Old F_ij method (commented out — replaced by reco bin overlap method)
     #loge_grid_plus = loge_grid .+ log10(1.02)
[…]
