### 2026-03-27 1d7d617
Remove R matrix renormalization, add linear-E CDF option for Super-K
- Remove response matrix row renormalization: events whose true energy
  falls outside the analysis grid (0.1-1000 GeV) are now correctly lost
  rather than artificially redistributed into the grid. This primarily
  affects upmu_thru bins where 8-40% of events are above 1 TeV.
- Add make_e_cdf() for building CDF splines in linear E (vs logE),
  configurable via energy_cdf=:linearE keyword in configure().
- Add R_3d (pre-flattened response matrix) to assets for diagnostics.
- Update regression test reference for Super-K.

Co-A
[…]
