+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -31,9 +31,9 @@ function default_physics()
     (; osc, atm_flux, earth_layers, xsec)
 end
 
-function configure(physics=default_physics())
+function configure(physics=default_physics(); energy_cdf=:logE)
     physics = (;physics.osc, physics.atm_flux, physics.earth_layers, physics.xsec)
-    assets = get_assets(physics)
+    assets = get_assets(physics; energy_cdf)
     return SuperKAtm(
         physics = physics,
         params = get_params(),
@@ -85,6 +85,46 @@ function make_log_e_cdf(bin; resolution_scale=1.0)
     return f_save
 end
 
+function make_e_cdf(bin; resolution_scale=1.0)
+    e = [bin.EnergyQuantile2_3Percent, bin.EnergyQuantile15_9Percent, bin.EnergyQuantile50_0Percent, bin.EnergyQuantile84_1Percent, bin.EnergyQuantile97_7Percent]
+
+    # Scale quantile distances from median to broaden/narrow the resolution
+    median = e[3]
+    e = median .+ resolution_scale .* (e .- median)
+
+    # Extrapolate tails, ensuring strict monotonicity
+    lo = max(0.0, 2*e[1] - e[2])
+    if lo >= e[1]
+        lo = e[1] * 0.5  # fallback: half the lowest quantile
+    end
+    hi = 2*e[end] - e[end-1]
+    energy_quantiles = [lo, e..., hi]
+    quantile_probs = [0.0, 0.023, 0.159, 0.5, 0.841, 0.977, 1.]
+
+    # Ensure strict monotonicity by nudging any non-increasing knots
+    for i in 2:length(energy_quantiles)
+        if energy_quantiles[i] <= energy_quantiles[i-1]
+            energy_quantiles[i] = energy_quantiles[i-1] + 1e-6
+        end
+    end
+
+    dy_dx = MonotonicSplines.estimate_dYdX(energy_quantiles, quantile_probs)
+    dy_dx[1] = 0
+    dy_dx[end] = 0
+    f = RQSpline(energy_quantiles, quantile_probs, dy_dx)
+
+    f_save = x -> begin
+        if x < energy_quantiles[1]
+            return 0.0
+        elseif x > energy_quantiles[end]
+            return 1.0
+        else
+            return f(x)
+        end
+    end
+
+    return f_save
+end
 
 function make_cosz_cdf(bin; resolution_scale=1.0)
     cosz = [bin.CosZQuantile2_3Percent, bin.CosZQuantile15_9Percent, bin.CosZQuantile50_0Percent, bin.CosZQuantile84_1Percent, bin.CosZQuantile97_7Percent]  # extrapolate tails
@@ -114,12 +154,15 @@ function make_cosz_cdf(bin; resolution_scale=1.0)
     return f_save
 end
 
-function make_response_matrix(MC_component, logE_grid, cosZ_grid; resolution_scale=1.0)
+function make_response_matrix(MC_component, logE_grid, cosZ_grid; resolution_scale=1.0, energy_cdf=:logE)
     n_bins = size(MC_component, 1)
     n_logE = length(logE_grid)
     n_cosZ = length(cosZ_grid)
 
     response_matrix = zeros(Float64, n_bins, n_logE-1, n_cosZ-1)
+
+    # For linear-E CDF, convert logE grid edges to linear E once
+    E_grid = energy_cdf == :linearE ? 10 .^ logE_grid : nothing
 
     for bin_idx in 1:n_bins
         bin = MC_component[bin_idx, :]
@@ -128,22 +171,21 @@ function make_response_matrix(MC_component, logE_grid, cosZ_grid; resolution_sca
[…]
