### 2026-03-27 74e853e
Fix NC flux treatment: use total neutrino flux instead of flat weights
NC interactions are flavor-blind, so the NC rate depends on the total
neutrino flux (nue + numu + nuebar + numubar). Previously used flat
weights (ones), which meant flux systematics were not properly
propagated to the NC component during profiling.

At nominal parameters the change is negligible (ratio w/w_nominal = 1),
but it matters when flux systematics are varied from nominal.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
