### 2026-03-27 21e5f60
new plots for thesis
