+++ config.txt
@@ -4,15 +4,14 @@
 
 # EXPERIMENT SELECTION
 
-experiments = katrin
-
+experiments = gerda
 # OSCILLATION MODEL
 
-model = NND
+model = NNM
 
 # MASS ORDERING
 # Choose ordering: NO (Normal) or IO (Inverted)
-ordering =IO
+ordering =NO
 
 # ============================================================
 # SCAN PARAMETERS
