### 2026-03-28 6096bd5
Overhaul Super-K response matrix: vMF cosZ + DSCB energy + Spray propagation
Replace the spline-based response matrix construction with physically
motivated distributions:

- cosZ: von Mises-Fisher marginal distribution (proper spherical geometry)
  fitted from unoscillated MC quantiles, stored in vmf_cosz_params.jld2
- Energy: Double-sided Crystal Ball in log(E) space (Gaussian core with
  power-law tails), fitted from unoscillated MC, stored in
  energy_response_params.jld2
- Oscillation smoothing: switch from post-hoc Gaussian blur to Spray
  propagation with σ_E matched to grid Nyquist and
[…]
