### 2026-03-29 503f7b6
Fix nutau CC and NC reweighting: proper nu/nubar mixture fractions
The SK data release lumps nutau+nutaubar CC and all NC into single
channels. Previously, nutau weights only included neutrino oscillations
(missing antineutrino contribution), and NC used a single cross-section
for both nu and nubar.

Now:
- Precompute energy-dependent nu/nubar mixture fractions from nominal
  flux × actual cross-sections (using numu xsec as proxy for nutau)
- Use these fractions in calc_weights to properly combine nu and nubar
  oscillated fluxes for nutau CC and NC channels
- NC now uses separate nu and nubar
[…]
