+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -11,6 +11,7 @@ using BAT
 using LaTeXStrings
 using Accessors
 using StatsBase
+using Statistics: mean
 using Printf
 using SpecialFunctions: besseli
 using JLD2
@@ -864,14 +865,23 @@ function calc_weights(params, assets, physics)
 
     nue     = contract_R(assets.R.nue,     nue_flux)
     numu    = contract_R(assets.R.numu,    numu_flux)
-    nutau   = contract_R(assets.R.nutau,   nutau_flux)
     nuebar  = contract_R(assets.R.nuebar,  nuebar_flux)
     numubar = contract_R(assets.R.numubar, numubar_flux)
-    # NC is flavor-blind: total flux is conserved under oscillations (unitarity),
-    # but still depends on flux systematics (Barr params, normalization)
-    flux_total_nc = (reshape(flux.nue, s) .+ reshape(flux.numu, s) .+
-                     reshape(flux.nuebar, s) .+ reshape(flux.numubar, s))
-    nunc    = contract_R(assets.R.nunc,    flux_total_nc .* xsec_nc .* flux_norm)
+
+    # nutau: SK MC lumps nutau + nutaubar CC into one channel.
+    # Use precomputed nu/nubar mixture fractions to combine with correct proportions.
+    nutau_combined = nutau_flux .* assets.nutau_nu_frac .+ nutaubar_flux .* (1 .- assets.nutau_nu_frac)
+    nutau   = contract_R(assets.R.nutau,   nutau_combined)
+
+    # NC: SK MC lumps all NC (nu + nubar, all flavors) into one channel.
+    # Use precomputed nu/nubar mixture fractions with proper cross-sections.
+    xsec_nc_anti = physics.xsec.scale(E, :nue, :NC, true, params)
+    flux_nu_total = reshape(flux.nue, s) .+ reshape(flux.numu, s)
+    flux_nubar_total = reshape(flux.nuebar, s) .+ reshape(flux.numubar, s)
+    nc_nu_flux = flux_nu_total .* xsec_nc
+    nc_nubar_flux = flux_nubar_total .* xsec_nc_anti
+    nc_combined = nc_nu_flux .* assets.nc_nu_frac .+ nc_nubar_flux .* (1 .- assets.nc_nu_frac)
+    nunc    = contract_R(assets.R.nunc,    nc_combined .* flux_norm)
 
     return (; nue, numu, nutau, nuebar, numubar, nunc)
 end
@@ -997,7 +1007,54 @@ function get_assets(physics; datadir = @__DIR__, energy_cdf=:logE)
             vmf_mu_z=vmf_params[key]["mu_z"], vmf_kappa=vmf_params[key]["kappa"]) for key in keys(MC))
     end
     R = flatten_R(R_3d)
-    nominal_weights = calc_weights(params_nominal, (;R, flux_nominal, paths, nominal_layers, loge_grid, cz_grid, cz_midpoints), physics)
+
+    # Compute nu/nubar mixture fractions for nutau CC and NC channels.
+    # The SK MC lumps nutau+nutaubar CC, and all NC, into single channels.
+    # We estimate the nu vs nubar fraction from nominal flux * actual cross-sections.
+    # For nutau CC, use numu cross-sections (nutau xsec not available separately).
+    E_mid = 10.0 .^ midpoints(loge_grid)
+
+    # Get actual cross-section curves from the xsec data file
[…]
