### 2026-03-29 a32f327
Add per-bin MC-stats-based energy and cosZ blur to R matrix
For bins with low MC statistics, the quantile-based R matrix is noisy.
Add Gaussian blur to both energy and cosZ dimensions of each R matrix
row, with width estimated from the MC statistics:

  sigma ~ IQR / sqrt(12) * sqrt(min(10, N_MC) / N_MC)

where N_MC ~ 5 * reported_counts (assumed MC/data ratio).
This mimics SK's per-event oscillation probability averaging using
~10 nearest neighbors.

Applied always (no threshold), but naturally vanishes for high-stats
bins where sigma_blur << bin width.

Co-Authored-By: Claude Opus 4.6 (1M co
[…]
