### 2026-03-29 fe87dee
Add R matrix coverage correction for out-of-grid events
Events whose true energy falls outside the logE grid [-1, 3] are not
captured by the R matrix (coverage < 1). Previously these events were
implicitly reweighted as if they were in-grid, causing over-correction
for bins near the grid edges (especially upmu_thru at high energy).

Now the reweight formula accounts for out-of-grid events:
  reweighted = MC * [f * (w_new/w_nom) + (1-f)]
where f = sum(R_row) is the coverage fraction per bin per flavor.

This correctly assumes out-of-grid events keep their nominal weight.
Upmu_thru numu coverage
[…]
