### 2026-03-29 0428cf4
Fix MC-stats blur: use reco bin width instead of quantile spread
The per-bin Gaussian blur should use the reconstructed momentum/cosZ
bin width (from bininfo), not the true energy quantile spread. The reco
bin width determines how spread out MC events are in true space.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
