### 2026-03-29 3abdbbd
Fix missing Jacobian in flux integration on logE grid
The HKKM atmospheric flux is differential in energy: Φ(E) in
(m² s sr GeV)⁻¹. On our logE grid, the bin content should be
Φ(E) × E × ln(10) × dlogE, but we were missing the factor of E
(the Jacobian dE/dlogE ∝ E).

This shifts the effective nue/numu flux ratio across the energy range,
causing systematic biases in the oscillation reweighting — particularly
for the nue appearance bins where the survival/appearance cancellation
is sensitive to the flux balance.

NO→IO roundtrip chi2 improved from 2.79 to 2.20 (direct: 2.42).
Bin 282 nue underc
[…]
