+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -650,20 +650,16 @@ function _build_R_from_params(MC_component, logE_grid, cosZ_grid, dscb_params, v
         p_e = diff(c_e)
 
         # Per-bin energy blur based on MC statistics
-        # SK uses ~10 nearest neighbors for per-event osc prob averaging.
-        # With N_MC events spanning IQR, the effective smearing is:
-        #   sigma_blur ~ IQR / sqrt(12) * sqrt(min(10, N_MC) / N_MC)
-        # In logE bins: sigma_blur_bins = sigma_blur_logE / dlogE
+        # SK uses ~10 nearest neighbors in energy for per-event osc prob averaging.
+        # With N_MC events distributed across a reco bin of width delta_logP,
+        # the 10 nearest neighbors span ~ delta_logP * min(10, N_MC) / N_MC.
+        # sigma ~ that span / sqrt(12).
         n_mc = mc_data_ratio * counts
-        iqr_logE = dscb_params["sigma"][bin_idx] * 2.0  # rough: DSCB sigma ~ half IQR in logE
         n_neighbors = min(10.0, n_mc)
-        sigma_blur_logE = iqr_logE / sqrt(12.0) * sqrt(n_neighbors / max(n_mc, 1.0))
+        sigma_blur_logE = reco_logP_width[bin_idx] * sqrt(n_neighbors / max(n_mc, 1.0)) / sqrt(12.0)
         sigma_blur_bins = sigma_blur_logE / dlogE
-
-        if sigma_blur_bins > 0.5  # only blur if significant (> half a bin)
-            K = make_gaussian_kernel_matrix(n_logE, sigma_blur_bins)
-            p_e = K * p_e
-        end
+        K_e = make_gaussian_kernel_matrix(n_logE, sigma_blur_bins)
+        p_e = K_e * p_e
 
         # CosZ: vMF CDF
         kappa = vmf_params["kappa"][bin_idx]
@@ -674,6 +670,13 @@ function _build_R_from_params(MC_component, logE_grid, cosZ_grid, dscb_params, v
             c_cosz = collect(range(0, 1, length=length(cosZ_grid)))
         end
         p_cosz = diff(c_cosz)
+
+        # CosZ blur: same logic with reco cosZ bin width
+        sigma_blur_cz = reco_cz_width[bin_idx] * sqrt(n_neighbors / max(n_mc, 1.0)) / sqrt(12.0)
+        dCZ = Float64(cosZ_grid[2] - cosZ_grid[1])
+        sigma_blur_cz_bins = sigma_blur_cz / dCZ
+        K_cz = make_gaussian_kernel_matrix(n_cosZ, sigma_blur_cz_bins)
+        p_cosz = K_cz * p_cosz
 
         response_matrix[bin_idx, :, :] .= p_e * p_cosz'
     end
@@ -869,18 +872,26 @@ function calc_weights(params, assets, physics)
     xsec_nutaubar= physics.xsec.scale(E, :nutau, :CC, true,  params)
     xsec_nc      = physics.xsec.scale(E, :nue,   :NC, false, params)
 
-    nue_flux   = (reshape(flux.nue,    s) .* p[:, :, 1, 1] .+
-                  reshape(flux.numu,   s) .* p[:, :, 2, 1]) .* xsec_nue .* flux_norm
-    numu_flux  = (reshape(flux.nue,    s) .* p[:, :, 1, 2] .+
-                  reshape(flux.numu,   s) .* p[:, :, 2, 2]) .* xsec_numu .* flux_norm
[…]
