### 2026-03-29 51f84bd
Update Super-K regression reference after flux Jacobian fix
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
