### 2026-03-29 ae21fae
Add second E factor for cross-section σ/E convention
The xsec data is stored as σ/E (cross-section per nucleon divided by
energy). The xsec.scale() function returns dimensionless ratios relative
to this σ/E baseline. To get the actual cross-section σ, we need an
additional factor of E.

Combined with the flux Jacobian (previous commit), the full weighting
on the logE grid is:
  event rate ∝ Φ(E)/GeV × (σ/E × E) × (E × dlogE) = Φ × σ × E × dlogE

So E² total: one from dE/dlogE Jacobian, one from σ/E → σ.

NO→IO roundtrip chi2: 2.17 → 1.77 (direct: 2.42, ratio: 0.73).
Bin 282 nue residual: +1.48
[…]
