### 2026-03-29 7852c2b
Update Super-K regression reference after E² flux weighting fix
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
