### 2026-03-30 c726e44
Revert coverage correction in reweight — let all events be reweighted
Clamping out-of-grid events to weight=1.0 over-constrains the
likelihood by removing all freedom for those event fractions.
Better to let the reweight ratio apply uniformly, even if slightly
inaccurate for out-of-grid events, than to lock them down.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
