### 2026-03-30 91955f6
Split PC stop/thru systematic into top/barrel/bottom exit directions
SK thesis (Sec 5.2.3) implements three independent uncertainties for the
PC stopping/through-going separation based on muon exit point (top,
barrel, bottom of detector). We had a single parameter covering all
directions.

The split gives the fit freedom to create directional asymmetry in the
PC sample — top-exiting events are downgoing, bottom-exiting are upgoing.
This directly affects the up/down ratio that constrains θ₂₃.

Bins assigned by cosZ midpoint: bottom (< -0.22), barrel (±0.22),
top (> 0.22), giving 4/2/4 bins per 
[…]
