### 2026-03-30 b4ad690
Make xsec nubar_ratio normalization-conserving
Previously, nubar_ratio only scaled the antineutrino cross-section
without compensating on the neutrino side. Now both ν and ν̄ are
scaled symmetrically: ν̄ × 2r/(1+r), ν × 2/(1+r), preserving the
total rate while shifting the ν/ν̄ balance.

At nominal (r=1), both factors equal 1.0 — no change to any
experiment's prediction. Applied to both CC (per channel) and NC.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
