### 2026-03-31 8d6f0ab
Switch Super-K to Differential_H2O xsec with nubar_ratio parameter
Add nubar_ratio parameter (5% uncertainty, normalization-conserving) to
the Differential_H2O cross-section model and switch Super-K from H2O_PCA
to Differential_H2O for simpler, less degenerate xsec parameterization.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
