### 2026-03-31 9bd3743
Add --nseeds for multi-start MLE fitting from randomized initial values
Each scan point spawns nseeds independent optimization jobs (seed 1 uses
nominal params, seeds 2..N sample from the prior). All jobs run as
independent work units via threads or Distributed.jl. The best fit per
point (highest log_posterior) is selected for the final result.

Uses a fixed RNG seed (Xoshiro(42)) so randomized starting values are
reproducible across interrupted/resumed runs, preserving cache hits.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
