### 2026-03-31 485c6d0
Keep original code path for nseeds=1, only randomize when nseeds > 1
nseeds=1 now uses the exact original _profile logic (single shared
params, original map_func signature) with no randomization. The
multi-seed expansion and best-fit selection only activates for nseeds>1,
where all seeds are randomized from the prior.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
