### 2026-03-31 a726126
Fix Xoshiro: add Random stdlib dependency and import
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
