+++ Project.toml
@@ -63,6 +63,7 @@ PolyesterForwardDiff = "98d1487c-24ca-40b6-b7ab-df2af84e126b"
 PositiveFactorizations = "85a6dd25-e78a-55b7-8502-1745935b8125"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
@@ -133,6 +134,7 @@ PolyesterForwardDiff = "0.1.3"
 PositiveFactorizations = "0.2.4"
 Printf = "1.11.0"
 ProgressMeter = "1.10.4"
+Random = "1.11.0"
 SpecialFunctions = "2.5.1"
 StaticArrays = "1.9.13"
 Statistics = "1.11.1"
