### 2026-04-01 78f3229
Fix Δχ² profile plot: use pure likelihood, not posterior
The profile Δχ² was computed from log_posterior (likelihood + prior),
which includes Gaussian penalty terms from all nuisance parameters.
As nuisance params shift during profiling, the cumulative prior penalty
inflates the Δχ² — roughly doubling it when ~50 constrained systematics
each contribute small pulls.

The standard HEP approach: optimize the posterior (prior constrains
nuisance params), but compute the test statistic from the pure
likelihood ratio. Now uses result.values.llh for Δχ² plots.

Co-Authored-By: Claude Opus 4.6 (1M con
[…]
