### 2026-04-01 f34bc39
Revert "Fix Δχ² profile plot: use pure likelihood, not posterior"
This reverts commit 78f32296b0893e5e77bdf0392f9df74d74103665.
