### 2026-04-01 68e69e6
Split flux ratio systematics into sub-GeV / 1-10 GeV / >10 GeV ranges
Replace single-parameter flux ratio uncertainties with per-energy-range
parameters using hard cuts in true energy:
- nuenuebar_sigma → _lo (E<1), _mid (1-10), _hi (≥10 GeV)
- numunumubar_sigma → same split
- nuenumu_sigma → same split
- norm_ratio → _lo (E<1), _hi (≥1 GeV)

Goes from 8 to 15 flux parameters. Each sub-range keeps the same prior
as the original. At nominal values, behavior is identical to before.

Adds energy-dependent freedom for the fit to absorb oscillation-like
effects at different energies, reducing artif
[…]
