+++ src/physics/atm_flux.jl
@@ -45,26 +45,40 @@ end
 
 function get_params(cfg::Barr)
     params = (
-        atm_flux_nuenuebar_sigma = 0.,
-        atm_flux_numunumubar_sigma = 0.,
-        atm_flux_nuenumu_sigma = 0.,
+        atm_flux_nuenuebar_sigma_lo = 0.,
+        atm_flux_nuenuebar_sigma_mid = 0.,
+        atm_flux_nuenuebar_sigma_hi = 0.,
+        atm_flux_numunumubar_sigma_lo = 0.,
+        atm_flux_numunumubar_sigma_mid = 0.,
+        atm_flux_numunumubar_sigma_hi = 0.,
+        atm_flux_nuenumu_sigma_lo = 0.,
+        atm_flux_nuenumu_sigma_mid = 0.,
+        atm_flux_nuenumu_sigma_hi = 0.,
         atm_flux_delta_spectral_index = 0.,
         atm_flux_uphorizontal_sigma = 0.,
         atm_flux_updown_sigma = 0.,
-        atm_flux_norm_ratio = 1.,
+        atm_flux_norm_ratio_lo = 1.,
+        atm_flux_norm_ratio_hi = 1.,
         atm_flux_K_pi_ratio = 1.,
         )
 end
 
 function get_priors(cfg::Barr)
     priors = (
-        atm_flux_nuenuebar_sigma = Truncated(Normal(0., 1.), -3, 3),
-        atm_flux_numunumubar_sigma = Truncated(Normal(0., 1.), -3, 3),
-        atm_flux_nuenumu_sigma = Truncated(Normal(0., 1.), -3, 3),
+        atm_flux_nuenuebar_sigma_lo = Truncated(Normal(0., 1.), -3, 3),
+        atm_flux_nuenuebar_sigma_mid = Truncated(Normal(0., 1.), -3, 3),
+        atm_flux_nuenuebar_sigma_hi = Truncated(Normal(0., 1.), -3, 3),
+        atm_flux_numunumubar_sigma_lo = Truncated(Normal(0., 1.), -3, 3),
+        atm_flux_numunumubar_sigma_mid = Truncated(Normal(0., 1.), -3, 3),
+        atm_flux_numunumubar_sigma_hi = Truncated(Normal(0., 1.), -3, 3),
+        atm_flux_nuenumu_sigma_lo = Truncated(Normal(0., 1.), -3, 3),
+        atm_flux_nuenumu_sigma_mid = Truncated(Normal(0., 1.), -3, 3),
+        atm_flux_nuenumu_sigma_hi = Truncated(Normal(0., 1.), -3, 3),
         atm_flux_delta_spectral_index = Truncated(Normal(0., 0.1), -0.3, 0.3),
         atm_flux_uphorizontal_sigma = Truncated(Normal(0., 1.), -3, 3),
         atm_flux_updown_sigma = Truncated(Normal(0., 1.), -3, 3),
-        atm_flux_norm_ratio = Normal(1, 0.15),
+        atm_flux_norm_ratio_lo = Normal(1, 0.15),
+        atm_flux_norm_ratio_hi = Normal(1, 0.15),
         atm_flux_K_pi_ratio = Normal(1, 0.1),
         )
 end
@@ -145,30 +159,46 @@ function get_sys_flux(cfg::Barr)
         log10e = flux.log10_true_energy
         cz = flux.true_coszen
 
+        # Energy range masks: sub-GeV (E < 1), 1-10 GeV, >10 GeV
+        mask_lo = e .< 1
+        mask_hi = e .>= 10
+        mask_mid = .!mask_lo .& .!mask_hi
+
         # spectral
         f_spectral_shift = (e ./ 24.0900951261) .^ params.atm_flux_delta_spectral_index
 
[…]
