### 2026-04-01 5e2f781
Switch Super-K default energy CDF to DSCB
The DSCB (Double-Sided Crystal Ball) energy CDF provides a better fit
to the per-bin energy response quantiles than the simple logE CDF.
Uses precomputed DSCB parameters for faster R matrix construction.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
