### 2026-04-09 fe34199
Add dσdE method to xsec module, use absolute cross-sections in Super-K
Add dσdE field to Xsec struct that returns actual σ/E values (not
dimensionless reweight ratios). Implemented for Differential_H2O using
the per-channel Wester CSV data with all systematic modifications
(MA_QE, MA_Res, FSI, channel norms, nubar_ratio).

Super-K calc_weights now uses physics.xsec.dσdE() instead of
physics.xsec.scale(), giving proper forward-folded cross-sections.

Also restores Differential_H2O as Super-K default (was accidentally
reverted to H2O_PCA during earlier revert).

Co-Authored-By: Claude Opus 4.6 (
[…]
