+++ src/physics/xsec.jl
@@ -25,6 +25,7 @@ const GENIE_H2O = H2O_PCA  # backward compat alias
     params::NamedTuple
     priors::NamedTuple
     scale::Function
+    dσdE::Union{Function, Nothing} = nothing
 end
 
 
@@ -33,9 +34,12 @@ function configure(cfg::XsecModel=SimpleScaling())
         cfg=cfg,
         params = get_params(cfg),
         priors = get_priors(cfg),
-        scale = get_scale(cfg)
+        scale = get_scale(cfg),
+        dσdE = get_dσdE(cfg)
         )
 end
+
+get_dσdE(::XsecModel) = nothing
 
 function get_params(cfg::SimpleScaling)
     (
@@ -186,6 +190,67 @@ function get_scale(cfg::Differential_H2O)
                 return s
             end
         end
+    end
+end
+
+function get_dσdE(cfg::Differential_H2O)
+
+    # Same CSV data as get_scale — σ/E per channel on water
+    df_nue = CSV.read(joinpath(@__DIR__, "xsec_nue_water.csv"), DataFrame, skipto=3);
+    df_nuebar = CSV.read(joinpath(@__DIR__, "xsec_nuebar_water.csv"), DataFrame, skipto=3);
+
+    function make_interpolation(name, df)
+        idx = findfirst(==(name), names(df))
+        x = collect(skipmissing(df[:,idx]))
+        y = collect(skipmissing(df[:,idx+1]))
+        itp = interpolate((x,), y, Gridded(Linear()))
+        m(x) = max.(0, x)
+        return m ∘ extrapolate(itp, Linear())
+    end
+
+    nue = (
+        CC1p1h = make_interpolation("CC1p1h", df_nue),
+        CC2p2h = make_interpolation("CC2p2h", df_nue),
+        CC1pi = make_interpolation("CC1pi", df_nue),
+        CCother = make_interpolation("CCother", df_nue),
+        CCDIS = make_interpolation("CCDIS", df_nue),
+        NC = make_interpolation("NC", df_nue),
+    )
+
+    nuebar = (
+        CC1p1h = make_interpolation("CC1p1h", df_nuebar),
+        CC2p2h = make_interpolation("CC2p2h", df_nuebar),
+        CC1pi = make_interpolation("CC1pi", df_nuebar),
+        CCother = make_interpolation("CCother", df_nuebar),
+        CCDIS = make_interpolation("CCDIS", df_nuebar),
+        NC = make_interpolation("NC", df_nuebar),
+    )
+
+    function dσdE(E::AbstractArray, flav::Symbol, interaction::Symbol, anti::Bool, params::NamedTuple)
+        funs = anti ? nuebar : nue
+
+        if interaction == :NC
+            return funs.NC.(E) .* params.nc_norm
+        end
+
+        ma_qe = ma_qe_ratio.(E, params.xsec_MA_QE)
+        ma_res = ma_res_ratio.(E, params.xsec_MA_Res)
+        fsi_1p1h = 1 .- 0.1 .* (params.xsec_fsi - 1)
+        fsi_1pi = 1 .+ 0.1 .* (params.xsec_fsi - 1)
+
+        s = (funs.CC1p1h.(E) .* params.cc1p1h_norm .* ma_qe .* fsi_1p1h .+
+             funs.CC2p2h.(E) .* params.cc2p2h_norm .+
+             funs.CC1pi.(E) .* params.cc1pi_norm .* ma_res .* fsi_1pi .+
+             funs.CCother.(E) .* params.ccother_norm .* params.xsec_I12 .+
+             funs.CCDIS.(E) .* params.ccdis_norm) .* params.cc_norm
+
+        r = params.nubar_ratio
[…]
