### 2026-04-09 0d5833e
Clean up: deduplicate xsec.jl, remove dead code, fix export typo
xsec.jl: Extract shared CSV loading and interpolation setup into
_load_h2o_interpolations(), and shared systematic application into
_apply_h2o_systematics(). Both get_scale and get_dσdE for
Differential_H2O now use these helpers instead of duplicating ~40 lines.

super_k.jl: Remove ~45 lines of commented-out F_ij code (old energy
scale method replaced by reco bin overlap approach).

Newtrinos.jl: Fix export typo whack_a_moles → whack_a_mole.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
