+++ src/Newtrinos.jl
@@ -24,7 +24,7 @@ abstract type Experiment end
 export Physics, Experiment
 export NewtrinosResult, plot
 export bin, rebin
-export make_init_samples, make_prior_samples, whack_a_moles, whack_many_moles
+export make_init_samples, make_prior_samples, whack_a_mole, whack_many_moles
 
 include("physics/osc.jl")
 using .osc
