### 2026-04-09 9b85453
Remove dead R matrix code, consolidate response param generation
super_k.jl: Remove ~430 lines of unused response matrix code:
- metalog, novosibirsk, log-spline, linear-spline energy CDFs
- make_cosz_cdf (spline fallback), make_vmf_cosz_cdf
- make_response_matrix (generic dispatcher)
- make_dscb_e_cdf, smooth_osc_prob
- energy_cdf parameter from configure/get_assets

Only the DSCB energy + vMF cosZ path remains, using precomputed
parameters from JLD2 files via _build_R_from_params.

Replace generate_cz_response.jl + generate_e_response.jl with a
single generate_response_params.jl that fits bo
[…]
