+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -39,9 +39,9 @@ function default_physics()
     (; osc, atm_flux, earth_layers, xsec)
 end
 
-function configure(physics=default_physics(); energy_cdf=:dscb)
+function configure(physics=default_physics())
     physics = (;physics.osc, physics.atm_flux, physics.earth_layers, physics.xsec)
-    assets = get_assets(physics; energy_cdf)
+    assets = get_assets(physics)
     return SuperKAtm(
         physics = physics,
         params = get_params(),
@@ -62,74 +62,6 @@ function read_sk_file(filepath::String)
     ])
     return df
 end
-
-"""
-    fit_metalog(qv, qp)
-
-Fit a 5-term metalog distribution from 5 quantile points.
-The metalog quantile function is:
-  Q(y) = a₁ + a₂·ln(y/(1-y)) + a₃·(y-0.5)·ln(y/(1-y)) + a₄·(y-0.5) + a₅·(y-0.5)²
-
-This is a linear system in a₁...a₅, so the fit is exact (no optimization needed).
-The CDF is obtained by numerical root-finding of Q(y) = x.
-"""
-function fit_metalog(qv, qp)
-    # Build the 5×5 basis matrix
-    n = length(qv)
-    M = zeros(n, n)
-    for i in 1:n
-        y = qp[i]
-        logit = log(y / (1 - y))
-        M[i, 1] = 1.0
-        M[i, 2] = logit
-        M[i, 3] = (y - 0.5) * logit
-        M[i, 4] = y - 0.5
-        M[i, 5] = (y - 0.5)^2
-    end
-    a = M \ qv
-    return a
-end
-
-function metalog_quantile(a, y)
-    logit = log(y / (1 - y))
-    ymh = y - 0.5
-    a[1] + a[2] * logit + a[3] * ymh * logit + a[4] * ymh + a[5] * ymh^2
-end
-
-function metalog_cdf_at(a, x; tol=1e-10, maxiter=100)
-    # Find y such that Q(y) = x via bisection
-    lo, hi = 1e-8, 1.0 - 1e-8
-    # Check bounds
-    metalog_quantile(a, lo) >= x && return 0.0
-    metalog_quantile(a, hi) <= x && return 1.0
-
-    for _ in 1:maxiter
-        mid = (lo + hi) / 2
-        (hi - lo) < tol && return mid
-        metalog_quantile(a, mid) < x ? (lo = mid) : (hi = mid)
-    end
-    return (lo + hi) / 2
-end
-
-function make_metalog_e_cdf(bin; resolution_scale=1.0)
-    e = [bin.EnergyQuantile2_3Percent, bin.EnergyQuantile15_9Percent, bin.EnergyQuantile50_0Percent, bin.EnergyQuantile84_1Percent, bin.EnergyQuantile97_7Percent]
-
-    median_e = e[3]
-    e = median_e .+ resolution_scale .* (e .- median_e)
-
-    # Fit in log-space for better behavior on highly skewed energy distributions
-    log_e = log10.(max.(e, 1e-30))
-    qp = [0.023, 0.159, 0.5, 0.841, 0.977]
-    a = fit_metalog(log_e, qp)
-
-    # CDF takes linear energy, converts to log10, then evaluates metalog CDF
-    f = x -> begin
-        x <= 0 && return 0.0
-        metalog_cdf_at(a, log10(x))
-    end
-    return f
-end
-
 
 
 # Double-sided Crystal Ball distribution in log(E)
@@ -256,277 +188,6 @@ function fit_dscb(qv, qp)
     return best
 end
 
-function make_dscb_e_cdf(bin; resolution_scale=1.0)
[…]
