+++ src/experiments/super_k/sk_atm_2023/generate_e_response.jl
@@ -1,182 +1,0 @@
-#!/usr/bin/env julia
-# Fit energy response distribution parameters for all bins and channels
-# using unoscillated MC. Stores fit parameters for each method so the
-# R matrix can be reconstructed without repeating the expensive fitting.
-#
-# Methods fitted:
-#   - DSCB (logE): Double-sided Crystal Ball in log10(E) space (6 params)
-#   - Novosibirsk (logE): Novosibirsk function in log10(E) space (3 params)
-#   - Novosibirsk (linE): Novosibirsk function in linear E space (3 params)
-
-using Pkg
-Pkg.activate(joinpath(@__DIR__, "../../../.."))
-
-using Newtrinos
-using JLD2
-using Printf
-
-const DATADIR = @__DIR__
-const QP = [0.023, 0.159, 0.5, 0.841, 0.977]
-
-# Load unoscillated MC
-println("Loading unoscillated MC...")
-mc = (
-    nue     = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNueNoOsc.txt")),
-    numu    = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNumuNoOsc.txt")),
-    nutau   = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNutauNoOsc.txt")),
-    nuebar  = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNueBarNoOsc.txt")),
-    numubar = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNumuBarNoOsc.txt")),
-    nunc    = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNCNoOsc.txt")),
-)
-
-n_bins = 930
-flavors = keys(mc)
-
-function get_energy_quantiles(bin)
-    [bin.EnergyQuantile2_3Percent, bin.EnergyQuantile15_9Percent,
-     bin.EnergyQuantile50_0Percent, bin.EnergyQuantile84_1Percent,
-     bin.EnergyQuantile97_7Percent]
-end
-
-# ─── DSCB fits ───
-println("\n--- Fitting DSCB (logE) ---")
-dscb_result = Dict{Symbol, Dict{String, Vector{Float64}}}()
-
-for flav in flavors
-    mc_comp = mc[flav]
-    mu_vec = zeros(n_bins)
-    sigma_vec = zeros(n_bins)
-    alphaL_vec = zeros(n_bins)
-    nL_vec = zeros(n_bins)
-    alphaR_vec = zeros(n_bins)
-    nR_vec = zeros(n_bins)
-
-    t0 = time()
-    for bin_idx in 1:n_bins
-        bin = mc_comp[bin_idx, :]
-        bin.Counts == 0 && continue
-        e_q = get_energy_quantiles(bin)
-        mu, sigma, aL, nL, aR, nR = Newtrinos.super_k.fit_dscb(e_q, QP)
-        mu_vec[bin_idx] = mu
-        sigma_vec[bin_idx] = sigma
-        alphaL_vec[bin_idx] = aL
-        nL_vec[bin_idx] = nL
-        alphaR_vec[bin_idx] = aR
-        nR_vec[bin_idx] = nR
-    end
-    dt = time() - t0
-
-    dscb_result[flav] = Dict(
[…]

[…]
