### 2026-04-09 82fe2d7
Fix per-bin R matrix blur: simplify energy, use kappa for cosZ
Energy blur: replace arbitrary min(10, n_mc)/max(n_mc, 1) clamps with
clean formula sigma = reco_bin_width / sqrt(N_MC).

CosZ blur: instead of applying a Gaussian kernel to the vMF PDF,
reduce kappa directly: kappa_eff = kappa / (1 + kappa * w² / N_MC).
This preserves the vMF functional form while broadening the
distribution to account for MC statistical uncertainty.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
