### 2026-04-09 3e3ceae
Remove remaining dead code from super_k.jl
- Remove dscb_pdf_unnorm (never called, only dscb_cdf_unnorm is used)
- Remove unused imports: MonotonicSplines, DataStructures, LaTeXStrings
- Remove R_3d and R_coverage from get_assets return (only R is used
  in the forward model; R_3d was only referenced by external plots)

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
