### 2026-04-09 ea505ea
Fix xsec extrapolation (Flat not Linear) and add dσdE for H2O_PCA
The Differential_H2O interpolations used Linear() extrapolation beyond
the CSV data range, causing σ/E to grow unboundedly at high energy
(banana artifact). Fixed to Flat() to match the H2O_PCA model.

Also implement get_dσdE(cfg::H2O_PCA) returning absolute σ/E values
with per-channel norm, shape PCA, and nubar_ratio systematics applied.
Uses the same Wester/GENIE blended nominal curves and shape PCA as
the existing scale() function.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
