+++ src/physics/xsec.jl
@@ -133,7 +133,7 @@ function _load_h2o_interpolations()
         y = collect(skipmissing(df[:,idx+1]))
         itp = interpolate((x,), y, Gridded(Linear()))
         m(x) = max.(0, x)
-        return m ∘ extrapolate(itp, Linear())
+        return m ∘ extrapolate(itp, Flat())
     end
 
     nue = (
@@ -471,6 +471,157 @@ function get_scale(cfg::H2O_PCA)
             ch_shape = process_shape_itps[ch]
             ch_w = max.(zero(T), f_ch .* ch_norm .* (one(T) .+ ch_eps .* ch_shape.(E)))
             # Normalization-conserving nu/nubar ratio: scale both sides symmetrically
+            r = getfield(params, getfield(nubar_ratio_syms, Symbol(ch)))
+            if anti
+                ch_w = ch_w .* (2 * r / (1 + r))
+            else
+                ch_w = ch_w .* (2 / (1 + r))
+            end
+            result .+= ch_w
+        end
+
+        if flav == :nutau
+            return result .* params.xsec_nutau_cc_norm
+        end
+        return result
+    end
+end
+
+function get_dσdE(cfg::H2O_PCA)
+    data = load(joinpath(@__DIR__, "xsec_genie_data.jld2"))
+    E_grid = data["E_grid"]
+    wester_xsec = data["wester_xsec"]
+    all_xsec = data["all_xsec"]
+
+    nominal_key = string(cfg.nominal)
+    cc_channels = ("CC1p1h", "CC2p2h", "CC1pi", "CCDIS", "CCother")
+
+    function make_itp(vals)
+        itp = interpolate((E_grid,), vals, Gridded(Linear()))
+        return extrapolate(itp, Flat())
+    end
+
+    # Blend Wester → GENIE above 20-30 GeV (same as get_scale)
+    E_blend_lo = 20.0
+    E_blend_hi = 30.0
+    blend_weight = [(E_grid[i] < E_blend_lo ? 1.0 :
+                      E_grid[i] > E_blend_hi ? 0.0 :
+                      1.0 - (E_grid[i] - E_blend_lo) / (E_blend_hi - E_blend_lo))
+                     for i in 1:length(E_grid)]
+
+    function get_nominal(flav, ch)
+        if nominal_key == "Wester"
+            w = wester_xsec[flav][ch]
+            g = all_xsec["G18_10a"][flav][ch]
+            return blend_weight .* w .+ (1.0 .- blend_weight) .* g
+        else
+            return all_xsec[nominal_key][flav][ch]
+        end
+    end
+
+    # Precompute absolute σ/E interpolations per channel per flavor
+    flav_keys = ("nue", "nuebar", "numu", "numubar")
+    nom_xsec_itps = Dict{String, NamedTuple}()
+    for fk in flav_keys
+        xsecs = NamedTuple{Symbol.(cc_channels)}(make_itp(get_nominal(fk, ch)) for ch in cc_channels)
+        nom_xsec_itps[fk] = xsecs
+    end
+
+    # NC σ/E per flavor
+    nc_itps = Dict{String, Any}()
+    for fk in flav_keys
+        nc_itps[fk] = make_itp(get_nominal(fk, "NC"))
+    end
+
+    # Reuse shape PCA and sym mappings from get_scale (computed at configure time)
+    # We need to recompute here since get_scale's closures aren't accessible
+    E_valid_max = 30.0
+    E_fade_start = 20.0
[…]
