### 2026-04-09 cdab5b0
Fix H2O_PCA Wester extrapolation: clamp above valid data range
The wester_xsec values in xsec_genie_data.jld2 were computed with
Linear() extrapolation beyond the ~26 GeV CSV data range, causing
DIS σ/E to rise unboundedly (banana artifact). Now clamp Wester
values above E=26 GeV to the last valid point before blending
with GENIE G18_10a. Applied consistently in get_nominal,
get_channel_curve (shape PCA), and NC shape PCA for both get_scale
and get_dσdE.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
