### 2026-04-09 9c69b91
Fix H2O_PCA Wester extension: use scaled GENIE shape above 26 GeV
Instead of clamping Wester values flat above the CSV data range and
blending to GENIE (which created a bump at the join), extend Wester
by scaling GENIE values to match exactly at the boundary (E=26 GeV).
This gives a smooth, continuous curve with no discontinuity.

Also trim the last 2 GENIE E_grid points (>950 GeV) which had
boundary artifacts (3% DIS drop in final grid step).

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
