### 2026-04-09 1e3dfd5
Switch Super-K to H2O_PCA xsec model, add nutau TODO
Switch Super-K default xsec from Differential_H2O to H2O_PCA for
per-channel shape PCA and nubar_ratio systematics.

Add TODO noting that nutau CC cross-sections currently use numu curves
as proxy — they should have a kinematic threshold at ~3.5 GeV (tau
mass) and different channel fractions at low energy.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
