### 2026-04-10 35176a6
Replace DSCB energy response with K=2 LogNormal mixture + E-dependent cosZ
Energy response:
- Replace DSCB (6 params, quantile-only) with K=2 LogNormal mixture in log10(E)
  (5 params, fits all 7 statistics: 5 quantiles + mean + std)
- Adaptive sigma floor prevents delta-like components
- Fallback to single LogNormal when K=2 produces spurious tails
- Mean/std errors drop from 2-88% (DSCB) to <0.1% (K=2)

CosZ response:
- Add energy-dependent vMF concentration: κ(E) = κ₀ × (E/E_ref)^β
- Per-bin β fitted via binary search with precomputed CDF lookup (~5s/flavor)
- Captures E-cosZ correlation fr
[…]
