+++ src/experiments/super_k/sk_atm_2023/generate_response_params.jl
@@ -1,5 +1,5 @@
 #!/usr/bin/env julia
-# Fit DSCB energy + vMF cosZ response parameters for all bins and channels
+# Fit LogNormal mixture (K=2) energy + vMF cosZ response parameters for all bins
 # using unoscillated MC. Saves precomputed parameters to JLD2 files so that
 # super_k.jl can build the response matrix without repeating these fits.
 #
@@ -36,38 +36,46 @@ function get_energy_quantiles(bin)
      bin.EnergyQuantile97_7Percent]
 end
 
-# ─── DSCB energy fits ───
-println("\n--- Fitting DSCB (logE) ---")
-dscb_result = Dict{Symbol, Dict{String, Vector{Float64}}}()
+# ─── LogNormal mixture (K=2) energy fits ───
+println("\n--- Fitting LogNormal mixture K=2 (logE) ---")
+mix_result = Dict{Symbol, Dict{String, Vector{Float64}}}()
 
 for flav in flavors
     mc_comp = mc[flav]
-    mu_vec = zeros(n_bins); sigma_vec = zeros(n_bins)
-    alphaL_vec = zeros(n_bins); nL_vec = zeros(n_bins)
-    alphaR_vec = zeros(n_bins); nR_vec = zeros(n_bins)
+    mu1_vec = zeros(n_bins); mu2_vec = zeros(n_bins)
+    sigma1_vec = zeros(n_bins); sigma2_vec = zeros(n_bins)
+    w1_vec = zeros(n_bins); w2_vec = zeros(n_bins)
 
     t0 = time()
     for bin_idx in 1:n_bins
         bin = mc_comp[bin_idx, :]
         bin.Counts == 0 && continue
-        mu, sigma, aL, nL, aR, nR = Newtrinos.super_k.fit_dscb(get_energy_quantiles(bin), QP)
-        mu_vec[bin_idx] = mu; sigma_vec[bin_idx] = sigma
-        alphaL_vec[bin_idx] = aL; nL_vec[bin_idx] = nL
-        alphaR_vec[bin_idx] = aR; nR_vec[bin_idx] = nR
+        eq = get_energy_quantiles(bin)
+        any(eq .<= 0) && continue
+        bin.EnergyAvg <= 0 && continue
+        bin.EnergyRMS <= 0 && continue
+
+        μs, σs, ws = Newtrinos.super_k.fit_energy_response(eq, bin.EnergyAvg, bin.EnergyRMS, QP)
+        mu1_vec[bin_idx] = μs[1]; mu2_vec[bin_idx] = μs[2]
+        sigma1_vec[bin_idx] = σs[1]; sigma2_vec[bin_idx] = σs[2]
+        w1_vec[bin_idx] = ws[1]; w2_vec[bin_idx] = ws[2]
     end
     dt = time() - t0
 
-    dscb_result[flav] = Dict(
-        "mu" => mu_vec, "sigma" => sigma_vec,
-        "alphaL" => alphaL_vec, "nL" => nL_vec,
-        "alphaR" => alphaR_vec, "nR" => nR_vec
+    mix_result[flav] = Dict(
+        "mu1" => mu1_vec, "mu2" => mu2_vec,
+        "sigma1" => sigma1_vec, "sigma2" => sigma2_vec,
+        "w1" => w1_vec, "w2" => w2_vec
     )
     println(@sprintf "  %8s: %.1f s" flav dt)
 end
 
-# ─── vMF cosZ fits ───
-println("\n--- Fitting vMF cosZ ---")
-vmf_result = Dict{Symbol, Dict{String, Vector{Float64}}}()
+# ─── vMF cosZ fits (two stages) ───
[…]
