+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -12,6 +12,7 @@ using Statistics: mean
 using Printf
 using SpecialFunctions: besseli
 using JLD2
+using Optim
 using ..Newtrinos
 
 @kwdef struct SuperKAtm <: Newtrinos.Experiment
@@ -61,117 +62,175 @@ function read_sk_file(filepath::String)
 end
 
 
-# Double-sided Crystal Ball distribution in log(E)
-# PDF: Gaussian core with power-law tails on both sides
-# Parameters: mu, sigma, alphaL, nL, alphaR, nR
-# Transition: left tail at t < -alphaL, right tail at t > alphaR (t = (x-mu)/sigma)
+# 2-component LogNormal mixture for energy response.
+# Equivalent to a 2-Gaussian mixture in log₁₀(E) space.
+# Fits all 7 data release statistics: 5 quantiles + mean + std (in linear E).
+# Two overlapping components produce smooth, nearly-unimodal distributions with
+# enough flexibility to capture asymmetry, while analytic moments keep fitting fast.
+# Falls back to single LogNormal when K=2 produces spurious tails.
 
-function dscb_cdf_unnorm(t, alphaL, nL, alphaR, nR)
-    # Integral of unnormalized PDF from -inf to t
-    if t < -alphaL
-        A = (nL / alphaL)^nL * exp(-alphaL^2 / 2)
-        B = nL / alphaL - alphaL
-        # Integral of A*(B-t')^(-nL) from -inf to t = A/(nL-1) * (B-t)^(1-nL)
-        return A / (nL - 1) * (B - t)^(1 - nL)
-    else
-        # Left tail integral from -inf to -alphaL
-        A_L = (nL / alphaL)^nL * exp(-alphaL^2 / 2)
-        B_L = nL / alphaL - alphaL
-        left_tail = A_L / (nL - 1) * (B_L + alphaL)^(1 - nL)
+const _LN10 = log(10.0)
 
-        if t <= alphaR
-            # Gaussian core integral from -alphaL to t
-            # = sqrt(2pi) * (Phi(t) - Phi(-alphaL))
-            gauss_part = sqrt(2 * pi) * (cdf(Normal(), t) - cdf(Normal(), -alphaL))
-            return left_tail + gauss_part
+function lognormal_mix_cdf(E, μs, σs, ws)
+    x = log10(E)
+    ws[1] * cdf(Normal(μs[1], σs[1]), x) + ws[2] * cdf(Normal(μs[2], σs[2]), x)
+end
+
+function lognormal_mix_mean(μs, σs, ws)
+    ws[1] * exp(μs[1] * _LN10 + 0.5 * σs[1]^2 * _LN10^2) +
+    ws[2] * exp(μs[2] * _LN10 + 0.5 * σs[2]^2 * _LN10^2)
+end
+
+function lognormal_mix_std(μs, σs, ws)
+    m1 = lognormal_mix_mean(μs, σs, ws)
+    m2 = ws[1] * exp(2 * μs[1] * _LN10 + 2 * σs[1]^2 * _LN10^2) +
+         ws[2] * exp(2 * μs[2] * _LN10 + 2 * σs[2]^2 * _LN10^2)
+    sqrt(max(m2 - m1^2, 0.0))
+end
+
+function _mix_unpack(x, σ_min)
+    μs = (x[1], x[2])
+    σs = (σ_min + exp(x[3]), σ_min + exp(x[4]))
+    w1 = 1.0 / (1.0 + exp(-x[5]))
+    ws = (w1, 1.0 - w1)
+    μs, σs, ws
+end
+
+function _mix_loss(x, eq, qp, mean_obs, std_obs, σ_min)
[…]
