### 2026-04-10 d5fbe2a
Add standalone script for E-dependent cosZ beta fitting
Reads existing energy + vMF params and fits β per bin without
redoing the energy or constant-vMF fits. Takes ~30s total.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
