### 2026-04-10 ab2e1a0
Constrain E-dep cosZ beta to [0.3, 1.5] based on physics
Angular resolution always improves with energy, so beta=0 (factorized)
is unphysical. Paper values give beta = 2*b_A where b_A ranges from
0.14 to 0.59 across samples, so beta in [0.3, 1.5] covers the
physical range. Removes the beta=0 fallback check.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
