### 2026-04-11 e98ec06
Use per-class beta for E-dependent cosZ, optimized from NoOsc roundtrip
Per-class beta values optimized to minimize NoOsc→NO roundtrip chi2:
- upmu_thru_nonshowering: β=2.21
- upmu_thru_showering:    β=3.19
- upmu_stop:              β=1.02
- pc_thru:                β=1.15
- pc_stop:                β=0.35
- default (all others):   β=0.50

κ₀ refitted per bin at the fixed class beta to maintain cosZ quantile
matching. generate_edep_params.jl updated to use per-class beta with
per-bin kappa scan (~2min total).

NoOsc→NO roundtrip (excl. nutau):
  chi2: 421.7 (DSCB) → 204 (per-bin β[0.3,1.5]) → 72
[…]
