### 2026-04-11 1ee7c40
Revert to per-bin beta in [0, 2] for E-dependent cosZ
Per-class beta with high values (2-3) improved the NoOsc→NO roundtrip
but worsened the actual oscillation fit. Revert to per-bin beta fitted
in [0, 2] which is more conservative and lets each bin find its own
optimum based on the cosZ quantile matching.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
