+++ src/experiments/super_k/sk_atm_2023/generate_edep_params.jl
@@ -1,7 +1,6 @@
 #!/usr/bin/env julia
-# Fit E-dependent κ using per-class β values
-# β is fixed per sample class (from NoOsc→NO roundtrip optimization)
-# κ₀ is fitted per bin via 1D scan to match cosZ quantiles
+# Fit E-dependent β per bin using existing energy + vMF params
+# β ∈ [0, 2], κ₀ scanned per bin at best β
 
 using Pkg
 Pkg.activate(joinpath(@__DIR__, "../../../.."))
@@ -28,25 +27,11 @@ mc = (
     nunc    = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins/unoscillated/sk_2023_MCNCNoOsc.txt")),
 )
 
-using CSV, DataFrames
-bininfo = CSV.read(joinpath(DATADIR, "bins/sk_2023_BinInfo.txt"), DataFrame; delim=' ', ignorerepeated=true, comment="#", header=false)
-rename!(bininfo, [:Sample, :logPMin, :logPMax, :CosZMin, :CosZMax])
-
 n_bins = 930
 flavors = keys(mc)
 loge_grid = collect(LinRange(-1, 3, 201))
 
-# Per-class β values (from NoOsc→NO roundtrip optimization)
-function get_class_beta(sample)
-    occursin("upmu_thru_nonshowering", sample) && return 2.21
-    occursin("upmu_thru_showering", sample) && return 3.19
-    occursin("upmu_stop", sample) && return 1.02
-    occursin("pc_thru", sample) && return 1.15
-    occursin("pc_stop", sample) && return 0.35
-    return 0.5  # default for all other samples
-end
-
-println("\n--- Fitting E-dependent κ with per-class β ---")
+println("\n--- Fitting per-bin E-dependent β ∈ [0, 2] ---")
 vmf_result = Dict{Symbol, Dict{String, Vector{Float64}}}()
 
 for flav in flavors
@@ -72,73 +57,11 @@ for flav in flavors
                 bin.CosZQuantile50_0Percent, bin.CosZQuantile84_1Percent,
                 bin.CosZQuantile97_7Percent]
 
-        beta = get_class_beta(bininfo.Sample[bin_idx])
-
-        # Use fit_vmf_edep but with fixed beta — just scan κ₀
-        # We override the beta search by passing it directly
-        kappa0, _, E_ref = Newtrinos.super_k.fit_vmf_edep(
+        kappa0, beta, E_ref = Newtrinos.super_k.fit_vmf_edep(
             cz_q, QP, mu_z_vec[bin_idx], kappa_vec[bin_idx],
             loge_grid, mix_result[flav], bin_idx)
 
-        # Now rescan κ₀ at the fixed class beta
-        μs = (mix_result[flav]["mu1"][bin_idx], mix_result[flav]["mu2"][bin_idx])
-        σs = (mix_result[flav]["sigma1"][bin_idx], mix_result[flav]["sigma2"][bin_idx])
-        ws = (mix_result[flav]["w1"][bin_idx], mix_result[flav]["w2"][bin_idx])
-        E_grid = 10.0 .^ loge_grid
-        c_e = [Newtrinos.super_k.lognormal_mix_cdf(e, μs, σs, ws) for e in E_grid]
-        p_E = max.(diff(c_e), 0.0)
-        s = sum(p_E); s > 0 && (p_E ./= s)
-        cdf_e = cumsum(p_E)
-        med_idx = clamp(searchsortedfirst(cdf_e, 0.5), 1, length(p_E))
-        E_ref = 10.0^((loge_grid[med_idx] + loge_grid[med_idx+1]) / 2)
-
-        active = findall(p_E .> 1e-4 * maximum(p_E))
[…]
