### 2026-04-11 6863f47
EXPERIMENTAL: Add E^0.6 energy PDF deconvolution correction
Multiply the fitted energy PDF by E^0.6 before building the response
matrix. This empirically corrects a suspected generation spectrum
mismatch — the MC may have been generated with a power law different
from the Honda flux, biasing the per-bin energy distributions.

Effect on NoOsc→NO roundtrip:
  - Total chi2: 164 → 52 (-68%)
  - numu: 125 → 56 (-55%), upmu_thru: 128 → 19 (-85%)
  - nutau: 114 → 16 (-86%), numubar: 89 → 22 (-75%)
  - upmu_stop regresses: 12 → 26 (+112%)

WARNING: This correction is HIGHLY EXPERIMENTAL. The E^0.6 exp
[…]
