### 2026-04-11 a5f6ee8
Revert E^0.6 energy PDF correction — worsened oscillation fit
The E^0.6 correction improved the NoOsc→NO roundtrip chi2 (164→52)
but made the actual oscillation fit slightly worse across all parameters.
The NoOsc roundtrip is not a reliable proxy for oscillation fit quality.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
