### 2026-04-11 f78a3d3
Switch energy PDF from K=2 to K=3 LogNormal mixture
K=3 has 8 free parameters for 7 constraints, matching all summary
statistics (5 quantiles + mean + std) much more precisely than K=2.
The adaptive sigma floor prevents delta-like components.

K=3 was previously tested only via the NoOsc roundtrip (which showed
it worse), but the roundtrip proved unreliable as a proxy for the
oscillation fit. Testing K=3 in the actual oscillation fit to see
if the better statistic matching translates to better physics.

May need to revert to K=2 if oscillation fit worsens.

Co-Authored-By: Claude Opus 4.6 (1M 
[…]
