+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -62,12 +62,10 @@ function read_sk_file(filepath::String)
 end
 
 
-# 2-component LogNormal mixture for energy response.
-# Equivalent to a 2-Gaussian mixture in log₁₀(E) space.
+# 3-component LogNormal mixture for energy response.
+# Gaussian mixture in log₁₀(E) space with K=3 components (8 free parameters).
 # Fits all 7 data release statistics: 5 quantiles + mean + std (in linear E).
-# Two overlapping components produce smooth, nearly-unimodal distributions with
-# enough flexibility to capture asymmetry, while analytic moments keep fitting fast.
-# Falls back to single LogNormal when K=2 produces spurious tails.
+# Adaptive σ floor prevents delta-like components.
 
 const _LN10 = log(10.0)
 
@@ -89,10 +87,13 @@ function lognormal_mix_std(μs, σs, ws)
 end
 
 function _mix_unpack(x, σ_min)
-    μs = (x[1], x[2])
-    σs = (σ_min + exp(x[3]), σ_min + exp(x[4]))
-    w1 = 1.0 / (1.0 + exp(-x[5]))
-    ws = (w1, 1.0 - w1)
+    μs = (x[1], x[2], x[3])
+    σs = (σ_min + exp(x[4]), σ_min + exp(x[5]), σ_min + exp(x[6]))
+    # softmax for 3 weights
+    max_w = max(x[7], x[8], 0.0)
+    e1 = exp(x[7] - max_w); e2 = exp(x[8] - max_w); e3 = exp(-max_w)
+    s = e1 + e2 + e3
+    ws = (e1/s, e2/s, e3/s)
     μs, σs, ws
 end
 
@@ -113,35 +114,46 @@ function _mix_loss(x, eq, qp, mean_obs, std_obs, σ_min)
     loss
 end
 
-function _fit_lognormal_mixture(eq, mean_obs, std_obs, qp, σ_min; n_restarts=20)
-    # Fit 2-component Gaussian mixture in log₁₀(E) to 5 quantiles + mean + std
+function fit_energy_response(eq, mean_obs, std_obs, qp; n_restarts=20)
+    # Fit 3-component Gaussian mixture in log₁₀(E) to 5 quantiles + mean + std
     log_q = log10.(max.(eq, 1e-30))
     log_median = log_q[3]
     log_iqr = max((log_q[4] - log_q[2]) / 2.0, 0.03)
     log_range = max(log_q[5] - log_q[1], 0.05)
     log_mean = log10(max(mean_obs, 1e-30))
+
+    σ_min = max(0.02, 0.05 * log_range)
     to_raw_σ(σ) = log(max(σ - σ_min, 1e-4))
 
     best_x = nothing
     best_loss = Inf
 
     inits = [
-        ([log_median, log_mean], [log_iqr, log_iqr*0.5], [1.0]),
-        ([log_median-0.3*log_iqr, log_median+0.3*log_iqr], [log_iqr*0.8, log_iqr*0.8], [0.0]),
-        ([log_q[2], log_q[4]], [log_iqr*0.5, log_iqr*0.5], [0.5]),
-        ([log_median, log_q[1]+0.3*log_range], [log_iqr*1.2, log_iqr*0.3], [2.0]),
-        ([log_median, log_q[5]-0.3*log_range], [log_iqr*1.0, log_iqr*0.5], [1.5]),
-        ([log_median-0.05*log_iqr, log_median+0.05*log_iqr], [log_iqr, log_iqr], [0.0]),
-        ([log_q[1]+0.2*log_range, log_q[5]-0.2*log_range], [log_iqr*0.6, log_iqr*0.6], [0.5]),
[…]
