### 2026-04-12 13833af
CompatHelper: bump compat for CUDA to 6, (keep existing compat)
