+++ Project.toml
@@ -83,7 +83,7 @@ BenchmarkTools = "1.6.3"
 Bessels = "0.2.8"
 Bijections = "0.1, 0.2"
 CSV = "0.10.15"
-CUDA = "5.7.3"
+CUDA = "5.7.3, 6"
 CairoMakie = "0.13, 0.14, 0.15"
 ChainRules = "1.72.6"
 ChainRulesCore = "1.26.0"
