### 2026-04-12 90a886f
Revert to K=2 energy PDFs, fit to NO MC instead of unoscillated
K=3 made the oscillation fit worse — reverted to K=2.

Switch from unoscillated to NO (normal ordering) MC for fitting the
energy and cosZ response parameters. The data release summary statistics
are oscillation-weighted (quantiles, mean, RMS all differ between
NoOsc/NO/IO), so fitting to the NO MC calibrates the response matrix
at the point closest to the true oscillation parameters.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
