+++ src/experiments/super_k/sk_atm_2023/generate_response_params.jl
@@ -16,15 +16,16 @@ using Printf
 const DATADIR = @__DIR__
 const QP = [0.023, 0.159, 0.5, 0.841, 0.977]
 
-# Load unoscillated MC
-println("Loading unoscillated MC...")
+# Load NO (normal ordering) MC — summary statistics are oscillation-weighted,
+# so fitting to NO gives a response calibrated near the true oscillation point.
+println("Loading NO MC...")
 mc = (
-    nue     = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNueNoOsc.txt")),
-    numu    = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNumuNoOsc.txt")),
-    nutau   = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNutauNoOsc.txt")),
-    nuebar  = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNueBarNoOsc.txt")),
-    numubar = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNumuBarNoOsc.txt")),
-    nunc    = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "unoscillated", "sk_2023_MCNCNoOsc.txt")),
+    nue     = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "normal", "sk_2023_MCNueNO.txt")),
+    numu    = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "normal", "sk_2023_MCNumuNO.txt")),
+    nutau   = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "normal", "sk_2023_MCNutauNO.txt")),
+    nuebar  = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "normal", "sk_2023_MCNueBarNO.txt")),
+    numubar = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "normal", "sk_2023_MCNumuBarNO.txt")),
+    nunc    = Newtrinos.super_k.read_sk_file(joinpath(DATADIR, "bins", "normal", "sk_2023_MCNCNO.txt")),
 )
 
 n_bins = 930
@@ -36,15 +37,15 @@ function get_energy_quantiles(bin)
      bin.EnergyQuantile97_7Percent]
 end
 
-# ─── LogNormal mixture (K=3) energy fits ───
-println("\n--- Fitting LogNormal mixture K=3 (logE) ---")
+# ─── LogNormal mixture (K=2) energy fits ───
+println("\n--- Fitting LogNormal mixture K=2 (logE) ---")
 mix_result = Dict{Symbol, Dict{String, Vector{Float64}}}()
 
 for flav in flavors
     mc_comp = mc[flav]
-    mu1_vec = zeros(n_bins); mu2_vec = zeros(n_bins); mu3_vec = zeros(n_bins)
-    sigma1_vec = zeros(n_bins); sigma2_vec = zeros(n_bins); sigma3_vec = zeros(n_bins)
-    w1_vec = zeros(n_bins); w2_vec = zeros(n_bins); w3_vec = zeros(n_bins)
+    mu1_vec = zeros(n_bins); mu2_vec = zeros(n_bins)
[…]
