### 2026-04-12 716d129
Revert to unoscillated MC for response fitting
NO MC made the oscillation fit slightly worse. Revert to unoscillated
MC which remains the best-performing version.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
