### 2026-04-12 c279dfe
Add neighbor regularization for energy and cosZ parameters
Smooth fitted parameters between neighboring bins within the same
sample. Each bin's params are blended with the average of its
neighbors (30% neighbor weight). This prevents jagged parameter
profiles where the per-bin optimizer finds very different local
optima for physically similar bins.

Smoothed parameters: mu1, mu2, sigma1, sigma2 (energy), kappa, beta (cosZ).
Weights w1/w2 and mu_z are left unsmoothed.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
