### 2026-04-12 8896684
Add sk_total_norm parameter to absorb overall MC normalization
The data release NO MC underpredicts data by ~7% overall. SK's full
analysis absorbs this with 193 systematic parameters; our ~60 params
can't fully cover it, forcing θ₂₃ to compensate.

sk_total_norm (prior: Normal(1.0, 0.2)) freely scales all MC channels,
decoupling the absolute rate from the oscillation parameters.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
