### 2026-04-13 e2b7478
Per-channel migration factors + fixed regularization
Two changes that significantly improve sin²θ₂₃ (0.485 → 0.459):

1. Per-channel systematic factors: Split get_all_factors into
   get_norm_factors (from total, shared) and get_migration_factors
   (from each channel's own counts). PID, ring counting, BDT,
   decay-e, neutron tagging, FC/PC separation, and pi0 systematics
   now respect the actual flavor composition in each bin, rather
   than scaling all flavors identically.

2. Fixed regularization: Canonicalize K=2 mixture component ordering
   (μ₁ < μ₂) before smoothing across cosZ neighb
[…]
