### 2026-04-13 2c89e97
Override pathological response fits from partner channels or cross-flavor median
Failed fits (low stats or cross-flavor outliers >3x median) are detected and
replaced: width params from CP-conjugate partner (nue↔nuebar, numu↔numubar),
or cross-flavor median if partner also failed. Final cleanup clamps any
remaining outliers after re-smoothing. Fixes ~2800 bin×channel combinations.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
