+++ src/experiments/super_k/sk_atm_2023/generate_response_params.jl
@@ -218,6 +218,188 @@ for flav in flavors
     println(@sprintf "  %8s: smoothed" flav)
 end
 
+# ─── Cross-flavor override: detect failed fits, replace from partner/median ───
+# Minority channels (e.g., nutau in nue-like bins) with very few events can get
+# pathological fits. Rather than blending (which dilutes but doesn't fix), we
+# detect failures and fully replace width parameters from partner channels or
+# the cross-flavor median. Location parameters (mu1, mu2, mu_z) are kept from
+# the channel's own fit when possible, since nutau/NC legitimately peak at
+# different energies.
+const MIN_COUNTS = 1.0
+const MAX_RATIO = 3.0
+
+PARTNERS = Dict(
+    :nue => :nuebar, :nuebar => :nue,
+    :numu => :numubar, :numubar => :numu,
+    :nutau => :numu,   # closest CC kinematics
+    :nunc => nothing,  # use cross-flavor median
+)
+
+println("\n--- Detecting failed fits (min_counts=$MIN_COUNTS, max_ratio=$MAX_RATIO) ---")
+
+# 1. Detect failed bins
+failed = Dict(flav => falses(n_bins) for flav in flavors)
+
+for bin_idx in 1:n_bins
+    # Collect active sigmas and kappas across flavors for median computation
+    e_sigmas = Dict{Symbol,Float64}()
+    cz_kappas = Dict{Symbol,Float64}()
+    for flav in flavors
+        mc[flav][bin_idx, :].Counts == 0 && continue
+        s = mix_result[flav]["sigma1"][bin_idx]
+        s > 0 && (e_sigmas[flav] = s)
+        k = vmf_result[flav]["kappa"][bin_idx]
+        k > 0 && (cz_kappas[flav] = k)
+    end
+
+    e_med = length(e_sigmas) >= 2 ? median(collect(values(e_sigmas))) : 0.0
+    cz_med = length(cz_kappas) >= 2 ? median(collect(values(cz_kappas))) : 0.0
+
+    for flav in flavors
+        counts = mc[flav][bin_idx, :].Counts
+        counts == 0 && continue
+
+        # Low statistics — insufficient to constrain 5 quantiles + mean + RMS
+        if counts < MIN_COUNTS
+            failed[flav][bin_idx] = true; continue
+        end
+
+        # Energy width outlier
+        if e_med > 0 && haskey(e_sigmas, flav)
+            ratio = e_sigmas[flav] / e_med
+            if ratio > MAX_RATIO || ratio < 1/MAX_RATIO
+                failed[flav][bin_idx] = true; continue
+            end
+        end
+
+        # CosZ kappa outlier
+        if cz_med > 0 && haskey(cz_kappas, flav)
+            ratio = cz_kappas[flav] / cz_med
+            if ratio > MAX_RATIO || ratio < 1/MAX_RATIO
+                failed[flav][bin_idx] = true; continue
+            end
+        end
+    end
+end
+
+for flav in flavors
+    n_fail = count(failed[flav])
+    n_active = count(mc[flav][i, :].Counts > 0 for i in 1:n_bins)
+    println(@sprintf "  %8s: %d / %d active bins failed" flav n_fail n_active)
+end
+
+# 2. Override from partner or cross-flavor median
+println("\n--- Overriding failed fits from partner/median ---")
+
+function get_cross_flavor_median(result_dict, key, bin_idx, flavors, failed)
+    vals = Float64[]
+    for f in flavors
+        failed[f][bin_idx] && continue
+        v = result_dict[f][key][bin_idx]
+        v > 0 && push!(vals, v)
+    end
+    isempty(vals) ? 0.0 : median(vals)
+end
+
[…]
