### 2026-04-13 a410067
new legend 1000 IO
