### 2026-04-15 befa342
Add Gaussian discriminant model for migration systematics
Replace linear per-channel migration factors with a physically motivated
Gaussian CDF model. Each flavor's discriminant distribution is modeled as
a Gaussian with shared width but flavor-dependent mean inferred from its
nominal fraction. Threshold shifts affect contamination channels more than
dominant channels, which is physically correct and avoids the negative
factors the linear model produced for extreme fractions.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
