+++ src/experiments/super_k/sk_atm_2023/super_k.jl
@@ -944,6 +944,63 @@ function get_double_factor(total, mask1, mask2, factor1)
     return factor
 end
 
+"""
+    get_gaussian_double_factor(total, channel, mask1, mask2, param) → Vector
+
+Gaussian discriminant model for binary classifier migration systematics.
+
+Models each flavor channel's discriminant distribution as a Gaussian with shared
+width σ=1 and flavor-dependent mean μ_f, inferred from the nominal event fraction
+on each side of the classification threshold. A systematic parameter shift moves
+the threshold, changing fractions non-linearly via the Gaussian CDF.
+
+This is more physical than the linear `get_double_factor`: a flavor that dominates
+one side (e.g., nue in e-like bins, 95% on correct side) is barely affected by a
+small threshold shift, while a contamination flavor near the threshold (e.g., numu
+in e-like, 20%) is affected much more.
+
+**Derivation**:
+- For each channel, the nominal fraction frac1 = n1/(n1+n2) determines the Gaussian
+  mean: μ = -Φ⁻¹(frac1), where Φ is the standard normal CDF.
+- The threshold shift δ is derived from the total (all-flavor) counts so that the
+  total-count behavior matches the linear model: Φ(δ - μ_total) = param × frac1_total.
+- The same δ is applied to each channel's Gaussian: new_frac1 = Φ(δ - μ_channel).
+- Per-side factors: factor1 = new_frac1/frac1, factor2 = (1-new_frac1)/(1-frac1).
+
+At param=1.0, δ=0 and all factors are exactly 1.0. For small deviations, the model
+reproduces the linear behavior for total counts but gives physically motivated
+non-linear per-channel responses.
+"""
+function get_gaussian_double_factor(total, channel, mask1, mask2, param)
+    # Threshold shift δ from total counts (backward compatible with linear model)
+    total1 = sum(total[mask1])
+    total2 = sum(total[mask2])
+    total_sum = total1 + total2
+    (total_sum < 1e-10) && return ones(eltype(param), length(channel))
+
+    frac1_total = total1 / total_sum
+    μ_total = -quantile(Normal(), clamp(frac1_total, 1e-6, 1 - 1e-6))
+    new_frac1_total = clamp(param * frac1_total, 1e-6, 1 - 1e-6)
+    δ = quantile(Normal(), new_frac1_total) + μ_total
+
+    # Apply same δ to this channel's Gaussian
+    ch1 = sum(channel[mask1])
+    ch2 = sum(channel[mask2])
+    ch_sum = ch1 + ch2
+    (ch_sum < 1e-10) && return ones(eltype(param), length(channel))
+
+    frac1_ch = ch1 / ch_sum
+    # Channel entirely on one side — no migration possible
+    (frac1_ch < 1e-10 || frac1_ch > 1 - 1e-10) && return ones(eltype(param), length(channel))
+    μ_ch = -quantile(Normal(), clamp(frac1_ch, 1e-6, 1 - 1e-6))
+    new_frac1_ch = cdf(Normal(), δ - μ_ch)
+
+    f1 = frac1_ch > 1e-10 ? new_frac1_ch / frac1_ch : one(param)
+    f2 = (1 - frac1_ch) > 1e-10 ? (1 - new_frac1_ch) / (1 - frac1_ch) : one(param)
+
+    return (mask1 * f1 .+ .!mask1) .* (mask2 * f2 .+ .!mask2)
+end
+
 function get_norm_factors(params, assets, total)
[…]
