### 2026-04-15 0cf09bd
Remove K/pi ratio and add coszen-dependent Barr flux uncertainties
Replace the K/pi ratio systematic with coszen-dependent modulation
functions for the three flux ratio uncertainties (nue/nuebar,
numu/numubar, nue/numu). The new functions suppress the uncertainty
near horizontal directions and enhance it at up/down-going angles.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
