### 2026-04-15 b68e1d3
Fix coszen flux functions: use uncert as input, scale by eff_sigma outside
Avoids DomainError from fractional powers of negative numbers by passing
only the positive uncertainty envelope into fun_nu* and multiplying by
the (possibly negative) pull parameter eff_sigma afterwards. Also unify
sign convention to use + for all three flux ratio corrections.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
