### 2026-04-16 17e3ecd
some analysis updates, and a bit sider priors
