### 2026-04-16 b5e8836
Split CCQE norm into Sub-GeV/Multi-GeV and widen 2p2h prior
Split xsec_cc1p1h_norm into xsec_cc1p1h_subgev_norm (5%) and
xsec_cc1p1h_multigev_norm (25%) with sharp step at 1.33 GeV,
matching the Wester thesis treatment. Change cc2p2h_norm prior
from Truncated Normal to Uniform(0, 2) per thesis.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
