+++ src/experiments/super_k/sk_atm_2023/run_mle_nofluxnorm.jl
@@ -25,7 +25,8 @@ likelihood = Newtrinos.generate_likelihood(experiments)
 @reset params.xsec_cc1pi_shape = 1.459204145906447
 @reset params.sk_flux_norm_high = 0.0
 @reset params.sk_flux_norm_low = 0.0
-@reset params.xsec_cc1p1h_norm = 0.8798531573440069
+@reset params.xsec_cc1p1h_subgev_norm = 0.8798531573440069
+@reset params.xsec_cc1p1h_multigev_norm = 0.8798531573440069
 @reset params.xsec_nc_norm = 1.1196641134274956
 @reset params.sk_fc_multigev_rel_norm = 1.0118968918009859
 @reset params.sk_multigev_ring_counting = 0.9677503494043178
@@ -60,7 +61,8 @@ println("\n=== Result: No bathtub flux norm ===")
 @printf "sk_total_norm = %.4f\n" opt_params.sk_total_norm
 @printf "xsec_cc1p1h_shape = %.4f\n" opt_params.xsec_cc1p1h_shape
 @printf "xsec_cc1pi_shape = %.4f\n" opt_params.xsec_cc1pi_shape
-@printf "xsec_cc1p1h_norm = %.4f\n" opt_params.xsec_cc1p1h_norm
+@printf "xsec_cc1p1h_subgev_norm = %.4f\n" opt_params.xsec_cc1p1h_subgev_norm
+@printf "xsec_cc1p1h_multigev_norm = %.4f\n" opt_params.xsec_cc1p1h_multigev_norm
 @printf "xsec_nc_norm = %.4f\n" opt_params.xsec_nc_norm
 @printf "sk_fiducial_norm = %.4f\n" opt_params.sk_fiducial_norm
 @printf "sk_fc_norm = %.4f\n" opt_params.sk_fc_norm
