+++ src/experiments/super_k/sk_atm_2023/run_mle_fixedreg.jl
@@ -23,7 +23,8 @@ likelihood = Newtrinos.generate_likelihood(experiments)
 @reset params.xsec_cc1pi_shape = 1.459204145906447
 @reset params.sk_flux_norm_high = 0.3994660286467456
 @reset params.sk_flux_norm_low = -0.7448957257833595
-@reset params.xsec_cc1p1h_norm = 0.8798531573440069
+@reset params.xsec_cc1p1h_subgev_norm = 0.8798531573440069
+@reset params.xsec_cc1p1h_multigev_norm = 0.8798531573440069
 @reset params.xsec_nc_norm = 1.1196641134274956
 @reset params.sk_fc_multigev_rel_norm = 1.0118968918009859
 @reset params.sk_multigev_ring_counting = 0.9677503494043178
