### 2026-04-16 349ed80
Add nue/numu CCQE ratio, tighten response matrix override, adjust priors
xsec.jl (H2O_PCA):
- Add xsec_cc1p1h_nue_numu_ratio parameter (5% prior) for νe/νμ CCQE
  flavor ratio uncertainty from nuclear model spread (RFG vs LFG)

generate_response_params.jl:
- Lower MAX_RATIO from 3.0 to 1.8 for more aggressive outlier detection
- Add coszen location (mu_z) outlier check using channel's own vMF width
  added in quadrature with std of other channels
- Override locations (mu1, mu2, mu_z) from cross-flavor median for
  low-count bins, not just from partner

super_k.jl:
- Dial back SK IV-V updown en
[…]
