### 2026-04-16 5a00c3a
Add solar activity flux systematic and widen SK I-III sub-GeV PID prior
- Add sk_solar_activity parameter (Uniform(0,1)) that interpolates the
  nominal flux between solar min and solar max using precomputed
  solmax/solmin ratio per flavor. Applied before Barr systematics.
- Widen sk_i_iii_subgev_pid prior from 2% to 3% to reduce excess δCP
  sensitivity. SK I-III is the least calibrated phase; thesis quotes
  up to 2-3% PID uncertainty for some phases.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
