### 2026-04-17 2fb4cc9
Exclude lowest momentum bins (logP=[2.0,2.4]) from Super-K analysis
These 83 bins at 100-250 MeV show a ~1.5% systematic bias in the
NO→IO reweighting roundtrip for the nue channel, likely due to
detector threshold effects not captured by the response matrix.
They contribute ~9 of 20 Δχ² in δCP sensitivity, inflating the
constraint beyond the published result.

Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
