### 2026-04-19 bc1c924
CompatHelper: bump compat for Optimization to 5, (keep existing compat)
