+++ Project.toml
@@ -123,7 +123,7 @@ Memoize = "0.4.4"
 MonotonicSplines = "0.3.3"
 NPZ = "0.4.3"
 Optim = "1.11.0"
-Optimization = "4.8.0"
+Optimization = "4.8.0, 5"
 PCHIPInterpolation = "0.2.1"
 PDMats = "0.11"
 PairPlots = "3.0.1"
