+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/analysis/#Analysis-API","page":"Analysis","title":"Analysis API","text":"","category":"section"},{"location":"api/analysis/#Likelihood-Construction","page":"Analysis","title":"Likelihood Construction","text":"","category":"section"},{"location":"api/analysis/#Scanning-and-Profiling","page":"Analysis","title":"Scanning and Profiling","text":"","category":"section"},{"location":"api/analysis/#Data-Generation","page":"Analysis","title":"Data Generation","text":"","category":"section"},{"location":"api/analysis/#Utilities","page":"Analysis","title":"Utilities","text":"","category":"section"},{"location":"api/analysis/#Newtrinos.generate_likelihood","page":"Analysis","title":"Newtrinos.generate_likelihood","text":"generate_likelihood(experiments[, observed])\n\nConstruct a joint likelihood from a NamedTuple of configured experiments. Combines all forward models and observed data into a single likelihood object compatible with DensityInterface.logdensityof.\n\n\n\n\n\n","category":"function"},{"location":"api/analysis/#Newtrinos.get_params","page":"Analysis","title":"Newtrinos.get_params","text":"get_params(x)\n\nExtract nominal parameter values as a sorted NamedTuple. Works on Physics, Experiment, Wrapper, or a NamedTuple of modules (merging all with conflict checking).\n\n\n\n\n\n","category":"function"},{"location":"api/analysis/#Newtrinos.get_priors","page":"Analysis","title":"Newtrinos.get_priors","text":"get_priors(x)\n\nExtract prior distributions as a sorted NamedTuple. Works on Physics, Experiment, Wrapper, or a NamedTuple of modules (merging all with conflict checking).\n\n\n\n\n\n","category":"function"},{"location":"api/analysis/#Newtrinos.condition","page":"Analysis","title":"Newtrinos.condition","text":"condition(priors, conditional_vars, params)\n\nFix parameters to specific values by replacing their priors with constants. conditional_vars can be an Array of symbols (uses values from params) or a Dict mapping symbols to values.\n\n\n\n\n\n","category":"function"},{"location":"api/analysis/#Newtrinos.profile","page":"Analysis","title":"Newtrinos.profile","text":"profile(likelihood, priors, vars_to_scan, params; cache_dir=nothing, map_func=nothing)\n\nRun a profile likelihood scan. At each grid point defined by vars_to_scan, optimizes over all other parameters. Use cache_dir to cache and resume results. Use map_func=pmap for distributed parallelism (default: Threads.@threads).\n\n\n\n\n\n","category":"function"},{"location":"api/analysis/#Newtrinos.scan","page":"Analysis","title":"Newtrinos.scan","text":"scan(likelihood, priors, vars_to_scan, params; gradient_map=false)\n\nRun a simple likelihood scan on a grid (no optimization over nuisance parameters). Faster than profile but does not account for nuisance parameter variations. Set gradient_map=true to also compute gradients at each point.\n\n\n\n\n\n","category":"function"},{"location":"api/analysis/#Newtrinos.find_mle","page":"Analysis","title":"Newtrinos.find_mle","text":"find_mle(likelihood, prior, params; adsel=AutoPolyesterForwardDiff())\n\nFind the Maximum Likelihood Estimator using LBFGS optimization via BAT. Returns (llh, log_posterior, optimized_params). Parameters with ConstValueDist priors are held fixed.\n\n\n\n\n\n","category":"function"},{"location":"api/analysis/#Newtrinos.fi
[…]
