### 2026-04-22 875b187
Add H2O_PCA per-event/grid xsec reweighting for DeepCore and ORCA
All three atmospheric experiments (Super-K, DeepCore, ORCA) now share a
consistent H2O_PCA cross-section model nominally at NEUT 5.4.0, enabling
correlated xsec systematics in a global fit.

Key changes:

xsec.jl:
- Rename Wester → NEUT5_4_0 throughout (more descriptive)
- Add mc_nominal field to H2O_PCA (cross-section model used to generate MC)
- Add scale_event: per-event reweighting using GENIE interaction codes
  (0=QE→CC1p1h, 1=RES→CC1pi, 2=DIS, else→CCother)
- Add event_weights: fast precomputed path for scale_event — inte
[…]
