+++ src/experiments/icecube/deepcore_9y_verification_sample/test.jl
@@ -23,7 +23,7 @@ priors = Newtrinos.get_priors(experiments)
 @reset priors.Δm²₂₁ = p.Δm²₂₁
 @reset priors.θ₁₂ = p.θ₁₂
 @reset priors.δCP = p.δCP
-@reset priors.nutau_cc_norm = p.nutau_cc_norm
+@reset priors.xsec_nutau_cc_norm = p.xsec_nutau_cc_norm
 @reset priors.θ₁₃ = Truncated(Normal(0.156, 0.008), 0.13, 0.18)
 @reset priors.Δm²₃₁ = Uniform(0.0022, 0.0027)
 @reset priors.θ₂₃ = Uniform(0.685, 0.9)
