+++ src/experiments/km3net/orca6_433kton/orca.jl
@@ -24,7 +24,7 @@ function default_physics()
     osc = Newtrinos.osc.configure(Newtrinos.osc.OscillationConfig(interaction=Newtrinos.osc.SI()))
     atm_flux = Newtrinos.atm_flux.configure(Newtrinos.atm_flux.AtmFluxConfig(nominal_model=Newtrinos.atm_flux.HKKM("frj-ally-20-01-mtn-solmin.d")))
     earth_layers = Newtrinos.earth_layers.configure()
-    xsec = Newtrinos.xsec.configure()
+    xsec = Newtrinos.xsec.configure(Newtrinos.xsec.H2O_PCA(mc_nominal=:G00_00a))
     (; osc, atm_flux, earth_layers, xsec)
 end
 
@@ -80,7 +80,28 @@ function get_assets(physics; datadir = @__DIR__)
     data_hist = permutedims(reshape(data.W, reco_shape[rs]), rs)
     muon_hist = permutedims(reshape(muons.W, reco_shape[rs]), rs);
 
-    assets = (;mc, muon_hist, observed=cut(data_hist), binning, true_shape, reco_shape, nominal_layers, flux_nominal)
+    xsec_eval = if !isnothing(physics.xsec.grid_weights)
+        gw = physics.xsec.grid_weights
+        @info "Precomputing per-energy xsec weight arrays for ORCA MC"
+        (
+            nue_cc      = gw(binning.e_fine, :nue,   :CC, false),
+            nuebar_cc   = gw(binning.e_fine, :nue,   :CC, true),
+            numu_cc     = gw(binning.e_fine, :numu,  :CC, false),
+            numubar_cc  = gw(binning.e_fine, :numu,  :CC, true),
+            nutau_cc    = gw(binning.e_fine, :nutau, :CC, false),
+            nutaubar_cc = gw(binning.e_fine, :nutau, :CC, true),
+            nue_nc      = gw(binning.e_fine, :nue,   :NC, false),
+            nuebar_nc   = gw(binning.e_fine, :nue,   :NC, true),
+            numu_nc     = gw(binning.e_fine, :numu,  :NC, false),
+            numubar_nc  = gw(binning.e_fine, :numu,  :NC, true),
+            nutau_nc    = gw(binning.e_fine, :nutau, :NC, false),
+            nutaubar_nc = gw(binning.e_fine, :nutau, :NC, true),
+        )
+    else
+        nothing
+    end
+
+    assets = (;mc, muon_hist, observed=cut(data_hist), binning, true_shape, reco_shape, nominal_layers, flux_nominal, xsec_eval)
 end
 
 function get_params()
@@ -124,9 +145,12 @@ function make_hist(e_idx, c_idx, p_idx, t_idx, w, size=(8,8,2,2))
     hist
 end
 
-function make_hist_per_channel(mc, osc_flux, lifetime_seconds, params, assets)
+function make_hist_per_channel(mc, osc_flux, lifetime_seconds, params, assets, xsec_cc=nothing, xsec_nc=nothing)
     w = lifetime_seconds * mc.W .* osc_flux .* (mc.he_mask * (params.orca_norm_he - 1.) .+ 1.0)
-    h = make_hist(mc.E_reco_bin, mc.Ct_reco_bin, mc.AnaClass, mc.IsCC .+ 1, w, assets.reco_shape)
+    if !isnothing(xsec_cc) && !isnothing(xsec_nc)
+        w = w .* ifelse.(mc.IsCC .== 1, xsec_cc[mc.E_true_bin], xsec_nc[mc.E_true_bin])
+    end
[…]
