+++ test/test_regression.jl
@@ -10,9 +10,9 @@ const ALL_REFERENCE_VALUES = (
     dayabay      = -168.90003273508322,
     kamland      = -63.111403860037875,
     minos        = -268.4046154382636,
-    deepcore     = -950.625021522011,
-    super_k      = -3370.642581,
-    orca         = -1164.3664813530054,
+    deepcore     = -957.0298637044948,
+    super_k      = -3433.766267280389,
+    orca         = -1173.9109711413255,
     coherent_csi = -574.3416032522271,
     coherent_lAr = -1754.9946939034153,
 )

[…]
