### 2026-04-23 1f851ed
Fix DeepCore hypersurface indexing for IO (negative Δm²₃₁)
The hypersurface grid (hs_dm31) only covers positive NO values
LinRange(0.0015, 0.0035, 20). With IO, Δm²₃₁ < 0, causing the
floor-index to be deeply negative (-38), resulting in a BoundsError.

Fix: use |Δm²₃₁| before indexing (detector response depends only on
the mass splitting magnitude, not the sign/ordering), and clamp to the
grid edges for robustness. Also fix the latent edge-case bug where
x = X[1] exactly would produce idx=0 (invalid in Julia 1-based indexing).

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
