### 2026-04-24 31b3a1a
Implement ORCA energy scale as reco-bin transfer matrix
Replace true-energy scaling in osc_prob with a bin-overlap transfer
matrix applied to reconstructed energy histograms. This is physically
correct: the energy scale uncertainty is a detector effect on reco
energy, not a shift in true neutrino energy.

- energy_transfer_matrix(scale, e_edges): T[k,j] = fraction of reco
  bin j mapped into measured bin k when reco energy is scaled by scale
- apply_energy_scale: applies T via matrix multiply along energy axis
- NaN guard added (max floor doesn't catch NaN; rare extreme param combos
  can prod
[…]
