+++ src/experiments/icecube/deepcore_9y_verification_sample/README.md
@@ -9,15 +9,15 @@ IceCube Collaboration, 2025, "Replication Data for: Measurement of atmospheric n
 ![DataMC](test_output/datamc.png)
 ## Meta Information
 - **repo_clean**: false
-- **exec_time**: 3506.3720021247864
+- **exec_time**: 8.058847904205322
 - **username**: peller
 - **nseeds**: 1
 - **repo**: /mnt/c/Users/peller/work/claude/Newtrinos.jl
 - **cache_dir**: test
 - **hostname**: flippy
-- **params**: (atm_flux_delta_spectral_index = 0.0, atm_flux_norm_ratio_hi = 1.0, atm_flux_norm_ratio_lo = 1.0, atm_flux_nuenuebar_sigma_hi = 0.0, atm_flux_nuenuebar_sigma_lo = 0.0, atm_flux_nuenuebar_sigma_mid = 0.0, atm_flux_nuenumu_sigma_hi = 0.0, atm_flux_nuenumu_sigma_lo = 0.0, atm_flux_nuenumu_sigma_mid = 0.0, atm_flux_numunumubar_sigma_hi = 0.0, atm_flux_numunumubar_sigma_lo = 0.0, atm_flux_numunumubar_sigma_mid = 0.0, atm_flux_updown_sigma = 0.0, atm_flux_uphorizontal_sigma = 0.0, deepcore_aeff_scale = 1.0, deepcore_atm_muon_scale = 1.0, deepcore_ice_absorption = 1.0, deepcore_ice_scattering = 1.0, deepcore_opt_eff_overall = 1.0, deepcore_rel_eff_p0 = 0.1, deepcore_rel_eff_p1 = -0.05, nc_norm = 1.0, nutau_cc_norm = 1.0, Δm²₂₁ = 7.53e-5, Δm²₃₁ = 0.0024752999999999997, δCP = 1.0, θ₁₂ = 0.5872523687443223, θ₁₃ = 0.1454258194533693, θ₂₃ = 0.8556288707523761)
-- **date**: 2026-04-22 11:07:36
[…]
