### 2026-04-24 81af850
Revert ORCA energy scale to true-energy shifting
Reco-bin transfer matrix approach gave incorrect results; reverting to
the original implementation that shifts true energy in osc_prob.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
