+++ src/experiments/km3net/orca6_433kton/README.md
@@ -7,15 +7,14 @@ Data source: https://opendata.km3net.de/dataset.xhtml?persistentId=doi:10.5072/F
 ![DataMC](test_output/datamc.png)
 ## Meta Information
 - **repo_clean**: false
-- **exec_time**: 296.1259431838989
+- **exec_time**: 29.718905210494995
 - **username**: peller
-- **nseeds**: 1
-- **repo**: /mnt/c/Users/peller/work/claude/Newtrinos.jl
+- **repo**: /mnt/c/Users/peller/work/Newtrinos
 - **cache_dir**: test
 - **hostname**: flippy
-- **params**: (atm_flux_delta_spectral_index = 0.0, atm_flux_norm_ratio_hi = 1.0, atm_flux_norm_ratio_lo = 1.0, atm_flux_nuenuebar_sigma_hi = 0.0, atm_flux_nuenuebar_sigma_lo = 0.0, atm_flux_nuenuebar_sigma_mid = 0.0, atm_flux_nuenumu_sigma_hi = 0.0, atm_flux_nuenumu_sigma_lo = 0.0, atm_flux_nuenumu_sigma_mid = 0.0, atm_flux_numunumubar_sigma_hi = 0.0, atm_flux_numunumubar_sigma_lo = 0.0, atm_flux_numunumubar_sigma_mid = 0.0, atm_flux_updown_sigma = 0.0, atm_flux_uphorizontal_sigma = 0.0, orca_energy_scale = 1.0, orca_norm_all = 1.0, orca_norm_he = 1.0, orca_norm_hpt = 1.0, orca_norm_muons = 1.0, orca_norm_showers = 1.0, xsec_cc1p1h_multigev_norm = 1.0, xsec_cc1p1h_nubar_ratio = 1.0, xsec_cc1p1h_nue_numu_ratio = 1.0, xsec_cc1p1h_shape = 0.0, xsec_cc1p1h_subgev_norm = 1.0, xsec_cc1pi_norm = 1.0, xsec_cc1pi_nubar_ratio = 1.0, xsec_cc1pi_shape = 0.0, xsec_cc2p2h_norm = 1.0, xsec_cc2p2h_nubar_ratio = 1.0, xsec_cc2p2h_shape = 0.0, xsec_ccdis_norm = 1.0, xsec_ccdis_nubar_ratio = 1.0, xsec_ccdis_shape = 0.0, xsec_ccother_norm = 1.0, xsec_ccother_nubar_ratio = 1.0, xsec_ccother_shape = 0.0, xsec_nc_norm = 1.0, xsec_nc_nubar_ratio = 1.0, xsec_nc_shape = 0.0, xsec_nutau_cc_norm = 1.0, Δm²₂₁ = 7.53e-5, Δm²₃₁ = 0.0024752999999999997, δCP = 1.0, θ₁₂ = 0.5872523687443223, θ₁₃ = 0.1454258194533693, θ₂₃ = 0.8556288707523761)
-- **date**: 2026-04-24 10:08:38
[…]
