### 2026-04-24 911e4bc
Add NaN guard to ORCA expected counts
max(1e-2, NaN) = NaN in Julia, so extreme parameter combinations that
produce NaN in oscillation probabilities would propagate to Poisson.
Use ifelse to catch NaN and replace with the floor value.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
