### 2026-04-27 0cf2310
Add refine_profile task with outlier repair and Gaussian smoothing
smooth_result(): detects failed scan-point fits (log_posterior much lower
than neighbours), replaces their parameter values with neighbour averages,
then applies a separable Gaussian blur across the grid.

make_seed_interpolants(): builds Gridded(Linear()) interpolants with linear
extrapolation for every parameter, so seeds can be evaluated at any grid
point including finer/coarser/shifted grids.

profile() gains seed_result keyword: interpolates smoothed parameter values
at each scan-point coordinate and passes them as per-poi
[…]
