### 2026-04-27 49a2c85
changing some priors in analysis.jl
