### 2026-04-27 4e9a687
Add sequential warm-starting for smooth profile scans
Implements BFS-ordered sequential scan where each grid point is seeded
by its nearest already-computed neighbour's optimised parameters. This
keeps the optimizer in the same basin of attraction across the grid,
eliminating jumpy profiles caused by LBFGS landing in different local
minima at neighbouring scan points.

New features:
- `_sequential_order(grid_shape, center_linear)`: BFS traversal of grid
  with parent tracking for warm-starting
- `profile(...; sequential=false, start_from=nothing)`: when sequential=true,
  runs BFS from the gri
[…]
