### 2026-04-27 59ca336
Update ORCA regression reference for spray propagation; add gradient diagnostic
Spray shifts the ORCA LLH from -1173.911 to -1174.064 at default params.

Also adds check_orca_energy_scale_gradient.jl which confirmed the orca_energy_scale
gradient is smooth and accurate (ForwardDiff ≈ finite differences throughout).
The profile jumpiness is therefore caused by local minima in the full nuisance
parameter space, not by bad energy-scale gradients.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
