### 2026-04-27 3c812fa
Switch DeepCore to Spray propagation; update regression reference
Same Nyquist-matched spray as Super-K and ORCA:
- dlogE = 4/200 = 0.02 (200 bins over 4 log10-E decades, matching LinRange(0,4,201))
- σ_E = 10^0.02 - 1 ≈ 0.047 (same as Super-K)
- σ_h = 10 km production height uncertainty

Regression: -957.030 → -958.074

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
