+++ test/test_regression.jl
@@ -10,7 +10,7 @@ const ALL_REFERENCE_VALUES = (
     dayabay      = -168.90003273508322,
     kamland      = -63.111403860037875,
     minos        = -268.4046154382636,
-    deepcore     = -957.0298637044948,
+    deepcore     = -958.0740617933973,
     super_k      = -3433.766267280389,
     orca         = -1174.064185239588,
     coherent_csi = -574.3416032522271,
