### 2026-04-29 38e46aa
Reparametrize, correct systematics, add t-efficiency
