+++ src/experiments/coherent/coherent_2020/coherent_csi.jl
@@ -24,7 +24,74 @@ import ..Newtrinos
     plot::Function
 end
 
-function configure(; datadir = @__DIR__, use_flux_data::Bool = true, ff_model::Symbol = :helm, ff_kwargs::NamedTuple = (;))
+function rebin_1d_projection(centers, weights, edges)
+    rebinned = zeros(eltype(weights), length(edges) - 1)
+    for i in eachindex(rebinned)
+        idx = findall(c -> c >= edges[i] && c < edges[i + 1], centers)
+        if !isempty(idx)
+            rebinned[i] = sum(weights[idx])
+        end
+    end
+    return rebinned
+end
+
+function edges_from_centers(centers)
+    n = length(centers)
+    @assert n >= 2
+    edges = similar(centers, n + 1)
+    edges[2:n] .= (centers[1:end-1] .+ centers[2:end]) ./ 2
+    edges[1] = centers[1] - (edges[2] - centers[1])
+    edges[end] = centers[end] + (centers[end] - edges[end-1])
+    return edges
+end
+
+function rebin_histogram_1d(source_weights, source_edges, target_edges)
+    T = promote_type(eltype(source_weights), eltype(source_edges), eltype(target_edges))
+    rebinned = zeros(T, length(target_edges) - 1)
+    dt = diff(source_edges)
+    density = source_weights ./ dt
+
+    for out_bin in eachindex(rebinned)
+        out_lo = target_edges[out_bin]
+        out_hi = target_edges[out_bin + 1]
+        for in_bin in eachindex(source_weights)
+            in_lo = source_edges[in_bin]
+            in_hi = source_edges[in_bin + 1]
+            overlap = min(out_hi, in_hi) - max(out_lo, in_lo)
+            if overlap > zero(overlap)
+                rebinned[out_bin] += density[in_bin] * overlap
+            end
+        end
+    end
+
+    return rebinned
+end
+
+function get_timing_efficiency(time_edges)
+    a_ns = 520.0
+    b_per_ns = 0.0494 / 1e3
+    eff = zeros(Float64, length(time_edges) - 1)
+
+    for i in eachindex(eff)
+        lo = time_edges[i]
+        hi = time_edges[i + 1]
+        width = hi - lo
+
+        integral = if hi <= a_ns
+            width
+        elseif lo >= a_ns
+            (exp(-b_per_ns * (lo - a_ns)) - exp(-b_per_ns * (hi - a_ns))) / b_per_ns
+        else
+            (a_ns - lo) + (1 - exp(-b_per_ns * (hi - a_ns))) / b_per_ns
+        end
+
+        eff[i] = integral / width
+    end
+
+    return eff
+end
+
+function configure(; datadir = @__DIR__, use_flux_data::Bool = true, ff_model::Symbol = :helm, ff_kwargs::NamedTuple = (;), sns_flux_kwargs::NamedTuple = (;))
     assets = get_assets(datadir)
 
     # Configure the SNS flux module
@@ -32,6 +99,7 @@ function configure(; datadir = @__DIR__, use_flux_data::Bool = true, ff_model::S
         exposure = assets.exposure,
         distance = assets.distance,
         use_data = use_flux_data,
+        sns_flux_kwargs...
     )
 
     # Reconfigure assets with data loaded from sns_flux
@@ -61,34 +129,24 @@ end
 
 function get_params(ss_bkg_nom, brn_nom, nin_nom)
     params = (
-        coherent_csi_eff_a = 1.32045,
-        coherent_csi_eff_b = 0.285979,
-        coherent_csi_eff_c = 10.8646,
-        coherent_csi_eff_d = -0.333322,
[…]
