### 2026-04-30 #9644 @matthewfeickert
👋 @philippeller, I can see that there's development happening on https://github.com/philippeller/Newtrinos.jl (which is good! 👍), but can you give us a check in update on how addressing issues is going?
