### 2026-04-30 ccc9d1d
Add IFT profile and multi-file seed merging for joint fits
- `ift_profile` in molewhacker.jl: IFT-guided profile posterior scan using
  MGVI covariance to predict nuisance parameter updates at each scan step,
  with optional LBFGS polish in prior-normalized z-space
- `_ift_predict_z`: helper that computes the cross-derivative via nested
  ForwardDiff and applies the MGVI covariance for the z-space prediction
- `toy_ift_profile.jl`: toy validation (2D correlated Gaussian) confirming
  the IFT predictor is exact to machine precision for quadratic posteriors
- `make_merged_seed_interpolants`: mer
[…]
