### 2026-04-30 3e0c8ec
Fix NaN propagation through Poisson floor in Super-K forward model
max(1e-3, NaN) returns NaN in IEEE 754, causing Poisson(NaN) DomainError
during cluster fits at extreme parameter combinations. Guard with ifelse
to match existing ORCA pattern. All 8 regression tests pass unchanged.

Also add IFT profile comparison scripts and algorithm PDF.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
