+++ docs/ift_profile.tex
@@ -1,0 +1,195 @@
+\documentclass[11pt, a4paper]{article}
+\usepackage[margin=2.5cm]{geometry}
+\usepackage{amsmath, amssymb}
+\usepackage{mathtools}
+\usepackage{bm}
+\usepackage{hyperref}
+\usepackage{xcolor}
+\usepackage{booktabs}
+\usepackage{enumerate}
+
+\hypersetup{colorlinks=true, linkcolor=blue, urlcolor=blue}
+
+\title{\textbf{IFT-Guided Profile Likelihood Scan}\\[0.5em]
+\large Algorithm Description}
+\author{Newtrinos.jl}
+\date{}
+
+\newcommand{\nupar}{\bm{\nu}}
+\newcommand{\zz}{\mathbf{z}}
+\newcommand{\Sig}{\boldsymbol{\Sigma}}
+\newcommand{\Lam}{\boldsymbol{\Lambda}}
+
+% Simple framed pseudocode using standard tabbing
+\newenvironment{pseudocode}[1]{%
+  \bigskip
+  \noindent\textbf{Algorithm:} #1\\[-0.3em]
+  \noindent\rule{\linewidth}{0.8pt}\\[-0.5em]
+  \noindent\begin{enumerate}
+}{%
+  \end{enumerate}
+  \noindent\rule{\linewidth}{0.4pt}
+  \bigskip
+}
+
+\begin{document}
+\maketitle
+
+\section*{Motivation}
+
+Profile likelihood scans require solving a separate optimisation at each
+grid point $\vartheta_k$:
+\begin{equation}
+  \hat{\nupar}(\vartheta_k)
+  \;=\; \arg\max_{\nupar}\;\bigl[\log\mathcal{L}(\vartheta_k, \nupar)
+                                  + \log\pi(\nupar)\bigr],
+\end{equation}
+where $\mathcal{L}$ is the likelihood and $\pi(\nupar)$ the nuisance prior.
+For experiments with $\mathcal{O}(40\text{--}90)$ nuisance parameters, each
+optimisation is expensive and a poor starting point can trap the optimiser in a
+local minimum.
+
+The IFT-guided scan uses a local Gaussian (MGVI) approximation of the posterior
+at the current optimum to \emph{predict} the nuisance optimum at the next grid
+point, avoiding cold restarts.
+
+\section{Prior-Normalised Coordinates ($z$-space)}
+
+Every nuisance parameter $\nu_i$ with prior $\pi_i$ is mapped to
+\begin{equation}
+  z_i \;=\; \Phi^{-1}\!\bigl(F_{\pi_i}(\nu_i)\bigr),
+\end{equation}
+where $F_{\pi_i}$ is the prior CDF and $\Phi^{-1}$ the standard-Normal quantile
+function.  Collecting all nuisance parameters gives
+$\zz \in \mathbb{R}^n$ with prior $\zz \sim \mathcal{N}(\mathbf{0}, \mathbf{I})$.
+
+The scan parameter $\vartheta_k$ is fixed to a point mass at each grid value and
+is \emph{excluded} from $\zz$; it never enters the $z$-space optimisation.
+
+At grid point $\vartheta_k$ the $z$-space posterior is
+\begin{equation}
+  \log p_k(\zz)
+  \;=\; \log\mathcal{L}\!\bigl(\vartheta_k,\, f(\zz)\bigr)
+        - \tfrac{1}{2}\|\zz\|^2 + \text{const},
+\end{equation}
+where $f:\mathbb{R}^n\!\to\!\mathbb{R}^n$ maps back to physical nuisance space.
+
+\section{Local MGVI Approximation}
+
[…]

[…]
