### 2026-04-30 f0dfad9
Fix make_merged_seed_interpolants dispatch on Vector{Any} from JLD2
FileIO.load returns Dict{String,Any}, so seed_results_list built from
a comprehension is Vector{Any}. Relax AbstractVector{NewtrinosResult}
to AbstractVector so dispatch succeeds. Also add linear-fit smoothing
script and improve profile/ift_profile comparison to use a shared MLE
seed for a fair comparison.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
