### 2026-04-30 2d71e57
Track optimizer convergence per scan point in NewtrinosResult
find_mle now captures res.info.retcode from bat_findmode and returns a
4th element (converged::Bool). find_mle_cached saves/loads it with
backward-compatible default=true for old cache files. assemble_profile_results
adds a converged::Bool array to result.values and warns if any points
failed. ift_profile captures the same from its z-space LBFGS polish.
smooth_result_linear skips the converged field (non-numeric).

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
