### 2026-04-30 5789daa
Switch LBFGS stopping criterion from reltol to g_tol
reltol=1e-7 (relative function change) terminates after 0-2 iterations
when warm-starting near the optimum, leaving nuisances un-re-optimized
for the new scan point value. This caused flat nuisance tracks in
sequential profile scans and jagged profiles in global fits.

g_tol=1e-5 (gradient norm) stops only at a true stationary point,
regardless of starting function value. maxiters raised to 2000.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
