### 2026-05-03 f2bc3f5
Add ift_profile task to analysis CLI
Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
