### 2026-05-05 793c459
Fix ORCA forward model: reco energy scale, oversampled grid, remove double-counted flux norm
Three physics fixes to the ORCA forward model:

1. Move energy scale from true-energy multiplication to reco-bin overlap shifting
   (same bin-transfer-matrix approach as Super-K), which is the physically correct
   treatment of a reconstructed energy miscalibration.

2. Oversample the true energy grid 5× (200 pts, matching DeepCore/SK resolution),
   evaluating flux systematics, oscillation probabilities, and xsec weights on the
   fine grid before averaging (downsampling) back to the 40-bin MC grid. 
[…]
