+++ src/experiments/km3net/orca6_433kton/README.md
@@ -6,8 +6,8 @@ Data source: https://opendata.km3net.de/dataset.xhtml?persistentId=doi:10.5072/F
 ![Comparison](test_output/contours.png)
 ![DataMC](test_output/datamc.png)
 ## Meta Information
-- **repo_clean**: true
-- **exec_time**: 13853.014309883118
+- **repo_clean**: false
+- **exec_time**: 3156.85036611557
 - **username**: peller
 - **nseeds**: 1
 - **repo**: /mnt/c/Users/peller/work/claude/Newtrinos.jl
@@ -15,8 +15,8 @@ Data source: https://opendata.km3net.de/dataset.xhtml?persistentId=doi:10.5072/F
 - **sequential**: false
 - **hostname**: flippy
 - **params**: (atm_flux_delta_spectral_index = 0.0, atm_flux_norm_ratio_hi = 1.0, atm_flux_norm_ratio_lo = 1.0, atm_flux_nuenuebar_sigma_hi = 0.0, atm_flux_nuenuebar_sigma_lo = 0.0, atm_flux_nuenuebar_sigma_mid = 0.0, atm_flux_nuenumu_sigma_hi = 0.0, atm_flux_nuenumu_sigma_lo = 0.0, atm_flux_nuenumu_sigma_mid = 0.0, atm_flux_numunumubar_sigma_hi = 0.0, atm_flux_numunumubar_sigma_lo = 0.0, atm_flux_numunumubar_sigma_mid = 0.0, atm_flux_updown_sigma = 0.0, atm_flux_uphorizontal_sigma = 0.0, orca_energy_scale = 1.0, orca_norm_all = 1.0, orca_norm_he = 1.0, orca_norm_hpt = 1.0, orca_norm_muons = 1.0, orca_norm_showers = 1.0, xsec_cc1p1h_multigev_norm = 1.0, xsec_cc1p1h_nubar_ratio = 1.0, xsec_cc1p1h_nue_numu_ratio = 1.0, xsec_cc1p1h_shape = 0.0, xsec_cc1p1h_subgev_norm = 1.0, xsec_cc1pi_norm = 1.0, xsec_cc1pi_nubar_ratio = 1.0, xsec_cc1pi_shape = 0.0, xsec_cc2p2h_norm = 1.0, xsec_cc2p2h_nubar_ratio = 1.0, xsec_cc2p2h_shape = 0.0, xsec_ccdis_norm = 1.0, xsec_ccdis_nubar_ratio = 1.0, xsec_ccdis_shape = 0.0, xsec_ccother_norm = 1.0, xsec_ccother_nubar_ratio = 1.0, xsec_ccother_shape = 0.0, xsec_nc_norm = 1.0, xsec_nc_nubar_ratio = 1.0, xsec_nc_shape = 0.0, xsec_nutau_cc_norm = 1.0, Δm²₂₁ = 7.53e-5, Δm²₃₁ = 0.0024752999999999997, δCP = 1.0, θ₁₂ = 0.5872523687443223, θ₁₃ = 0.1454258194533693, θ₂₃ = 0.8556288707523761)
-- **date**: 2026-04-28 13:17:15
+- **date**: 2026-05-04 17:45:05
 - **task**: profile
-- **vars_to_scan**: OrderedDict{Any, Any}(:θ₂₃ => 11, :Δm²₃₁ => 11)
-- **commit_hash**: 3c812fa952a46f7bf6b2229105820765652ef107
[…]
