### 2026-05-06 1e8afed
Switch find_mle optimizer from OptimizationAlg/Optimization.LBFGS to OptimAlg/Optim.LBFGS
Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
