### 2026-05-07 171e117
Add find_mle_ext (BAT-free Fminbox optimizer) and fix ORCA get_expected
- Add find_mle_ext: bypasses BAT's bat_findmode entirely, using
  Optim.Fminbox(LBFGS(m=20)) with bounds from prior distributions and
  ForwardDiff gradients. Convergence criterion is g_tol=1e-6 (gradient-
  based) rather than f_abstol, which is more robust for combined fits
  where one experiment dominates the gradient.
- Switch find_mle_cached to use find_mle_ext for all profile/scan points.
- Fix ORCA get_expected: nc_norm and nutau_cc_norm were not applied to
  NC and nutau CC histograms, giving zero gradients for thos
[…]
