### 2026-05-07 d61f19f
Revert orca.jl to pre-session state
Undo apply_energy_scale rewrite and get_expected nc_norm/nutau_cc_norm
changes — those were debugging artifacts for SimpleScaling xsec and are
incorrect with H2O_PCA (which handles NC/nutau normalisation internally
via xsec_w_os weights and has no params.nc_norm field).

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
