+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"category":"section","location":"api/analysis/#Analysis-API","page":"Analysis","text":"","title":"Analysis API"},{"category":"section","location":"api/analysis/#Likelihood-Construction","page":"Analysis","text":"","title":"Likelihood Construction"},{"category":"section","location":"api/analysis/#Scanning-and-Profiling","page":"Analysis","text":"","title":"Scanning and Profiling"},{"category":"section","location":"api/analysis/#Data-Generation","page":"Analysis","text":"","title":"Data Generation"},{"category":"section","location":"api/analysis/#Utilities","page":"Analysis","text":"","title":"Utilities"},{"category":"function","location":"api/analysis/#Newtrinos.generate_likelihood","page":"Analysis","text":"generate_likelihood(experiments[, observed])\n\nConstruct a joint likelihood from a NamedTuple of configured experiments. Combines all forward models and observed data into a single likelihood object compatible with DensityInterface.logdensityof.\n\n\n\n\n\n","title":"Newtrinos.generate_likelihood"},{"category":"function","location":"api/analysis/#Newtrinos.get_params","page":"Analysis","text":"get_params(x)\n\nExtract nominal parameter values as a sorted NamedTuple. Works on Physics, Experiment, Wrapper, or a NamedTuple of modules (merging all with conflict checking).\n\n\n\n\n\n","title":"Newtrinos.get_params"},{"category":"function","location":"api/analysis/#Newtrinos.get_priors","page":"Analysis","text":"get_priors(x)\n\nExtract prior distributions as a sorted NamedTuple. Works on Physics, Experiment, Wrapper, or a NamedTuple of modules (merging all with conflict checking).\n\n\n\n\n\n","title":"Newtrinos.get_priors"},{"category":"function","location":"api/analysis/#Newtrinos.condition","page":"Analysis","text":"condition(priors, conditional_vars, params)\n\nFix parameters to specific values by replacing their priors with constants. conditional_vars can be an Array of symbols (uses values from params) or a Dict mapping symbols to values.\n\n\n\n\n\n","title":"Newtrinos.condition"},{"category":"function","location":"api/analysis/#Newtrinos.profile","page":"Analysis","text":"profile(likelihood, priors, vars_to_scan, params; cache_dir=nothing, map_func=nothing)\n\nRun a profile likelihood scan. At each grid point defined by vars_to_scan, optimizes over all other parameters. Use cache_dir to cache and resume results. Use map_func=pmap for distributed parallelism (default: Threads.@threads).\n\n\n\n\n\n","title":"Newtrinos.profile"},{"category":"function","location":"api/analysis/#Newtrinos.scan","page":"Analysis","text":"scan(likelihood, priors, vars_to_scan, params; gradient_map=false)\n\nRun a simple likelihood scan on a grid (no optimization over nuisance parameters). Faster than profile but does not account for nuisance parameter variations. Set gradient_map=true to also compute gradients at each point.\n\n\n\n\n\n","title":"Newtrinos.scan"},{"category":"function","location":"api/analysis/#Newtrinos.find_mle","page":"Analysis","text":"find_mle(likelihood, prior, params; adsel=AutoPolyesterForwardDiff())\n\nFind the Maximum Likelihood Estimator using LBFGS optimization via BAT. Returns (llh, log_posterior, optimized_params). Parameters with ConstValueDist priors are held fixed.\n\n\n\n\n\n","title":"Newtrinos.find_mle"},{"category":"function","location":"api/an
[…]
