### 2026-05-10 43dc1f1
added tests for analysis_tools.jl
