+++ test/test_analysis.jl
@@ -1,14 +1,50 @@
 using DataStructures
+using Test
+using Newtrinos
+using BAT
+using Distributions
+using LinearAlgebra
+using DensityInterface
+using MeasureBase
+using ValueShapes
+
+struct MockPhysics <: Newtrinos.Physics
+    params::NamedTuple
+    priors::NamedTuple
+end
+
+struct MockExperiment <: Newtrinos.Experiment
+    physics::MockPhysics
+    params::NamedTuple
+    priors::NamedTuple
+    assets::NamedTuple
+    forward_model::Function
+    plot::Function
+end
+
+struct ArgumentErrorDensity end
+DensityInterface.DensityKind(::ArgumentErrorDensity) = IsDensity()
+DensityInterface.logdensityof(::ArgumentErrorDensity, x) = throw(ArgumentError("forced"))
 
 @testset "Analysis Tools" begin
 
+    @testset "NewtrinosResult" begin
+        axes = (x=[1.0, 2.0, 3.0],)
+        values = (llh=[-10.0, -5.0, -8.0], log_posterior=[-10.0, -5.0, -8.0])
+        result = Newtrinos.NewtrinosResult(axes=axes, values=values)
+
+        @test result.axes == axes
+        @test result.values == values
+        @test result.meta isa Dict
+    end
+
     @testset "sort_nt" begin
-        nt = (c=3, a=1, b=2)
+        nt = (b=3, a=1, c=2)
         sorted = Newtrinos.sort_nt(nt)
         @test keys(sorted) == (:a, :b, :c)
         @test sorted.a == 1
-        @test sorted.b == 2
-        @test sorted.c == 3
+        @test sorted.b == 3
+        @test sorted.c == 2
 
         # Already sorted
         nt2 = (a=1, b=2)
@@ -45,55 +81,610 @@ using DataStructures
         @test keys(merged3) == (:a, :m, :z)
     end
 
-    @testset "NewtrinosResult" begin
-        axes = (x=[1.0, 2.0, 3.0],)
-        values = (llh=[-10.0, -5.0, -8.0], log_posterior=[-10.0, -5.0, -8.0])
-        result = Newtrinos.NewtrinosResult(axes=axes, values=values)
+    @testset "Wrapper and Base.getproperty" begin
+        mock_physics = MockPhysics(
+            (mu = 5.0,),
+            (mu = Uniform(0.0, 10.0),)
+        )
+        mock_exp = MockExperiment(
+            mock_physics,
+            (scale = 1.0,),
+            (scale = Uniform(0.5, 2.0),),
+            (observed = [5.0],),
+            p -> MvNormal([p.mu * p.scale], I(1)),
+            (p, data=nothing) -> nothing
+        )
 
-        @test result.axes == axes
-        @test result.values == values
-        @test result.meta isa Dict
+        aliases = Dict(:mu => :mean_value)
+        wrapper = Newtrinos.Wrapper(mock_exp, aliases)
 
-        # bestfit should find the maximum log_posterior
-        bf = Newtrinos.bestfit(result)
-        @test bf.log_posterior == -5.0
-        @test bf.x == 2.0  # axis value at best fit
+        params = Newtrinos.get_params(wrapper)
+        @test haskey(params, :mean_value)
+        @test !haskey(params, :mu)
+        @test params.mean_value == 5.0
+        @test params.scale == 1.0
+
+        priors = Newtrinos.get_priors(wrapper)
+        @test haskey(priors, :mean_value)
+        @test priors.mean_value isa Uniform
+
+        # test correct call of Base.getproperty branches
+        @test Newtrinos.Base.getproperty(wrapper, :forward_model) == wrapper.forward_model
+        @test Newtrinos.Base.getproperty(wrapper, :plot) == wrapper.plot
[…]
