### 2026-05-16 e58cfaf
added tests for cli_common.jl
