### 2026-05-17 be0a257
CompatHelper: bump compat for Optim to 2, (keep existing compat)
