+++ Project.toml
@@ -122,7 +122,7 @@ MeasureBase = "0.14.13"
 Memoize = "0.4.4"
 MonotonicSplines = "0.3.3"
 NPZ = "0.4.3"
-Optim = "1.11.0"
+Optim = "1.11.0, 2"
 Optimization = "4.8.0"
 PCHIPInterpolation = "0.2.1"
 PDMats = "0.11"
