### 2026-05-20 18803a2
added tests for molewhacker.jl
