### 2026-05-21 214df67
added documentation for analysis_tools.jl, molewhacker.jl, and cli_common.jl. in the markdown file mentioned cli_common and analysis.jl as Command Line entry for analyses and described how to use it
