+++ docs/src/api/analysis.md
@@ -1,25 +1,42 @@
 # Analysis API
 
-## Likelihood Construction
+The analysis layer provides inference tools that treat experiments as black boxes.
+It is composed of three source files:
+
+- **`analysis_tools.jl`** — core types, parameter utilities, likelihood generation,
+  scanning, profiling, and data generation. All symbols are part of the `Newtrinos`
+  module.
+- **`molewhacker.jl`** — adaptive importance sampling via MGVI Gaussian mixture
+  refinement. All symbols are part of the `Newtrinos` module.
+- **`cli_common.jl`** — helper functions for CLI scripts. These are standalone
+  utilities loaded via `include` and are **not** exported from the `Newtrinos` module.
+
+---
+
+## Core Types
 
 ```@docs
-Newtrinos.generate_likelihood
+Newtrinos.NewtrinosResult
+Newtrinos.Wrapper
+```
+
+---
+
+## Parameter Utilities
+
+These functions collect and merge parameters and priors across physics and experiment
+modules. `condition` and `correlated_priors_vars` allow fixing or correlating
+individual parameters before constructing a posterior.
+
+```@docs
 Newtrinos.get_params
 Newtrinos.get_priors
 Newtrinos.condition
+Newtrinos.safe_merge
+Newtrinos.correlated_priors_vars
 ```
 
-## Scanning and Profiling
-
-```@docs
-Newtrinos.profile
-Newtrinos.scan
-Newtrinos.find_mle
-Newtrinos.find_mle_cached
-Newtrinos.generate_scanpoints
-Newtrinos.bestfit
-Newtrinos.assemble_profile_results
-```
+---
 
 ## Data Generation
 
@@ -28,10 +45,197 @@ Newtrinos.generate_asimov_data
 Newtrinos.generate_toy_data
 ```
 
+---
+
+## Adaptive Importance Sampling
+
+The `molewhacker` tools implement an adaptive importance sampling strategy. Starting
+from an initial Gaussian mixture proposal, each iteration fits a new local MGVI
+Gaussian component at the highest-weight sample point and adds it to the mixture. 
+This "whack-a-mole" strategy adaptively targets undersampled regions of the posterior.
+`whack_many_moles` is the recommended entry point for production runs. 
+
+```@docs
+Newtrinos.make_prior_samples
+Newtrinos.make_init_samples
+Newtrinos.whack_a_mole
+Newtrinos.whack_many_moles
+Newtrinos.importance_sampling
+Newtrinos.local_MGVI_approx
+```
+
+---
+
+## Likelihood Construction
+
+```@docs
+Newtrinos.generate_likelihood
+Newtrinos.get_observed
+Newtrinos.get_fwd_model
+```
+
+---
+
+## Scanning and Profiling
+
+`scan` evaluates the joint likelihood on a fixed grid without optimization.
+`profile` minimizes over nuisance parameters at each grid point.
+`find_mle` / `find_mle_cached` locate the global maximum-likelihood estimate.
+
+```@docs
+Newtrinos.find_mle
+Newtrinos.find_mle_cached
+Newtrinos.profile
+Newtrinos.scan
+Newtrinos.bestfit
+Newtrinos.generate_scanpoints
+Newtrinos.assemble_profile_results
+```
+
+---
+
 ## Utilities
 
 ```@docs
-Newtrinos.safe_merge
-Newtrinos.correlated_priors_vars
 Newtrinos.add_meta!
+```
+
+---
+
+## Internal API
+
+Low-level helpers used internally by the scanning and profiling pipeline.
+
+```@docs
+Newtrinos.sort_nt
+Newtrinos._generate_grid
+Newtrinos._profile
+```
+
+---
+
+## CLI Usage
+
[…]
