### 2026-05-21 01b38fb
fixed double docstring call of types that were already called in types.md and fixed some warnings
