+++ README.md
@@ -1,7 +1,8 @@
-<!--- badges of CI, coverage, license, docs(missing)-->
 [![CI](https://github.com/davschu/Newtrinos.jl/actions/workflows/CI.yml/badge.svg)](https://github.com/davschu/Newtrinos.jl/actions/workflows/CI.yml)
 [![codecov](https://codecov.io/github/davschu/Newtrinos.jl/graph/badge.svg?token=OTXXQIR8GW)](https://codecov.io/github/davschu/Newtrinos.jl)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+[![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://davschu.github.io/Newtrinos.jl/stable/)
+[![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://davschu.github.io/Newtrinos.jl/dev/)
 
 # Newtrinos.jl
 
