### 2026-06-02 aba606c
before paper changes
