+++ juno_test_NND_instrumented.jl
@@ -1,0 +1,195 @@
+using LinearAlgebra
+using Distributions
+using LaTeXStrings
+using Printf
+using FileIO
+import JLD2
+using DataFrames
+using Setfield
+using Accessors
+
+using Revise
+using Newtrinos
+using CairoMakie
+
+# For memory tracking
+using Base: GC
+
+osc_cfg = Newtrinos.osc.OscillationConfig(
+    flavour=Newtrinos.osc.NNM(three_flavour=Newtrinos.osc.ThreeFlavour(ordering=:NO)),
+    propagation=Newtrinos.osc.Basic(),
+    states=Newtrinos.osc.All(),
+    interaction=Newtrinos.osc.SI()
+    )
+
+osc = Newtrinos.osc.configure(osc_cfg)
+
+atm_flux = Newtrinos.atm_flux.configure()
+erath_layers = Newtrinos.earth_layers.configure()
+xsec=Newtrinos.xsec.configure()
+
+physics = (; osc, atm_flux, earth_layers, xsec);
+
+experiments = (
+ 
+   juno= Newtrinos.juno.configure(physics;livetime_years = 6.0),
+   tao=Newtrinos.tao.configure(physics;livetime_years = 6.0),
+);
+
+p = Newtrinos.get_params(experiments)
+
+all_priors = Newtrinos.get_priors(experiments)
+
+p_complete = merge(p, (
+        tao_detection_epsilon = 1.0,
+        tao_res_a = 0.015,
+        tao_res_b = 0.0,
+        tao_res_c = 0.0,
+        tao_accidental_norm = 1.0,
+        tao_fast_neutron_norm = 1.0,
+        tao_lihe_norm = 1.0,
+    
+))
+
+m0_values=[1e-2]
+
+for i in 1:length(m0_values)
+
+    m0 =m0_values[i]
+    p_complete_new= merge(p_complete, (m₀=m0,))
+    
+    vars_to_scan = (r=31, N=31)  
+    
+    modified_priors = merge(p_complete_new,(N=DiscreteUniform(2,200), r =LogUniform(1e-8,1), 
+        junotao_flux_scale = Truncated(Normal(1.0, 0.02), 0.0, Inf), 
+        junotao_energy_scale = Truncated(Normal(1.0, 0.005), 0.0, Inf),
+        juno_detection_epsilon = Truncated(Normal(1.0, 0.01), 0.0, Inf),
+
+        juno_res_a = Truncated(Normal(0.0261, 0.0002), 0.0, Inf),
+        juno_res_b = Truncated(Normal(0.0082, 0.0001), 0.0, Inf),
+        juno_res_c = Truncated(Normal(0.0123, 0.0004), 0.0, Inf),
+        
+        junotao_shape_eps = Normal(0,1),
+        juno_geo_shape_eps = Normal(0,1),
+        
+        juno_geo_rate_norm = Truncated(Normal(1.0, 0.30), 0.0, Inf),
+        juno_accidental_norm = Truncated(Normal(1.0, 0.01), 0.0, Inf),     
+        juno_world_reactor_norm = Truncated(Normal(1.0, 0.02), 0.0, Inf),  
+        juno_lihe_norm = Truncated(Normal(1.0, 0.20), 0.0, Inf),      
+        juno_co_norm = Truncated(Normal(1.0, 0.50), 0.0, Inf),         
+        juno_atmnc_norm = Truncated(Normal(1.0, 0.50), 0.0, Inf),   
+        juno_fast_neutron_norm = Truncated(Normal(1.0, 1.0), 0.0, Inf), 
+        
[…]
