+++ src/Newtrinos.jl
@@ -25,7 +25,8 @@ include("experiments/minos/minos_sterile_16e20_POT/minos.jl")
 include("experiments/icecube/deepcore_3y_highstats_sample_b/deepcore.jl")
 include("experiments/kamland/kamland_7years/kamland.jl")
 include("experiments/km3net/orca6_433kton/orca.jl")
-include("experiments/nova/nova_try.jl")
+include("experiments/nova/nova_last.jl")
+include("experiments/t2k/t2k.jl")
 
 include("experiments/katrin/katrin_NN.jl")
 include("experiments/coherent/coherent_2020/coherent_csi.jl")

[…]
