### 2026-06-04 36cc6ad
modified content of getting started to enable an easier start into the package and also added tutorials for the different workflows needed to work with the package
