+++ docs/src/getting_started.md
@@ -1,55 +1,76 @@
-# Getting Started
+# First steps
+This section provides a simple starting example to give you a basic idea of how the Newtrinos.jl package works. It is NOT thought as a tutorial. We just want to guide you through how it basically works and is structured. Like riding a bike for the first time, but with training wheels.
 
-## Installation
+## Structure of the package
+The Newtrinos.jl architecture is organized into three orthogonal layers:
+- Physics layer: configure the theoretical physics models
+- Experiment layer: configure the experimental likelihoods
+- Analysis layer: analyse the data from the given experiments for a given physics model
 
-Newtrinos.jl is not yet registered in the Julia General registry. Install it directly from GitHub:
+Each layer treats the other layers as black boxes. This way, the package allows a flexible composition for statistical analysis of neutrino data.
 
-```julia
-using Pkg
-Pkg.add(url="https://github.com/philippeller/Newtrinos.jl.git")
-```
+## Example: 
+In this first example we define a physics model that contains decoherent neutrino propagation, standard matter interactions, and uses the inverted mass ordering. As experiment we choose the IceCube deepcore neutrino experiment. We want to evaluate the log-likelihood for this case. 
 
-Or for development:
-
-```bash
-git clone https://github.com/philippeller/Newtrinos.jl.git
-cd Newtrinos.jl
-julia --project -e 'using Pkg; Pkg.instantiate()'
-```
-
-## First Steps
-
-Load the package:
+### Load the package:
+If Newtrinos.jl is not installed yet on your machine you need to install it first. You can find the installation instructions [here](installation.md). If it's already installed load the package via: 
 
 ```julia
 using Newtrinos
-using DensityInterface
+using DensityInterface #need this for calculating the log-likelihood later
 ```
 
-Configure an experiment with default physics:
+### Configure the physics model:
+We want now to specify the physics model we want to investigate. 
 
-```julia
-exp = Newtrinos.dayabay.configure()
+`Newtrinos.osc.OscillationConfig(...)` creates an oscillation configuration by combining a flavour model, propagation model, eigenstate selection strategy, interaction model, and eigendecomposition algorithm. Calling it without arguments returns the default configuration (3-flavour, vacuum interactions, basic propagation, all states, default eigendecomposition).
+
+```julia 
+osc_cfg = Newtrinos.osc.OscillationConfig(
+    flavour     = Newtrinos.osc.ThreeFlavour(ordering=:IO), # use inverted mass ordering
+    propagation = Newtrinos.osc.Decoherent(), # use decoherent propagation
+    interaction = Newtrinos.osc.SI(), # use standard MSW effect
+)
 ```
 
-Each experiment returns a struct containing physics models, parameters, priors, data assets, and a forward model. Extract parameters and evaluate the likelihood:
+Then use `Newtrinos.osc.configure(oscillation_config)` to pass all relevant parameters, priors, matrices from the configuration into the Osc struct, i.e. Create a fully configured oscillation physics model from the given configuration. 
+
+```julia 
+osc_model = Newtrinos.osc.configure(osc_cfg)
+```
+
+Assemble the full physics model by combining the oscillation model with flux models, earth/matter models, and cross-section models.
 
 ```julia
-experiments = (; dayabay = exp)
+atm_flux     = Newtrinos.atm_flux.configure() # atmospherical flux
+earth_layers = Newtrinos.earth_layers.configure() # earth layer density model
+xsec         = Newtrinos.xsec.configure() # standard cross section model
+
+physics = (; osc=osc_model, atm_flux, earth_layers, xsec)
+```
+
+### Configure the experiment: 
[…]
