+++ docs/src/tutorials/physics_model.md
@@ -1,0 +1,262 @@
+# Physics Model Tutorial
+
+This page shows how to configure, inspect, and use the physics modules in Newtrinos.jl directly — without attaching any experiment. For the underlying theory, see [Neutrino Oscillation Physics](neutrino_physics.md).
+
+---
+
+## Configuring the Oscillation Module
+
+### Default Configuration
+
+The simplest way to get a fully configured oscillation module is to call `configure()` with no arguments. This uses standard three-flavour vacuum oscillations with basic propagation, all oscillating states, and best-fit parameter values as defaults.
+
+```julia
+using Newtrinos
+
+osc = Newtrinos.osc.configure()
+```
+
+The returned `Osc` struct contains the nominal parameter values and prior distributions:
+
+```julia
+osc.params   # NamedTuple: θ₁₂, θ₁₃, θ₂₃, δCP, Δm²₂₁, Δm²₃₁, ...
+osc.priors   # NamedTuple: prior distributions for each parameter
+```
+
+### Custom Configuration
+
+Pass an `OscillationConfig` to choose a different physics model. The config accepts four independent axes: flavour model, interaction model, propagation model, and eigendecomposition algorithm.
+
+```julia
+osc_cfg = Newtrinos.osc.OscillationConfig(
+    flavour     = Newtrinos.osc.ThreeFlavour(ordering = :IO), # inverted mass ordering
+    interaction = Newtrinos.osc.SI(),                         # standard MSW matter effects
+    propagation = Newtrinos.osc.Basic(),                      # coherent propagation
+)
+
+osc = Newtrinos.osc.configure(osc_cfg)
+```
+
+Available sub-models per axis:
+
+| Axis | Options |
+|---|---|
+| `flavour` | `ThreeFlavour`, `Sterile`, `ADD`, `Darkdim_*` |
+| `interaction` | `Vacuum`, `SI`, `NSI` |
+| `propagation` | `Basic`, `Decoherent`, `Damping` |
+
+---
+
+## Assembling the Full Physics Module
+
+Atmospheric-neutrino experiments require not just oscillations but also a flux model, an Earth density model, and a cross-section model. Configure each independently and assemble them into a `NamedTuple`:
+
+```julia
+osc          = Newtrinos.osc.configure()
+atm_flux     = Newtrinos.atm_flux.configure()
+earth_layers = Newtrinos.earth_layers.configure()
+xsec         = Newtrinos.xsec.configure()
+
+physics = (; osc, atm_flux, earth_layers, xsec)
+```
+
+To collect all parameters and priors across the full physics module:
+
+```julia
+params = Newtrinos.get_params(physics)
+priors = Newtrinos.get_priors(physics)
+```
+
+`get_params` merges the parameter `NamedTuple`s from every sub-module using `safe_merge`, which errors if any parameter name appears twice (preventing silent conflicts).
+
+---
+
+## Modifying Parameters
+
+All parameter collections are immutable `NamedTuple`s. Use `@reset` from [Accessors.jl](https://github.com/JuliaObjects/Accessors.jl) to create a modified copy without mutating the original:
+
+```julia
+using Accessors
+
+p = Newtrinos.get_params(physics)
+
+# Set θ₂₃ to maximal mixing
+@reset p.θ₂₃ = pi/4
+
+# Switch to CP conserving model by setting δCP = 0
+@reset p.δCP = 0 
+
[…]
