+++ docs/src/index.md
@@ -1,12 +1,10 @@
 # Newtrinos.jl
 
-**Newtrinos.jl** is a Julia package for the **global analysis of neutrino data**, fully open source and free to use under the MIT license.
-
-## Overview
+**Newtrinos.jl** is a Julia package for the **global analysis of neutrino data**. It is fully open source and free to use under the MIT license.
 
 The package is built to support flexible and modular analysis of neutrino physics, combining experimental data with physics models and inference tools. It supports both **Frequentist** (profile likelihood) and **Bayesian** (nested sampling, importance sampling) inference methods.
 
-## Key Features
+The key features of the package are: 
 
 - **11 experiments**: IceCube DeepCore, Daya Bay, KamLAND, MINOS, Super-Kamiokande, ORCA, JUNO, TAO, COHERENT CsI/LAr, IceCube Upgrade
 - **Modular physics**: Oscillations (3-flavour, sterile, ADD, dark dimensions), matter effects (SI, NSI), atmospheric fluxes, cross-sections
@@ -32,6 +30,13 @@ params = Newtrinos.get_params(experiments)
 # Evaluate
 logdensityof(likelihood, params)
 ```
+
+## Julia programming language
+
+The package is written in Julia. If you're not yet familiar with Julia and want to learn more about the language, here are a few resources to get started:
+- the [Julia Website](https://julialang.org/) offers many links to introductory videos and written tutorials. 
+- there also exists a [MATLAB-Python-Julia cheatsheet](https://cheatsheets.quantecon.org/)
+- this [article](https://www.stochasticlifestyle.com/why-numba-and-cython-are-not-substitutes-for-julia/) explains how Julia adresses several fundamental challenges inherent to scientific high-performance computing
 
 ## References
 

[…]
