### 2026-06-10 f5041ed
paper modified version 1
