+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Newtrinos.jl: A Julia Package for Global Neutrino Oscillation Fits'
+title: 'Newtrinos.jl: A Julia Package for Global Analysis of Neutrino data'
 tags:
   - julia
   - neutrino
@@ -24,12 +24,18 @@ bibliography: paper.bib
 ---
 
 # Summary
-*Newtrinos.jl* is a Julia package for performing global fits of neutrino oscillation data. It is designed to be easy to use, fast, and flexible. The software allows researchers to combine data from different experiments and test a variety of theoretical models. It supports automatic differentiation and is written entirely in Julia, making it suitable for modern scientific workflows.
+*Newtrinos.jl* is an open-source Julia package for performing global analyses of neutrino data. It is designed to be easy to use, fast, and flexible. The software allows researchers to combine data from different experiments and test a variety of theoretical models. It also allows adding experiments and theoretical models via Julias multiple dispatch without critical changes to the source code. It supports automatic differentiation and is written entirely in Julia, making it suitable for modern scientific workflows.
 
 # Statement of Need
-Global fits of neutrino oscillation data are essential for understanding neutrino properties and testing new physics [@Capozzi:2025wyn; @Esteban:2024eli; @deSalas:2020pgw]. These existing studies rely on proprietary, closed-source implementations, which makes their results difficult to verify and the software impossible to reuse. Some of these efforts also face technical limitations that restrict the complexity of datasets they can include. For example, the analysis in [@Capozzi:2025wyn] excludes atmospheric neutrino experiments entirely due to such limitations.
-Another recent study highlights the necessity of combining different datasets in performing a first joint fit between the T2K and the NOvA experiment [@T2K:2025wet].
-It is likely, that urgent questions in the field, such as the Neutrino Mass Ordering or CP violation in the lepton sector, can be addressed in global fits in the coming years.
+As neutrino physics transitions into an era of high-precision measurements, global fits of experimental neutrino data have become essential for understanding and extracting neutrino properties and probing new physics beyond the Standard Model [@Capozzi:2025wyn; @Esteban:2024eli; @deSalas:2020pgw]. While single-experiment analyses provide highly precise measurements of specific physical channels, global analyses complement them by systematically combining disparate datasets to break parameter degeneracies and maximize overall statistical sensitivity. Thus, urgent questions in the field, such as the Neutrino Mass Ordering or CP violation in the lepton sector, are also addressed in global fits. Recent efforts underscore this necessity; for example, the first joint fit between the T2K and NOvA experiments demonstrated that combining distinct experimental signatures is crucial to breaking parameter degeneracies [@T2K:2025wet].
+
+Performing a global neutrino fit requires managing the massive computational overhead of calculating oscillation probabilities across diverse, experiment-specific detector models, beam fluxes, and matter density profiles. Furthermore, the process is complicated by the need to rigorously account for a vast number of correlated systematic uncertainties and nuisance parameters for each contained experiment to ensure the statistical consistency of the joint data. On top of that, the relatively bad availability of public data and information regarding the different experiments complicates the process even further. So global fits require huge efforts and thus, analyses sometimes seem to be unfeasible. For example the analysis in [@Capozzi:2025wyn] excludes atmospheric neutrino experiments entirely due to such limitations, stating "For atmospheric neutrinos, currently involving hundreds of bins, dozens of systematic uncertainties, and refined statistical separation of event classes by flavor proxies, the construction of $\chi^2$ maps based on public information has become eventually unfeasible outside the experimental collaborations”.
+
[…]
