### 2026-06-12 59c1b3e
modified paper version 2
