+++ paper/paper.bib
@@ -154,39 +154,6 @@
     year = "2020"
 }
 
-@incollection{Neal:2011hmc,
-  author       = {Radford M. Neal},
-  title        = {{MCMC} Using {Hamiltonian} Dynamics},
-  booktitle    = {Handbook of {Markov} Chain {Monte Carlo}},
-  editor       = {Brooks, Steve and Gelman, Andrew and Jones, Galin L. and Meng, Xiao-Li},
-  publisher    = {Chapman and Hall/{CRC}},
-  year         = {2011},
-  pages        = {113--162},
-  eprint       = {1206.1901},
-  archivePrefix = {arXiv}
-}
-
-@article{Roberts:1996mala,
-  author  = {Gareth O. Roberts and Richard L. Tweedie},
-  title   = {Exponential Convergence of {Langevin} Distributions and Their Discrete Approximations},
-  journal = {Bernoulli},
-  volume  = {2},
-  number  = {4},
-  pages   = {341--363},
-  year    = {1996},
-  doi     = {10.2307/3318418}
-}
-
-@article{James:1975dr,
-  author  = {James, F. and Roos, M.},
-  title   = {Minuit: A System for Function Minimization and Analysis of the Parameter Errors and Correlations},
-  journal = {Comput. Phys. Commun.},
-  volume  = {10},
-  pages   = {343--367},
-  year    = {1975},
-  doi     = {10.1016/0010-4655(75)90039-9}
-}
-
 @article{Ettengruber:2024fcq,
     author = "Ettengruber, Manuel and Zander, Alan and Eller, Philipp",
     title = "{Testing the number of neutrino species with a global fit of neutrino data}",
@@ -226,4 +193,14 @@
     number = "5",
     pages = "055009",
     year = "2025"
+}
+
+@misc{eller2026atmosphericneutrinooscillationspicture,
+      title={Atmospheric Neutrino Oscillations: the Full Picture}, 
+      author={Philipp Eller},
+      year={2026},
+      eprint={2606.09714},
+      archivePrefix={arXiv},
+      primaryClass={hep-ex},
+      url={https://arxiv.org/abs/2606.09714}, 
 }
