+++ paper/paper.md
@@ -24,42 +24,50 @@ bibliography: paper.bib
 ---
 
 # Summary
-*Newtrinos.jl* is an open-source Julia package for performing global analyses of neutrino data. It is designed to be easy to use, fast, and flexible. The software allows researchers to combine data from different experiments and test a variety of theoretical models. It also allows adding experiments and theoretical models via Julias multiple dispatch without critical changes to the source code. It supports automatic differentiation and is written entirely in Julia, making it suitable for modern scientific workflows.
+Newtrinos.jl is an open-source Julia package for performing global analyses of neutrino data. It provides a modular, three-layer architecture that separates theoretical physics models, experiment configurations, and statistical inference methods. This allows researchers to freely combine data from a wide range of supported neutrino experiments and test them against a variety of theoretical models. New experiments and physics models can be easily added without modifying core code, making the package expandable. The Out-of-the-box usability of the package makes it easy to use and appealing for a broad range of researchers in the field. For the inference methods, full statistical likelihoods, that include all relevant systematic uncertainties as nuisance parameters, are constructed for each experiment alone and then are composed into a joint likelihood over the complete dataset. The unified parameter handling system automatically merges physics and nuisance parameters across the modules and between the experiments, ensuring consistency and eliminating redundancy in the joint fit. For inference, the package provides profile likelihood scanning, Bayesian posterior sampling, and Asimov-based sensitivity projections, all parallelisable across CPU threads or distributed workers. Written entirely in Julia, the complete forward model chain is compatible with ForwardDiff.jl out of the box, enabling exact gradient computation making it suitable for modern scientific workflows.
 
 # Statement of Need
 As neutrino physics transitions into an era of high-precision measurements, global fits of experimental neutrino data have become essential for understanding and extracting neutrino properties and probing new physics beyond the Standard Model [@Capozzi:2025wyn; @Esteban:2024eli; @deSalas:2020pgw]. While single-experiment analyses provide highly precise measurements of specific physical channels, global analyses complement them by systematically combining disparate datasets to break parameter degeneracies and maximize overall statistical sensitivity. Thus, urgent questions in the field, such as the Neutrino Mass Ordering or CP violation in the lepton sector, are also addressed in global fits. Recent efforts underscore this necessity; for example, the first joint fit between the T2K and NOvA experiments demonstrated that combining distinct experimental signatures is crucial to breaking parameter degeneracies [@T2K:2025wet].
 
-Performing a global neutrino fit requires managing the massive computational overhead of calculating oscillation probabilities across diverse, experiment-specific detector models, beam fluxes, and matter density profiles. Furthermore, the process is complicated by the need to rigorously account for a vast number of correlated systematic uncertainties and nuisance parameters for each contained experiment to ensure the statistical consistency of the joint data. On top of that, the relatively bad availability of public data and information regarding the different experiments complicates the process even further. So global fits require huge efforts and thus, analyses sometimes seem to be unfeasible. For example the analysis in [@Capozzi:2025wyn] excludes atmospheric neutrino experiments entirely due to such limitations, stating "For atmospheric neutrinos, currently involving hundreds of bins, dozens of systematic uncertainties, and refined statistical separation of event classes by flavor proxies, the construction of $\chi^2$ maps based on public information has become eventually unfeasible outside the experimental collaborations”.
[…]
