### 2026-06-12 64af7ca
paper version 3 + diagrams
