+++ src/experiments/gerda/gerda.jl
@@ -42,7 +42,7 @@ function get_assets(physics; datadir = @__DIR__)
 
     assets = (
 
-        observed =1 *1e28, #gerda 
+        observed =1.8 *1e26, #gerda 
         # 1*1e27,, # #legend 1000
         
     )
@@ -172,7 +172,7 @@ end
 
 
 
-function get_neutrinomassSTD(cfg=NNM(three_flavour=Newtrinos.osc.ThreeFlavour(ordering=:IO)))
+function get_neutrinomassSTD(cfg=NNM(three_flavour=Newtrinos.osc.ThreeFlavour(ordering=:NO)))
     function NeutrinoMassNNM(params::NamedTuple)
 
         
@@ -243,7 +243,7 @@ function get_neutrinomassSTD(cfg=NNM(three_flavour=Newtrinos.osc.ThreeFlavour(or
 end
 
 
-function get_neutrinomass(cfg=NNM(three_flavour=Newtrinos.osc.ThreeFlavour(ordering=:IO)))
+function get_neutrinomass(cfg=NNM(three_flavour=Newtrinos.osc.ThreeFlavour(ordering=:NO)))
     function NeutrinoMassNNM(params::NamedTuple)
 
         
@@ -395,7 +395,7 @@ function get_forward_model_correct(physics, assets)
     function forward_model(params)
         # Use the flavour model provided in physics to respect model and ordering from config
         cfg = physics.osc.cfg.flavour
-        observed = 1 * 1e28 # gerda
+        observed = 1.8 * 1e26# gerda
         # Use the appropriate neutrinomass/halftime function for this flavour model
         fun = get_halftime(cfg)
         predicted_value_T = fun(params)
@@ -403,7 +403,7 @@ function get_forward_model_correct(physics, assets)
         if predicted_value_T >= observed 
            predicted_value_T=observed
         end   
-        sigma = 0.01*1e28 #0.1*1e26 #0.01*1e26#
+        sigma = 0.01*1e26 #0.1*1e26 #0.01*1e26#
         #println("Predicted m_nu: ", predicted_value_T)
        
         return Normal(predicted_value_T, sigma)
