### 2026-08-18 #9644 @matthewfeickert
@rafaelab, @szabo137 thank you for your reviews!

@philippeller Apologies for the delayed response time. As Newtrinos.jl looks like it has passed all the reviewer comments I'll review the paper this week for editorial comments and then we can move forward with publication. I will also open up issues on things related to the repository that need to be addressed as well. c.f. https://github.com/philippeller/Newtrinos.jl/issues/241
