### 2026-08-25 #9644 @davschu
@rafaelab, @szabo137 thank you a lot for your time and reviews and the positive feedback!
@matthewfeickert we fixed and solved the problem issue at https://github.com/Newtrinos-org/Newtrinos.jl/issues/241. Did you find any other issues with the repository that we should fix?
Thank you!
